Struct openssl::aes::AesKey

pub struct AesKey(_);

Methods

impl AesKey

fn new_encrypt(key: &[u8]) -> Result<AesKey, KeyError>

Prepares a key for encryption.

Failure

Returns an error if the key is not 128, 192, or 256 bits.

fn new_decrypt(key: &[u8]) -> Result<AesKey, KeyError>

Prepares a key for decryption.

Failure

Returns an error if the key is not 128, 192, or 256 bits.