Struct openssl::bio::MemBio

pub struct MemBio {
    // some fields omitted
}

Methods

impl MemBio

fn new() -> Result<MemBio, SslError>

Creates a new owned memory based BIO

fn borrowed(bio: *mut BIO) -> MemBio

Returns a "borrow", i.e. it has no ownership

unsafe fn unwrap(self) -> *mut BIO

Consumes current bio and returns wrapped value Note that data ownership is lost and should be managed manually

unsafe fn get_handle(&self) -> *mut BIO

Temporarily gets wrapped value