Enum openssl::bn::RNGProperty

pub enum RNGProperty {
    MsbMaybeZero,
    MsbOne,
    TwoMsbOne,
}

Specifies the desired properties of a randomly generated BigNum.

Variants
MsbMaybeZero

The most significant bit of the number is allowed to be 0.

MsbOne

The MSB should be set to 1.

TwoMsbOne

The two most significant bits of the number will be set to 1, so that the product of two such random numbers will always have 2 * bits length.