Module openssl::cms

SMIME implementation using CMS

CMS (PKCS#7) is an encryption standard. It allows signing and encrypting data using X.509 certificates. The OpenSSL implementation of CMS is used in email encryption generated from a Vec of bytes. This Vec follows the smime protocol standards. Data accepted by this module will be smime type enveloped-data.

Structs

CMSOptions

CmsContentInfo

High level CMS wrapper

CmsContentInfoRef

Reference to CMSContentInfo