Struct openssl::cms::CmsContentInfoRef

pub struct CmsContentInfoRef(_);

Reference to CMSContentInfo

Implementations

impl CmsContentInfoRef

pub fn decrypt<T>(
&self,
pkey: &PKeyRef<T>,
cert: &X509
) -> Result<Vec<u8>, ErrorStack>where
T: HasPrivate,

Given the sender’s private key, pkey and the recipient’s certificate, cert, decrypt the data in self.

This corresponds to CMS_decrypt.

pub fn decrypt_without_cert_check<T>(
&self,
pkey: &PKeyRef<T>
) -> Result<Vec<u8>, ErrorStack>where
T: HasPrivate,

Given the sender’s private key, pkey, decrypt the data in self without validating the recipient certificate.

Warning: Not checking the recipient certificate may leave you vulnerable to Bleichenbacher’s attack on PKCS#1 v1.5 RSA padding.

This corresponds to CMS_decrypt.

pub fn to_der(&self) -> Result<Vec<u8>, ErrorStack>

Serializes this CmsContentInfo using DER.

This corresponds to i2d_CMS_ContentInfo.

pub fn to_pem(&self) -> Result<Vec<u8>, ErrorStack>

Serializes this CmsContentInfo using DER.

This corresponds to PEM_write_bio_CMS.