Struct openssl::crypto::hash::Hasher

pub struct Hasher {
    // some fields omitted
}

Methods

impl Hasher

fn new(ht: HashType) -> Hasher

fn with_context(ctx: HasherContext, ht: HashType) -> Hasher

fn update(&mut self, data: &[u8])

Update this hasher with more input bytes

fn finalize(self) -> Vec<u8>

Return the digest of all bytes added to this hasher since its last initialization

fn finalize_reuse(self) -> (Vec<u8>, HasherContext)

Return the digest of all bytes added to this hasher since its last initialization and its context for reuse