Function openssl::crypto::pkcs5::evp_bytes_to_key_pbkdf1_compatible

pub fn evp_bytes_to_key_pbkdf1_compatible(typ: Type, message_digest_type: Type, data: &[u8], salt: Option<&[u8]>, count: u32) -> KeyIvPair

Derives a key and an IV from various parameters.

If specified salt must be 8 bytes in length.

If the total key and IV length is less than 16 bytes and MD5 is used then the algorithm is compatible with the key derivation algorithm from PKCS#5 v1.5 or PBKDF1 from PKCS#5 v2.0.

New applications should not use this and instead use pbkdf2_hmac_sha1 or another more modern key derivation algorithm.