Struct openssl::crypto::symm::Crypter

pub struct Crypter { /* fields omitted */ }

Represents a symmetric cipher context.

Methods

impl Crypter

pub fn new(t: Type) -> Crypter

pub fn pad(&self, padding: bool)

Enables or disables padding. If padding is disabled, total amount of data encrypted must be a multiple of block size.

pub fn init(&self, mode: Mode, key: &[u8], iv: &[u8])

Initializes this crypter.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn update(&self, data: &[u8]) -> Vec<u8>

Update this crypter with more data to encrypt or decrypt. Returns encrypted or decrypted bytes.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn finalize(&self) -> Vec<u8>

Finish crypting. Returns the remaining partial block of output, if any.