Struct openssl::dh::Dh

pub struct Dh<T>(_, _);

Implementations

impl Dh<Params>

pub fn from_params(
p: BigNum,
g: BigNum,
q: BigNum
) -> Result<Dh<Params>, ErrorStack>

pub fn from_pqg(
prime_p: BigNum,
prime_q: Option<BigNum>,
generator: BigNum
) -> Result<Dh<Params>, ErrorStack>

Creates a DH instance based upon the given primes and generator params.

This corresponds to DH_set0_pqg.

pub fn set_private_key(self, priv_key: BigNum) -> Result<Dh<Private>, ErrorStack>

Sets the private key on the DH object and recomputes the public key.

pub fn generate_params(
prime_len: u32,
generator: u32
) -> Result<Dh<Params>, ErrorStack>

Generates DH params based on the given prime_len and a fixed generator value.

This corresponds to DH_generate_parameters_ex.

pub fn generate_key(self) -> Result<Dh<Private>, ErrorStack>

Generates a public and a private key based on the DH params.

This corresponds to DH_generate_key.

pub fn params_from_pem(pem: &[u8]) -> Result<Dh<Params>, ErrorStack>

Deserializes a PEM-encoded PKCS#3 DHpararameters structure.

The input should have a header of -----BEGIN DH PARAMETERS-----.

This corresponds to PEM_read_bio_DHparams.

pub fn params_from_der(der: &[u8]) -> Result<Dh<Params>, ErrorStack>

Deserializes a DER-encoded PKCS#3 DHparameters structure.

This corresponds to d2i_DHparams.

pub fn get_1024_160() -> Result<Dh<Params>, ErrorStack>

Requires OpenSSL 1.0.2 or newer.

This corresponds to DH_get_1024_160.

pub fn get_2048_224() -> Result<Dh<Params>, ErrorStack>

Requires OpenSSL 1.0.2 or newer.

This corresponds to DH_get_2048_224.

pub fn get_2048_256() -> Result<Dh<Params>, ErrorStack>

Requires OpenSSL 1.0.2 or newer.

This corresponds to DH_get_2048_256.

impl<T> Dh<T> where
T: HasParams,

pub fn prime_p(&self) -> &BigNumRef

Returns the prime p from the DH instance.

This corresponds to DH_get0_pqg.

pub fn prime_q(&self) -> Option<&BigNumRef>

Returns the prime q from the DH instance.

This corresponds to DH_get0_pqg.

pub fn generator(&self) -> &BigNumRef

Returns the generator from the DH instance.

This corresponds to DH_get0_pqg.