Struct openssl::dsa::DsaRef

pub struct DsaRef<T>(_, _);

Reference to Dsa.

Implementations

impl<T> DsaRef<T> where
T: HasPublic,

pub fn public_key_to_pem(&self) -> Result<Vec<u8>, ErrorStack>

Serialies the public key into a PEM-encoded SubjectPublicKeyInfo structure.

The output will have a header of -----BEGIN PUBLIC KEY-----.

This corresponds to PEM_write_bio_DSA_PUBKEY.

pub fn public_key_to_der(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the public key into a DER-encoded SubjectPublicKeyInfo structure.

This corresponds to i2d_DSA_PUBKEY.

pub fn pub_key(&self) -> &BigNumRef

Returns a reference to the public key component of self.

impl<T> DsaRef<T> where
T: HasPrivate,

pub fn private_key_to_pem(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the private key to a PEM-encoded DSAPrivateKey structure.

The output will have a header of -----BEGIN DSA PRIVATE KEY-----.

This corresponds to PEM_write_bio_DSAPrivateKey.

pub fn private_key_to_pem_passphrase(
&self,
cipher: Cipher,
passphrase: &[u8]
) -> Result<Vec<u8>, ErrorStack>

Serializes the private key to a PEM-encoded encrypted DSAPrivateKey structure.

The output will have a header of -----BEGIN DSA PRIVATE KEY-----.

This corresponds to PEM_write_bio_DSAPrivateKey.

pub fn priv_key(&self) -> &BigNumRef

Returns a reference to the private key component of self.

impl<T> DsaRef<T> where
T: HasParams,

pub fn size(&self) -> u32

Returns the maximum size of the signature output by self in bytes.

OpenSSL documentation at DSA_size

pub fn p(&self) -> &BigNumRef

Returns the DSA prime parameter of self.

pub fn q(&self) -> &BigNumRef

Returns the DSA sub-prime parameter of self.

pub fn g(&self) -> &BigNumRef

Returns the DSA base parameter of self.