Struct openssl::ec::Asn1Flag

pub struct Asn1Flag(_);

Named Curve or Explicit

This type acts as a boolean as to whether the EcGroup is named or explicit.

Methods

impl Asn1Flag

pub const EXPLICIT_CURVE: Asn1Flag

EXPLICIT_CURVE: Asn1Flag = Asn1Flag(0)

Curve defined using polynomial parameters

Most applications use a named EC_GROUP curve, however, support is included to explicitly define the curve used to calculate keys This information would need to be known by both endpoint to make communication effective.

OPENSSL_EC_EXPLICIT_CURVE, but that was only added in 1.1. Man page documents that 0 can be used in older versions.

OpenSSL documentation at EC_GROUP

pub const NAMED_CURVE: Asn1Flag

NAMED_CURVE: Asn1Flag = Asn1Flag(ffi::OPENSSL_EC_NAMED_CURVE)

Standard Curves

Curves that make up the typical encryption use cases. The collection of curves are well known but extensible.

OpenSSL documentation at EC_GROUP