Struct openssl::ec::PointConversionForm

pub struct PointConversionForm(_);

Compressed or Uncompressed conversion

Conversion from the binary value of the point on the curve is performed in one of compressed, uncompressed, or hybrid conversions. The default is compressed, except for binary curves.

Further documentation is available in the X9.62 standard.

Methods

impl PointConversionForm

pub const COMPRESSED: PointConversionForm

COMPRESSED: PointConversionForm = PointConversionForm(ffi::point_conversion_form_t::POINT_CONVERSION_COMPRESSED)

Compressed conversion from point value.

pub const UNCOMPRESSED: PointConversionForm

UNCOMPRESSED: PointConversionForm = PointConversionForm(ffi::point_conversion_form_t::POINT_CONVERSION_UNCOMPRESSED)

Uncompressed conversion from point value.

pub const HYBRID: PointConversionForm

HYBRID: PointConversionForm = PointConversionForm(ffi::point_conversion_form_t::POINT_CONVERSION_HYBRID)

Performs both compressed and uncompressed conversions.