Struct openssl::ecdsa::EcdsaSig

pub struct EcdsaSig(_);

A low level interface to ECDSA

OpenSSL documentation at ECDSA_sign

Implementations

impl EcdsaSig

pub fn sign<T>(data: &[u8], eckey: &EcKeyRef<T>) -> Result<EcdsaSig, ErrorStack> where
T: HasPrivate,

Computes a digital signature of the hash value data using the private EC key eckey.

OpenSSL documentation at ECDSA_do_sign

pub fn from_private_components(
r: BigNum,
s: BigNum
) -> Result<EcdsaSig, ErrorStack>

Returns a new EcdsaSig by setting the r and s values associated with a ECDSA signature.

OpenSSL documentation at ECDSA_SIG_set0

pub fn from_der(der: &[u8]) -> Result<EcdsaSig, ErrorStack>

Decodes a DER-encoded ECDSA signature.

This corresponds to d2i_ECDSA_SIG.