Struct openssl::error::Error

pub struct Error { /* fields omitted */ }

An error reported from OpenSSL.

Methods

impl Error

pub fn get() -> Option<Error>

Returns the first error on the OpenSSL error stack.

pub fn put(&self)

Pushes the error back onto the OpenSSL error stack.

pub fn code(&self) -> c_ulong

Returns the raw OpenSSL error code for this error.

pub fn library(&self) -> Option<&'static str>

Returns the name of the library reporting the error, if available.

pub fn function(&self) -> Option<&'static str>

Returns the name of the function reporting the error.

pub fn reason(&self) -> Option<&'static str>

Returns the reason for the error.

pub fn file(&self) -> &'static str

Returns the name of the source file which encountered the error.

pub fn line(&self) -> u32

Returns the line in the source file which encountered the error.

pub fn data(&self) -> Option<&str>

Returns additional data describing the error.