Struct openssl::hash::Hasher

pub struct Hasher { /* fields omitted */ }

Provides message digest (hash) computation.

Examples

Calculate a hash in one go:

use openssl::hash::{hash, MessageDigest};

let data = b"\x42\xF4\x97\xE0";
let spec = b"\x7c\x43\x0f\x17\x8a\xef\xdf\x14\x87\xfe\xe7\x14\x4e\x96\x41\xe2";
let res = hash(MessageDigest::md5(), data).unwrap();
assert_eq!(&*res, spec);

Supply the input in chunks:

use openssl::hash::{Hasher, MessageDigest};

let data = [b"\x42\xF4", b"\x97\xE0"];
let spec = b"\x7c\x43\x0f\x17\x8a\xef\xdf\x14\x87\xfe\xe7\x14\x4e\x96\x41\xe2";
let mut h = Hasher::new(MessageDigest::md5()).unwrap();
h.update(data[0]).unwrap();
h.update(data[1]).unwrap();
let res = h.finish().unwrap();
assert_eq!(&*res, spec);

Warning

Don't actually use MD5 and SHA-1 hashes, they're not secure anymore.

Don't ever hash passwords, use the functions in the pkcs5 module or bcrypt/scrypt instead.

Methods

impl Hasher

pub fn new(ty: MessageDigest) -> Result<Hasher, ErrorStack>

Creates a new Hasher with the specified hash type.

pub fn update(&mut self, data: &[u8]) -> Result<(), ErrorStack>

Feeds data into the hasher.

pub fn finish(&mut self) -> Result<DigestBytes, ErrorStack>

Returns the hash of the data written and resets the hasher.