Struct openssl::pkcs12::Pkcs12Builder

pub struct Pkcs12Builder { /* fields omitted */ }

Methods

impl Pkcs12Builder

fn key_algorithm(&mut self, nid: Nid) -> &mut Self

The encryption algorithm that should be used for the key

fn cert_algorithm(&mut self, nid: Nid) -> &mut Self

The encryption algorithm that should be used for the cert

fn key_iter(&mut self, iter: u32) -> &mut Self

Key iteration count, default is 2048 as of this writing

fn mac_iter(&mut self, mac_iter: u32) -> &mut Self

MAC iteration count, default is the same as key_iter.

Old implementations don't understand MAC iterations greater than 1, (pre 1.0.1?), if such compatibility is required this should be set to 1.

fn ca(&mut self, ca: Stack<X509>) -> &mut Self

An additional set of certificates to include in the archive beyond the one provided to build.

fn build(self,
password: &str,
friendly_name: &str,
pkey: &PKeyRef,
cert: &X509)
-> Result<Pkcs12, ErrorStack>

Builds the PKCS #12 object

Arguments

password - the password used to encrypt the key and certificate

friendly_name - user defined name for the certificate

pkey - key to store

cert - certificate to store