Struct openssl::pkcs7::Pkcs7

pub struct Pkcs7(_);

A PKCS#7 structure.

Contains signed and/or encrypted data.

Methods

impl Pkcs7

pub fn from_pem(pem: &[u8]) -> Result<Pkcs7, ErrorStack>

Deserializes a PEM-encoded PKCS#7 signature

The input should have a header of -----BEGIN PKCS7-----.

This corresponds to PEM_read_bio_PKCS7.

pub fn from_smime(input: &[u8]) -> Result<(Pkcs7, Option<Vec<u8>>), ErrorStack>

Parses a message in S/MIME format.

Returns the loaded signature, along with the cleartext message (if available).

This corresponds to SMIME_read_PKCS7.

pub fn encrypt(
certs: &StackRef<X509>,
input: &[u8],
cipher: Cipher,
flags: Pkcs7Flags
) -> Result<Pkcs7, ErrorStack>

Creates and returns a PKCS#7 envelopedData structure.

certs is a list of recipient certificates. input is the content to be encrypted. cipher is the symmetric cipher to use. flags is an optional set of flags.

This corresponds to PKCS7_encrypt.

pub fn sign<PT>(
signcert: &X509Ref,
pkey: &PKeyRef<PT>,
certs: &StackRef<X509>,
input: &[u8],
flags: Pkcs7Flags
) -> Result<Pkcs7, ErrorStack> where
PT: HasPrivate,

Creates and returns a PKCS#7 signedData structure.

signcert is the certificate to sign with, pkey is the corresponding private key. certs is an optional additional set of certificates to include in the PKCS#7 structure (for example any intermediate CAs in the chain).

This corresponds to PKCS7_sign.