Struct openssl::pkey::PKey

pub struct PKey(_);

Methods

impl PKey

fn from_rsa(rsa: Rsa) -> Result<PKey, ErrorStack>

Creates a new PKey containing an RSA key.

fn from_dsa(dsa: Dsa) -> Result<PKey, ErrorStack>

Creates a new PKey containing a DSA key.

fn from_dh(dh: Dh) -> Result<PKey, ErrorStack>

Creates a new PKey containing a Diffie-Hellman key.

fn from_ec_key(ec_key: EcKey) -> Result<PKey, ErrorStack>

Creates a new PKey containing an elliptic curve key.

fn hmac(key: &[u8]) -> Result<PKey, ErrorStack>

Creates a new PKey containing an HMAC key.

Note

To compute HMAC values, use the sign module.

fn private_key_from_pem(pem: &[u8]) -> Result<PKey, ErrorStack>

Deserializes a PEM-formatted private key.

fn private_key_from_pem_passphrase(
pem: &[u8],
passphrase: &[u8]
) -> Result<PKey, ErrorStack>

Deserializes a PEM-formatted private key, using the supplied password if the key is encrypted.

Panics

Panics if passphrase contains an embedded null.

fn private_key_from_pem_callback<F>(
pem: &[u8],
callback: F
) -> Result<PKey, ErrorStack> where
F: FnOnce(&mut [u8]) -> Result<usize, ErrorStack>,

Deserializes a PEM-formatted private key, using a callback to retrieve a password if the key is encrypted.

The callback should copy the password into the provided buffer and return the number of bytes written.

fn public_key_from_pem(pem: &[u8]) -> Result<PKey, ErrorStack>

Deserializes a public key from PEM-formatted data.

fn private_key_from_pkcs8_callback<F>(
der: &[u8],
callback: F
) -> Result<PKey, ErrorStack> where
F: FnOnce(&mut [u8]) -> Result<usize, ErrorStack>,

Deserializes a DER-formatted PKCS#8 private key, using a callback to retrieve the password if the key is encrpyted.

The callback should copy the password into the provided buffer and return the number of bytes written.

fn private_key_from_pkcs8_passphrase(
der: &[u8],
passphrase: &[u8]
) -> Result<PKey, ErrorStack>

Deserializes a DER-formatted PKCS#8 private key, using the supplied password if the key is encrypted.

Panics

Panics if passphrase contains an embedded null.

fn private_key_from_pem_cb<F>(
buf: &[u8],
pass_cb: F
) -> Result<PKey, ErrorStack> where
F: FnOnce(&mut [c_char]) -> usize,

Deprecated since 0.9.2
: use private_key_from_pem_callback