openssl::pkey_ctx

Struct PkeyCtx

pub struct PkeyCtx<T>(/* private fields */);

A context object which can perform asymmetric cryptography operations.

Implementations

impl<T> PkeyCtx<T>

pub fn new(pkey: &PKeyRef<T>) -> Result<Self, ErrorStack>

Creates a new pkey context using the provided key.

This corresponds to EVP_PKEY_CTX_new.

impl PkeyCtx<()>

pub fn new_id(id: Id) -> Result<Self, ErrorStack>

Creates a new pkey context for the specified algorithm ID.

This corresponds to EVP_PKEY_CTX_new_id.