Struct openssl::rsa::RsaPrivateKeyBuilder

pub struct RsaPrivateKeyBuilder { /* private fields */ }

Implementations

impl RsaPrivateKeyBuilder

pub fn new( n: BigNum, e: BigNum, d: BigNum ) -> Result<RsaPrivateKeyBuilder, ErrorStack>

Creates a new RsaPrivateKeyBuilder.

n is the modulus common to both public and private key. e is the public exponent and d is the private exponent.

This corresponds to RSA_new and uses RSA_set0_key.

pub fn set_factors( self, p: BigNum, q: BigNum ) -> Result<RsaPrivateKeyBuilder, ErrorStack>

Sets the factors of the Rsa key.

p and q are the first and second factors of n.

This corresponds to RSA_set0_factors.

pub fn set_crt_params( self, dmp1: BigNum, dmq1: BigNum, iqmp: BigNum ) -> Result<RsaPrivateKeyBuilder, ErrorStack>

Sets the Chinese Remainder Theorem params of the Rsa key.

dmp1, dmq1, and iqmp are the exponents and coefficient for CRT calculations which is used to speed up RSA operations.

This corresponds to RSA_set0_crt_params.

pub fn build(self) -> Rsa<Private>

Returns the Rsa key.