Struct openssl::sign::Signer

pub struct Signer<'a> { /* private fields */ }

A type which computes cryptographic signatures of data.

Implementations

impl<'a> Signer<'a>

pub fn new<T>(
type_: MessageDigest,
pkey: &'a PKeyRef<T>
) -> Result<Signer<'a>, ErrorStack> where
T: HasPrivate,

Creates a new Signer.

This cannot be used with Ed25519 or Ed448 keys. Please refer to new_without_digest.

OpenSSL documentation at EVP_DigestSignInit.

pub fn new_without_digest<T>(
pkey: &'a PKeyRef<T>
) -> Result<Signer<'a>, ErrorStack> where
T: HasPrivate,

Creates a new Signer without a digest.

This is the only way to create a Verifier for Ed25519 or Ed448 keys. It can also be used to create a CMAC.

OpenSSL documentation at EVP_DigestSignInit.

pub fn rsa_padding(&self) -> Result<Padding, ErrorStack>

Returns the RSA padding mode in use.

This is only useful for RSA keys.

This corresponds to EVP_PKEY_CTX_get_rsa_padding.

pub fn set_rsa_padding(&mut self, padding: Padding) -> Result<(), ErrorStack>

Sets the RSA padding mode.

This is only useful for RSA keys.

This corresponds to EVP_PKEY_CTX_set_rsa_padding.

pub fn set_rsa_pss_saltlen(
&mut self,
len: RsaPssSaltlen
) -> Result<(), ErrorStack>

Sets the RSA PSS salt length.

This is only useful for RSA keys.

This corresponds to EVP_PKEY_CTX_set_rsa_pss_saltlen.

pub fn set_rsa_mgf1_md(&mut self, md: MessageDigest) -> Result<(), ErrorStack>

Sets the RSA MGF1 algorithm.

This is only useful for RSA keys.

This corresponds to EVP_PKEY_CTX_set_rsa_mgf1_md.

pub fn update(&mut self, buf: &[u8]) -> Result<(), ErrorStack>

Feeds more data into the Signer.

Please note that PureEdDSA (Ed25519 and Ed448 keys) do not support streaming. Use sign_oneshot instead.

OpenSSL documentation at EVP_DigestUpdate.

pub fn len(&self) -> Result<usize, ErrorStack>

Computes an upper bound on the signature length.

The actual signature may be shorter than this value. Check the return value of sign to get the exact length.

OpenSSL documentation at EVP_DigestSignFinal.

pub fn sign(&self, buf: &mut [u8]) -> Result<usize, ErrorStack>

Writes the signature into the provided buffer, returning the number of bytes written.

This method will fail if the buffer is not large enough for the signature. Use the len method to get an upper bound on the required size.

OpenSSL documentation at EVP_DigestSignFinal.

pub fn sign_to_vec(&self) -> Result<Vec<u8>, ErrorStack>

Returns the signature.

This is a simple convenience wrapper over len and sign.

pub fn sign_oneshot(
&mut self,
sig_buf: &mut [u8],
data_buf: &[u8]
) -> Result<usize, ErrorStack>

Signs the data in data_buf and writes the signature into the buffer sig_buf, returning the number of bytes written.

For PureEdDSA (Ed25519 and Ed448 keys) this is the only way to sign data.

This method will fail if the buffer is not large enough for the signature. Use the len method to get an upper bound on the required size.

OpenSSL documentation at EVP_DigestSign.

pub fn sign_oneshot_to_vec(
&mut self,
data_buf: &[u8]
) -> Result<Vec<u8>, ErrorStack>

Returns the signature.

This is a simple convenience wrapper over len and sign_oneshot.