Struct openssl::sign::Verifier

pub struct Verifier<'a> { /* private fields */ }

A type which can be used to verify the integrity and authenticity of data given the signature.

Implementations

impl<'a> Verifier<'a>

A type which verifies cryptographic signatures of data.

pub fn new<T>(
type_: MessageDigest,
pkey: &'a PKeyRef<T>
) -> Result<Verifier<'a>, ErrorStack>where
T: HasPublic,

Creates a new Verifier.

This cannot be used with Ed25519 or Ed448 keys. Please refer to Verifier::new_without_digest.

OpenSSL documentation at EVP_DigestVerifyInit.

pub fn new_without_digest<T>(
pkey: &'a PKeyRef<T>
) -> Result<Verifier<'a>, ErrorStack>where
T: HasPublic,

Creates a new Verifier without a digest.

This is the only way to create a Verifier for Ed25519 or Ed448 keys.

OpenSSL documentation at EVP_DigestVerifyInit.

pub fn rsa_padding(&self) -> Result<Padding, ErrorStack>

Returns the RSA padding mode in use.

This is only useful for RSA keys.

This corresponds to EVP_PKEY_CTX_get_rsa_padding.

pub fn set_rsa_padding(&mut self, padding: Padding) -> Result<(), ErrorStack>

Sets the RSA padding mode.

This is only useful for RSA keys.

This corresponds to EVP_PKEY_CTX_set_rsa_padding.

pub fn set_rsa_pss_saltlen(
&mut self,
len: RsaPssSaltlen
) -> Result<(), ErrorStack>

Sets the RSA PSS salt length.

This is only useful for RSA keys.

This corresponds to EVP_PKEY_CTX_set_rsa_pss_saltlen.

pub fn set_rsa_mgf1_md(&mut self, md: MessageDigest) -> Result<(), ErrorStack>

Sets the RSA MGF1 algorithm.

This is only useful for RSA keys.

This corresponds to EVP_PKEY_CTX_set_rsa_mgf1_md.

pub fn update(&mut self, buf: &[u8]) -> Result<(), ErrorStack>

Feeds more data into the Verifier.

Please note that PureEdDSA (Ed25519 and Ed448 keys) do not support streaming. Use Verifier::verify_oneshot instead.

OpenSSL documentation at EVP_DigestUpdate.

pub fn verify(&self, signature: &[u8]) -> Result<bool, ErrorStack>

Determines if the data fed into the Verifier matches the provided signature.

OpenSSL documentation at EVP_DigestVerifyFinal.

pub fn verify_oneshot(
&mut self,
signature: &[u8],
buf: &[u8]
) -> Result<bool, ErrorStack>

Determines if the data given in buf matches the provided signature.

OpenSSL documentation at EVP_DigestVerify.