Enum openssl::ssl::error::SslError

pub enum SslError {
    StreamError(IoError),
    SslSessionClosed,
    OpenSslErrors(Vec<OpensslError>),
}

An SSL error

Variants

StreamError

The underlying stream has reported an error

SslSessionClosed

The SSL session has been closed by the other end

OpenSslErrors

An error in the OpenSSL library

Methods

impl SslError

fn get() -> SslError

Creates a new OpenSslErrors with the current contents of the error stack.