Function openssl::ssl::select_next_proto

pub fn select_next_proto<'a>(
    server: &[u8], 
    client: &'a [u8]
) -> Option<&'a [u8]>

A standard implementation of protocol selection for Application Layer Protocol Negotiation (ALPN).

server should contain the server's list of supported protocols and client the client's. They must both be in the ALPN wire format. See the documentation for SslContextBuilder::set_alpn_protos for details.

It will select the first protocol supported by the server which is also supported by the client.

This corresponds to SSL_select_next_proto.