Struct openssl::ssl::ErrorCode

pub struct ErrorCode(_);

An error code returned from SSL functions.

Methods

impl ErrorCode

pub fn from_raw(raw: c_int) -> ErrorCode

pub fn as_raw(&self) -> c_int

pub const ZERO_RETURN: ErrorCode

ZERO_RETURN: ErrorCode = ErrorCode(ffi::SSL_ERROR_ZERO_RETURN)

The SSL session has been closed.

pub const WANT_READ: ErrorCode

WANT_READ: ErrorCode = ErrorCode(ffi::SSL_ERROR_WANT_READ)

An attempt to read data from the underlying socket returned WouldBlock.

Wait for read readiness and retry the operation.

pub const WANT_WRITE: ErrorCode

WANT_WRITE: ErrorCode = ErrorCode(ffi::SSL_ERROR_WANT_WRITE)

An attempt to write data to the underlying socket returned WouldBlock.

Wait for write readiness and retry the operation.

pub const SYSCALL: ErrorCode

SYSCALL: ErrorCode = ErrorCode(ffi::SSL_ERROR_SYSCALL)

A non-recoverable IO error occurred.

pub const SSL: ErrorCode

SSL: ErrorCode = ErrorCode(ffi::SSL_ERROR_SSL)

An error occurred in the SSL library.