Struct openssl::ssl::Ssl

pub struct Ssl { /* fields omitted */ }

Methods

impl Ssl

pub fn new(ctx: &SslContext) -> Result<Ssl, SslError>

pub fn set_verify(&mut self, mode: SslVerifyMode)

Sets the verification mode to be used during the handshake process.

Use set_verify_callback to additionally add a callback.

pub fn set_verify_callback<F>(&mut self, mode: SslVerifyMode, verify: F) where
F: Fn(bool, &X509StoreContext) -> bool + Any + 'static + Sync + Send,

Sets the certificate verification callback to be used during the handshake process.

The callback is provided with a boolean indicating if the preveification process was successful, and an object providing access to the certificate chain. It should return true if the certificate chain is valid and false otherwise.

pub fn get_current_cipher<'a>(&'a self) -> Option<SslCipher<'a>>

pub fn state_string(&self) -> &'static str

pub fn state_string_long(&self) -> &'static str

pub fn set_hostname(&self, hostname: &str) -> Result<(), SslError>

Sets the host name to be used with SNI (Server Name Indication).

pub fn peer_certificate(&self) -> Option<X509>

Returns the certificate of the peer, if present.

pub fn version(&self) -> &'static str

Returns the name of the protocol used for the connection, e.g. "TLSv1.2", "SSLv3", etc.

pub fn pending(&self) -> usize

Returns the number of bytes remaining in the currently processed TLS record.

pub fn compression(&self) -> Option<String>

Returns the compression currently in use.

The result will be either None, indicating no compression is in use, or a string with the compression name.

pub fn get_ssl_method(&self) -> Option<SslMethod>

pub fn get_servername(&self) -> Option<String>

Returns the server's name for the current connection

pub fn set_ssl_context(&self, ctx: &SslContext) -> SslContext

change the context corresponding to the current connection

Returns a clone of the SslContext @ctx (ie: the new context). The old context is freed.

pub fn get_ssl_context(&self) -> SslContext

obtain the context corresponding to the current connection