Struct openssl::ssl::SslConnector

pub struct SslConnector(_);

A type which wraps client-side streams in a TLS session.

OpenSSL's default configuration is highly insecure. This connector manages the OpenSSL structures, configuring cipher suites, session options, hostname verification, and more.

OpenSSL's built in hostname verification is used when linking against OpenSSL 1.0.2 or 1.1.0, and a custom implementation is used when linking against OpenSSL 1.0.1.

Methods

impl SslConnector

pub fn builder(method: SslMethod) -> Result<SslConnectorBuilder, ErrorStack>

Creates a new builder for TLS connections.

The default configuration is subject to change, and is currently derived from Python.

pub fn connect<S>(
&self,
domain: &str,
stream: S
) -> Result<SslStream<S>, HandshakeError<S>> where
S: Read + Write,

Initiates a client-side TLS session on a stream.

The domain is used for SNI and hostname verification.

pub fn configure(&self) -> Result<ConnectConfiguration, ErrorStack>

Returns a structure allowing for configuration of a single TLS session before connection.