Struct openssl::ssl::SslContextBuilder

pub struct SslContextBuilder(_);

A builder for SslContexts.

Methods

impl SslContextBuilder

pub fn new(method: SslMethod) -> Result<SslContextBuilder, ErrorStack>

Creates a new SslContextBuilder.

This corresponds to SSL_CTX_new.

pub unsafe fn from_ptr(ctx: *mut SSL_CTX) -> SslContextBuilder

Creates an SslContextBuilder from a pointer to a raw OpenSSL value.

pub fn as_ptr(&self) -> *mut SSL_CTX

Returns a pointer to the raw OpenSSL value.

pub fn set_verify(&mut self, mode: SslVerifyMode)

Configures the certificate verification method for new connections.

This corresponds to SSL_CTX_set_verify.

pub fn set_verify_callback<F>(&mut self, mode: SslVerifyMode, verify: F) where
F: Fn(bool, &mut X509StoreContextRef) -> bool + 'static + Sync + Send,

Configures the certificate verification method for new connections and registers a verification callback.

The callback is passed a boolean indicating if OpenSSL's internal verification succeeded as well as a reference to the X509StoreContext which can be used to examine the certificate chain. It should return a boolean indicating if verification succeeded.

This corresponds to SSL_CTX_set_verify.

pub fn set_servername_callback<F>(&mut self, callback: F) where
F: Fn(&mut SslRef, &mut SslAlert) -> Result<(), SniError> + 'static + Sync + Send,

Configures the server name indication (SNI) callback for new connections.

SNI is used to allow a single server to handle requests for multiple domains, each of which has its own certificate chain and configuration.

Obtain the server name with the servername method and then set the corresponding context with set_ssl_context

This corresponds to SSL_CTX_set_tlsext_servername_callback.

pub fn set_verify_depth(&mut self, depth: u32)

Sets the certificate verification depth.

If the peer's certificate chain is longer than this value, verification will fail.

This corresponds to SSL_CTX_set_verify_depth.

pub fn set_read_ahead(&mut self, read_ahead: bool)

Controls read ahead behavior.

If enabled, OpenSSL will read as much data as is available from the underlying stream, instead of a single record at a time.

It has no effect when used with DTLS.

This corresponds to SSL_CTX_set_read_ahead.

pub fn set_mode(&mut self, mode: SslMode) -> SslMode

Sets the mode used by the context, returning the previous mode.

This corresponds to SSL_CTX_set_mode.

pub fn set_tmp_dh(&mut self, dh: &DhRef<Params>) -> Result<(), ErrorStack>

Sets the parameters to be used during ephemeral Diffie-Hellman key exchange.

This corresponds to SSL_CTX_set_tmp_dh.

pub fn set_tmp_dh_callback<F>(&mut self, callback: F) where
F: Fn(&mut SslRef, bool, u32) -> Result<Dh<Params>, ErrorStack> + 'static + Sync + Send,

Sets the callback which will generate parameters to be used during ephemeral Diffie-Hellman key exchange.

The callback is provided with a reference to the Ssl for the session, as well as a boolean indicating if the selected cipher is export-grade, and the key length. The export and key length options are archaic and should be ignored in almost all cases.

This corresponds to SSL_CTX_set_tmp_dh_callback.

pub fn set_tmp_ecdh(&mut self, key: &EcKeyRef<Params>) -> Result<(), ErrorStack>

Sets the parameters to be used during ephemeral elliptic curve Diffie-Hellman key exchange.

This corresponds to SSL_CTX_set_tmp_ecdh.

pub fn set_tmp_ecdh_callback<F>(&mut self, callback: F) where
F: Fn(&mut SslRef, bool, u32) -> Result<EcKey<Params>, ErrorStack> + 'static + Sync + Send,

Sets the callback which will generate parameters to be used during ephemeral elliptic curve Diffie-Hellman key exchange.

The callback is provided with a reference to the Ssl for the session, as well as a boolean indicating if the selected cipher is export-grade, and the key length. The export and key length options are archaic and should be ignored in almost all cases.

Requires OpenSSL 1.0.1 or 1.0.2.

This corresponds to SSL_CTX_set_tmp_ecdh_callback.

pub fn set_default_verify_paths(&mut self) -> Result<(), ErrorStack>

Use the default locations of trusted certificates for verification.

These locations are read from the SSL_CERT_FILE and SSL_CERT_DIR environment variables if present, or defaults specified at OpenSSL build time otherwise.

This corresponds to SSL_CTX_set_default_verify_paths.

pub fn set_ca_file<P: AsRef<Path>>(&mut self, file: P) -> Result<(), ErrorStack>

Loads trusted root certificates from a file.

The file should contain a sequence of PEM-formatted CA certificates.

This corresponds to SSL_CTX_load_verify_locations.

pub fn set_client_ca_list(&mut self, list: Stack<X509Name>)

Sets the list of CA names sent to the client.

The CA certificates must still be added to the trust root - they are not automatically set as trusted by this method.

This corresponds to SSL_CTX_set_client_CA_list.

pub fn set_session_id_context(
&mut self,
sid_ctx: &[u8]
) -> Result<(), ErrorStack>

Set the context identifier for sessions.

This value identifies the server's session cache to clients, telling them when they're able to reuse sessions. It should be be set to a unique value per server, unless multiple servers share a session cache.

This value should be set when using client certificates, or each request will fail its handshake and need to be restarted.

This corresponds to SSL_CTX_set_session_id_context.

pub fn set_certificate_file<P: AsRef<Path>>(
&mut self,
file: P,
file_type: SslFiletype
) -> Result<(), ErrorStack>

Loads a leaf certificate from a file.

Only a single certificate will be loaded - use add_extra_chain_cert to add the remainder of the certificate chain, or set_certificate_chain_file to load the entire chain from a single file.

This corresponds to SSL_CTX_use_certificate_file.

pub fn set_certificate_chain_file<P: AsRef<Path>>(
&mut self,
file: P
) -> Result<(), ErrorStack>

Loads a certificate chain from a file.

The file should contain a sequence of PEM-formatted certificates, the first being the leaf certificate, and the remainder forming the chain of certificates up to and including the trusted root certificate.

This corresponds to SSL_CTX_use_certificate_chain_file.

pub fn set_certificate(&mut self, cert: &X509Ref) -> Result<(), ErrorStack>

Sets the leaf certificate.

Use add_extra_chain_cert to add the remainder of the certificate chain.

This corresponds to SSL_CTX_use_certificate.

pub fn add_extra_chain_cert(&mut self, cert: X509) -> Result<(), ErrorStack>

Appends a certificate to the certificate chain.

This chain should contain all certificates necessary to go from the certificate specified by set_certificate to a trusted root.

This corresponds to SSL_CTX_add_extra_chain_cert.

pub fn set_private_key_file<P: AsRef<Path>>(
&mut self,
file: P,
file_type: SslFiletype
) -> Result<(), ErrorStack>

Loads the private key from a file.

This corresponds to SSL_CTX_use_PrivateKey_file.

pub fn set_private_key<T>(&mut self, key: &PKeyRef<T>) -> Result<(), ErrorStack> where
T: HasPrivate,

Sets the private key.

This corresponds to SSL_CTX_use_PrivateKey.

pub fn set_cipher_list(&mut self, cipher_list: &str) -> Result<(), ErrorStack>

Sets the list of supported ciphers for protocols before TLSv1.3.

The set_ciphersuites method controls the cipher suites for TLSv1.3.

See ciphers for details on the format.

This corresponds to SSL_CTX_set_cipher_list.

pub fn set_options(&mut self, option: SslOptions) -> SslOptions

Sets the options used by the context, returning the old set.

This corresponds to SSL_CTX_set_options.

Note

This enables the specified options, but does not disable unspecified options. Use clear_options for that.

pub fn options(&self) -> SslOptions

Returns the options used by the context.

This corresponds to SSL_CTX_get_options.

pub fn clear_options(&mut self, option: SslOptions) -> SslOptions

Clears the options used by the context, returning the old set.

This corresponds to SSL_CTX_clear_options.

pub fn set_tlsext_use_srtp(&mut self, protocols: &str) -> Result<(), ErrorStack>

Enables the DTLS extension "use_srtp" as defined in RFC5764.

This corresponds to SSL_CTX_set_tlsext_use_srtp.

pub fn check_private_key(&self) -> Result<(), ErrorStack>

Checks for consistency between the private key and certificate.

This corresponds to SSL_CTX_check_private_key.

pub fn cert_store(&self) -> &X509StoreBuilderRef

Returns a shared reference to the context's certificate store.

This corresponds to SSL_CTX_get_cert_store.

pub fn cert_store_mut(&mut self) -> &mut X509StoreBuilderRef

Returns a mutable reference to the context's certificate store.

This corresponds to SSL_CTX_get_cert_store.

pub fn set_status_callback<F>(&mut self, callback: F) -> Result<(), ErrorStack> where
F: Fn(&mut SslRef) -> Result<bool, ErrorStack> + 'static + Sync + Send,

Sets the callback dealing with OCSP stapling.

On the client side, this callback is responsible for validating the OCSP status response returned by the server. The status may be retrieved with the SslRef::ocsp_status method. A response of Ok(true) indicates that the OCSP status is valid, and a response of Ok(false) indicates that the OCSP status is invalid and the handshake should be terminated.

On the server side, this callback is resopnsible for setting the OCSP status response to be returned to clients. The status may be set with the SslRef::set_ocsp_status method. A response of Ok(true) indicates that the OCSP status should be returned to the client, and Ok(false) indicates that the status should not be returned to the client.

This corresponds to SSL_CTX_set_tlsext_status_cb.

pub fn set_psk_client_callback<F>(&mut self, callback: F) where
F: Fn(&mut SslRef, Option<&[u8]>, &mut [u8], &mut [u8]) -> Result<usize, ErrorStack> + 'static + Sync + Send,

Sets the callback for providing an identity and pre-shared key for a TLS-PSK client.

The callback will be called with the SSL context, an identity hint if one was provided by the server, a mutable slice for each of the identity and pre-shared key bytes. The identity must be written as a null-terminated C string.

This corresponds to SSL_CTX_set_psk_client_callback.

pub fn set_psk_callback<F>(&mut self, callback: F) where
F: Fn(&mut SslRef, Option<&[u8]>, &mut [u8], &mut [u8]) -> Result<usize, ErrorStack> + 'static + Sync + Send,

Deprecated since 0.10.10
: renamed to set_psk_client_callback

pub fn set_psk_server_callback<F>(&mut self, callback: F) where
F: Fn(&mut SslRef, Option<&[u8]>, &mut [u8]) -> Result<usize, ErrorStack> + 'static + Sync + Send,

Sets the callback for providing an identity and pre-shared key for a TLS-PSK server.

The callback will be called with the SSL context, an identity provided by the client, and, a mutable slice for the pre-shared key bytes. The callback returns the number of bytes in the pre-shared key.

This corresponds to SSL_CTX_set_psk_server_callback.

pub fn set_new_session_callback<F>(&mut self, callback: F) where
F: Fn(&mut SslRef, SslSession) + 'static + Sync + Send,

Sets the callback which is called when new sessions are negotiated.

This can be used by clients to implement session caching. While in TLSv1.2 the session is available to access via SslRef::session immediately after the handshake completes, this is not the case for TLSv1.3. There, a session is not generally available immediately, and the server may provide multiple session tokens to the client over a single session. The new session callback is a portable way to deal with both cases.

Note that session caching must be enabled for the callback to be invoked, and it defaults off for clients. set_session_cache_mode controls that behavior.

This corresponds to SSL_CTX_sess_set_new_cb.

pub fn set_remove_session_callback<F>(&mut self, callback: F) where
F: Fn(&SslContextRef, &SslSessionRef) + 'static + Sync + Send,

Sets the callback which is called when sessions are removed from the context.

Sessions can be removed because they have timed out or because they are considered faulty.

This corresponds to SSL_CTX_sess_set_remove_cb.

pub unsafe fn set_get_session_callback<F>(&mut self, callback: F) where
F: Fn(&mut SslRef, &[u8]) -> Option<SslSession> + 'static + Sync + Send,

Sets the callback which is called when a client proposed to resume a session but it was not found in the internal cache.

The callback is passed a reference to the session ID provided by the client. It should return the session corresponding to that ID if available. This is only used for servers, not clients.

This corresponds to SSL_CTX_sess_set_get_cb.

Safety

The returned SslSession must not be associated with a different SslContext.

pub fn set_session_cache_mode(
&mut self,
mode: SslSessionCacheMode
) -> SslSessionCacheMode

Sets the session caching mode use for connections made with the context.

Returns the previous session caching mode.

This corresponds to SSL_CTX_set_session_cache_mode.

pub fn set_cookie_generate_cb<F>(&mut self, callback: F) where
F: Fn(&mut SslRef, &mut [u8]) -> Result<usize, ErrorStack> + 'static + Sync + Send,

Sets the callback for generating a DTLSv1 cookie

The callback will be called with the SSL context and a slice into which the cookie should be written. The callback should return the number of bytes written.

This corresponds to SSL_CTX_set_cookie_generate_cb.

pub fn set_cookie_verify_cb<F>(&mut self, callback: F) where
F: Fn(&mut SslRef, &[u8]) -> bool + 'static + Sync + Send,

Sets the callback for verifying a DTLSv1 cookie

The callback will be called with the SSL context and the cookie supplied by the client. It should return true if and only if the cookie is valid.

This corresponds to SSL_CTX_set_cookie_verify_cb.

pub fn set_ex_data<T>(&mut self, index: Index<SslContext, T>, data: T)

Sets the extra data at the specified index.

This can be used to provide data to callbacks registered with the context. Use the SslContext::new_ex_index method to create an Index.

This corresponds to SSL_CTX_set_ex_data.

pub fn build(self) -> SslContext

Consumes the builder, returning a new SslContext.