Struct openssl::ssl::SslSessionRef

pub struct SslSessionRef(_);

Reference to SslSession.

Methods

impl SslSessionRef

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn id(&self) -> &[u8]

Returns the SSL session ID.

This corresponds to SSL_SESSION_get_id.

pub fn master_key_len(&self) -> usize

Returns the length of the master key.

This corresponds to SSL_SESSION_get_master_key.

pub fn master_key(&self, buf: &mut [u8]) -> usize

Copies the master key into the provided buffer.

Returns the number of bytes written, or the size of the master key if the buffer is empty.

This corresponds to SSL_SESSION_get_master_key.