Struct openssl::ssl::SslStreamBuilder

pub struct SslStreamBuilder<S> { /* fields omitted */ }

A partially constructed SslStream, useful for unusual handshakes.

Methods

impl<S> SslStreamBuilder<S> where
S: Read + Write,

pub fn new(ssl: Ssl, stream: S) -> Self

Begin creating an SslStream atop stream

pub fn set_connect_state(&mut self)

Configure as an outgoing stream from a client.

This corresponds to SSL_set_connect_state.

pub fn set_accept_state(&mut self)

Configure as an incoming stream to a server.

This corresponds to SSL_set_accept_state.

pub fn connect(self) -> Result<SslStream<S>, HandshakeError<S>>

See Ssl::connect

pub fn accept(self) -> Result<SslStream<S>, HandshakeError<S>>

See Ssl::accept

pub fn handshake(self) -> Result<SslStream<S>, HandshakeError<S>>

Initiates the handshake.

This will fail if set_accept_state or set_connect_state was not called first.

This corresponds to SSL_do_handshake.

impl<S> SslStreamBuilder<S>

pub fn get_ref(&self) -> &S

Returns a shared reference to the underlying stream.

pub fn get_mut(&mut self) -> &mut S

Returns a mutable reference to the underlying stream.

Warning

It is inadvisable to read from or write to the underlying stream as it will most likely corrupt the SSL session.

pub fn ssl(&self) -> &SslRef

Returns a shared reference to the Ssl object associated with this builder.