Struct openssl::ssl::SslVerifyMode

pub struct SslVerifyMode {
    // some fields omitted
}

Methods

impl SslVerifyMode

fn empty() -> SslVerifyMode

Returns an empty set of flags.

fn all() -> SslVerifyMode

Returns the set containing all flags.

fn bits(&self) -> i32

Returns the raw value of the flags currently stored.

fn from_bits(bits: i32) -> Option<SslVerifyMode>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: i32) -> SslVerifyMode

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Returns true if all flags are currently set.

fn intersects(&self, other: SslVerifyMode) -> bool

Returns true if there are flags common to both self and other.

fn contains(&self, other: SslVerifyMode) -> bool

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: SslVerifyMode)

Inserts the specified flags in-place.

fn remove(&mut self, other: SslVerifyMode)

Removes the specified flags in-place.

fn toggle(&mut self, other: SslVerifyMode)

Toggles the specified flags in-place.