Function openssl::symm::encrypt_aead

pub fn encrypt_aead(
    t: Cipher, 
    key: &[u8], 
    iv: Option<&[u8]>, 
    aad: &[u8], 
    data: &[u8], 
    tag: &mut [u8]
) -> Result<Vec<u8>, ErrorStack>

Like encrypt, but for AEAD ciphers such as AES GCM.

Additional Authenticated Data can be provided in the aad field, and the authentication tag will be copied into the tag field.

The size of the tag buffer indicates the required size of the tag. While some ciphers support a range of tag sizes, it is recommended to pick the maximum size. For AES GCM, this is 16 bytes, for example.