Trait openssl::types::OpenSslType

pub trait OpenSslType: Sized {
    type CType;
    type Ref: OpenSslTypeRef<CType=Self::CType>;
    unsafe fn from_ptr(ptr: *mut Self::CType) -> Self;
}

A type implemented by wrappers over OpenSSL types.

This should not be implemented by anything outside of this crate; new methods may be added at any time.

Associated Types

type CType

The raw C type.

type Ref: OpenSslTypeRef<CType=Self::CType>

The type representing a reference to this type.

Required Methods

unsafe fn from_ptr(ptr: *mut Self::CType) -> Self

Constructs an instance of this type from its raw type.