Struct openssl::x509::extension::BasicConstraints

pub struct BasicConstraints { /* fields omitted */ }

An extension which indicates whether a certificate is a CA certificate.

Methods

impl BasicConstraints

pub fn new() -> BasicConstraints

Construct a new BasicConstraints extension.

pub fn critical(&mut self) -> &mut BasicConstraints

Sets the critical flag to true. The extension will be critical.

pub fn ca(&mut self) -> &mut BasicConstraints

Sets the ca flag to true.

pub fn pathlen(&mut self, pathlen: u32) -> &mut BasicConstraints

Sets the pathlen to an optional non-negative value. The pathlen is the maximum number of CAs that can appear below this one in a chain.

pub fn build(&self) -> Result<X509Extension, ErrorStack>

Return the BasicConstraints extension as an X509Extension.