Struct openssl::x509::extension::ExtendedKeyUsage

pub struct ExtendedKeyUsage { /* fields omitted */ }

An extension consisting of a list of usages indicating purposes for which the certificate public key can be used for.

Methods

impl ExtendedKeyUsage

pub fn new() -> ExtendedKeyUsage

Construct a new ExtendedKeyUsage extension.

pub fn critical(&mut self) -> &mut ExtendedKeyUsage

Sets the critical flag to true. The extension will be critical.

pub fn server_auth(&mut self) -> &mut ExtendedKeyUsage

Sets the serverAuth flag to true.

pub fn client_auth(&mut self) -> &mut ExtendedKeyUsage

Sets the clientAuth flag to true.

pub fn code_signing(&mut self) -> &mut ExtendedKeyUsage

Sets the codeSigning flag to true.

pub fn time_stamping(&mut self) -> &mut ExtendedKeyUsage

Sets the timeStamping flag to true.

pub fn ms_code_ind(&mut self) -> &mut ExtendedKeyUsage

Sets the msCodeInd flag to true.

pub fn ms_code_com(&mut self) -> &mut ExtendedKeyUsage

Sets the msCodeCom flag to true.

pub fn ms_ctl_sign(&mut self) -> &mut ExtendedKeyUsage

Sets the msCTLSign flag to true.

pub fn ms_sgc(&mut self) -> &mut ExtendedKeyUsage

Sets the msSGC flag to true.

pub fn ms_efs(&mut self) -> &mut ExtendedKeyUsage

Sets the msEFS flag to true.

pub fn ns_sgc(&mut self) -> &mut ExtendedKeyUsage

Sets the nsSGC flag to true.

pub fn other(&mut self, other: &str) -> &mut ExtendedKeyUsage

Sets a flag not already defined.

pub fn build(&self) -> Result<X509Extension, ErrorStack>

Return the ExtendedKeyUsage extension as an X509Extension.