Struct openssl::x509::Extensions

struct Extensions {
    extensions: Vec<Extension>,
    indexes: HashMap<ExtensionType, usize>,
}

A collection of X.509 extensions.

Upholds the invariant that a certificate MUST NOT include more than one instance of a particular extension, according to RFC 3280 4.2. Also ensures that extensions are added to the certificate during signing in the order they were inserted, which is required for certain extensions like SubjectKeyIdentifier and AuthorityKeyIdentifier.

Fields

extensions

The extensions contained in the collection.

indexes

A map of used to keep track of added extensions and their indexes in self.extensions.

Methods

impl Extensions

fn new() -> Extensions

Creates a new Extensions.

fn add(&mut self, ext: Extension)

Adds a new Extension, replacing any existing one of the same ExtensionType.

fn iter(&self) -> ExtensionsIter

Returns an ExtensionsIter for the collection.