Struct openssl::x509::X509Ref

pub struct X509Ref(_);

Reference to X509.

Implementations

impl X509Ref

pub fn subject_name(&self) -> &X509NameRef

Returns this certificate’s subject name.

This corresponds to X509_get_subject_name.

pub fn subject_name_hash(&self) -> u32

Returns the hash of the certificates subject

This corresponds to X509_subject_name_hash.

pub fn issuer_name(&self) -> &X509NameRef

Returns this certificate’s issuer name.

This corresponds to X509_get_issuer_name.

pub fn subject_alt_names(&self) -> Option<Stack<GeneralName>>

Returns this certificate’s subject alternative name entries, if they exist.

This corresponds to X509_get_ext_d2i called with NID_subject_alt_name.

pub fn issuer_alt_names(&self) -> Option<Stack<GeneralName>>

Returns this certificate’s issuer alternative name entries, if they exist.

This corresponds to X509_get_ext_d2i called with NID_issuer_alt_name.

pub fn authority_info(&self) -> Option<Stack<AccessDescription>>

Returns this certificate’s authority information access entries, if they exist.

This corresponds to X509_get_ext_d2i called with NID_info_access.

pub fn public_key(&self) -> Result<PKey<Public>, ErrorStack>

pub fn digest(
&self,
hash_type: MessageDigest
) -> Result<DigestBytes, ErrorStack>

Returns a digest of the DER representation of the certificate.

This corresponds to X509_digest.

pub fn fingerprint(
&self,
hash_type: MessageDigest
) -> Result<Vec<u8>, ErrorStack>

👎 Deprecated since 0.10.9:
renamed to digest

pub fn not_after(&self) -> &Asn1TimeRef

Returns the certificate’s Not After validity period.

pub fn not_before(&self) -> &Asn1TimeRef

Returns the certificate’s Not Before validity period.

pub fn signature(&self) -> &Asn1BitStringRef

Returns the certificate’s signature

pub fn signature_algorithm(&self) -> &X509AlgorithmRef

Returns the certificate’s signature algorithm.

pub fn ocsp_responders(&self) -> Result<Stack<OpensslString>, ErrorStack>

Returns the list of OCSP responder URLs specified in the certificate’s Authority Information Access field.

pub fn issued(&self, subject: &X509Ref) -> X509VerifyResult

Checks that this certificate issued subject.

pub fn version(&self) -> i32

Returns certificate version. If this certificate has no explicit version set, it defaults to version 1.

Note that 0 return value stands for version 1, 1 for version 2 and so on.

This corresponds to X509_get_version.

pub fn verify<T>(&self, key: &PKeyRef<T>) -> Result<bool, ErrorStack> where
T: HasPublic,

Check if the certificate is signed using the given public key.

Only the signature is checked: no other checks (such as certificate chain validity) are performed.

Returns true if verification succeeds.

This corresponds to [`X509_verify“].

pub fn serial_number(&self) -> &Asn1IntegerRef

Returns this certificate’s serial number.

This corresponds to X509_get_serialNumber.

pub fn to_pem(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the certificate into a PEM-encoded X509 structure.

The output will have a header of -----BEGIN CERTIFICATE-----.

This corresponds to PEM_write_bio_X509.

pub fn to_der(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the certificate into a DER-encoded X509 structure.

This corresponds to i2d_X509.