Struct openssl::x509::X509StoreContext

pub struct X509StoreContext(_);

An X509 certificate store context.

Methods

impl X509StoreContext

pub fn ssl_idx() -> Result<Index<X509StoreContext, SslRef>, ErrorStack>

Returns the index which can be used to obtain a reference to the Ssl associated with a context.