openssl::aes
Function aes_ige

pub fn aes_ige(
    in_: &[u8],
    out: &mut [u8],
    key: &AesKey,
    iv: &mut [u8],
    mode: Mode,
)

Performs AES IGE encryption or decryption

AES IGE (Infinite Garble Extension) is a form of AES block cipher utilized in OpenSSL. Infinite Garble refers to propagating forward errors. IGE, like other block ciphers implemented for AES requires an initialization vector. The IGE mode allows a stream of blocks to be encrypted or decrypted without having the entire plaintext available. For more information, visit AES IGE Encryption.

This block cipher uses 16 byte blocks. The rust implementation will panic if the input or output does not meet this 16-byte boundary. Attention must be made in this low level implementation to pad the value to the 128-bit boundary.

Panics

Panics if in_ is not the same length as out, if that length is not a multiple of 16, or if iv is not at least 32 bytes.

This corresponds to AES_ige_encrypt.