Struct openssl::asn1::Asn1Time

pub struct Asn1Time(_);

Time storage and comparison

Asn1Time should be used to store and share time information using certificates. If Asn1Time is set using a string, it must be in either YYMMDDHHMMSSZ, YYYYMMDDHHMMSSZ, or another ASN.1 format.

ASN_TIME_set documentation at OpenSSL explains the ASN.1 implementation used by OpenSSL.

Implementations

impl Asn1Time

pub fn days_from_now(days: u32) -> Result<Asn1Time, ErrorStack>

Creates a new time on specified interval in days from now

pub fn from_unix(time: time_t) -> Result<Asn1Time, ErrorStack>

Creates a new time from the specified time_t value

This corresponds to ASN1_TIME_set.

pub fn from_str(s: &str) -> Result<Asn1Time, ErrorStack>

Creates a new time corresponding to the specified ASN1 time string.

This corresponds to ASN1_TIME_set_string.

pub fn from_str_x509(s: &str) -> Result<Asn1Time, ErrorStack>

Creates a new time corresponding to the specified X509 time string.

Requires OpenSSL 1.1.1 or newer.

This corresponds to ASN1_TIME_set_string_X509.