Struct openssl::asn1::Asn1TimeRef

pub struct Asn1TimeRef(_);

Reference to an Asn1Time

Implementations

impl Asn1TimeRef

pub fn diff(&self, compare: &Self) -> Result<TimeDiff, ErrorStack>

Find difference between two times

This corresponds to ASN1_TIME_diff.

pub fn compare(&self, other: &Self) -> Result<Ordering, ErrorStack>

Compare two times

This corresponds to ASN1_TIME_compare but is implemented using diff so that it is also supported on older versions of OpenSSL.