Struct openssl::bn::BigNum

pub struct BigNum(_);

Dynamically sized large number impelementation

Perform large number mathematics. Create a new BigNum with new. Perform stanard mathematics on large numbers using methods from Dref<Target = BigNumRef>

OpenSSL documenation at BN_new.

Examples

use openssl::bn::BigNum;
let little_big = BigNum::from_u32(std::u32::MAX)?;
assert_eq!(*&little_big.num_bytes(), 4);

Methods

impl BigNum

fn new() -> Result<BigNum, ErrorStack>

Creates a new BigNum with the value 0.

fn from_u32(n: u32) -> Result<BigNum, ErrorStack>

Creates a new BigNum with the given value.

OpenSSL documentation at BN_set_word

fn from_dec_str(s: &str) -> Result<BigNum, ErrorStack>

Creates a BigNum from a decimal string.

OpenSSL documentation at BN_dec2bn

fn from_hex_str(s: &str) -> Result<BigNum, ErrorStack>

Creates a BigNum from a hexadecimal string.

OpenSSL documentation at BN_hex2bn

fn get_rfc2409_prime_768() -> Result<BigNum, ErrorStack>

Returns a constant used in IKE as defined in RFC 2409. This prime number is in the order of magnitude of 2 ^ 768. This number is used during calculated key exchanges such as Diffie-Hellman. This number is labeled Oakley group id 1.

OpenSSL documentation at BN_get_rfc2409_prime_768

fn get_rfc2409_prime_1024() -> Result<BigNum, ErrorStack>

Returns a constant used in IKE as defined in RFC 2409. This prime number is in the order of magnitude of 2 ^ 1024. This number is used during calculated key exchanges such as Diffie-Hellman. This number is labeled Oakly group 2.

OpenSSL documentation at BN_get_rfc2409_prime_1024

fn get_rfc3526_prime_1536() -> Result<BigNum, ErrorStack>

Returns a constant used in IKE as defined in RFC 3526. The prime is in the order of magnitude of 2 ^ 1536. This number is used during calculated key exchanges such as Diffie-Hellman. This number is labeled MODP group 5.

OpenSSL documentation at BN_get_rfc3526_prime_1536

fn get_rfc3526_prime_2048() -> Result<BigNum, ErrorStack>

Returns a constant used in IKE as defined in RFC 3526. The prime is in the order of magnitude of 2 ^ 2048. This number is used during calculated key exchanges such as Diffie-Hellman. This number is labeled MODP group 14.

OpenSSL documentation at BN_get_rfc3526_prime_2048

fn get_rfc3526_prime_3072() -> Result<BigNum, ErrorStack>

Returns a constant used in IKE as defined in RFC 3526. The prime is in the order of magnitude of 2 ^ 3072. This number is used during calculated key exchanges such as Diffie-Hellman. This number is labeled MODP group 15.

OpenSSL documentation at BN_get_rfc3526_prime_3072

fn get_rfc3526_prime_4096() -> Result<BigNum, ErrorStack>

Returns a constant used in IKE as defined in RFC 3526. The prime is in the order of magnitude of 2 ^ 4096. This number is used during calculated key exchanges such as Diffie-Hellman. This number is labeled MODP group 16.

OpenSSL documentation at BN_get_rfc3526_prime_4096

fn get_rfc3526_prime_6144() -> Result<BigNum, ErrorStack>

Returns a constant used in IKE as defined in RFC 3526. The prime is in the order of magnitude of 2 ^ 6144. This number is used during calculated key exchanges such as Diffie-Hellman. This number is labeled MODP group 17.

OpenSSL documentation at BN_get_rfc3526_prime_6144

fn get_rfc3526_prime_8192() -> Result<BigNum, ErrorStack>

Returns a constant used in IKE as defined in RFC 3526. The prime is in the order of magnitude of 2 ^ 8192. This number is used during calculated key exchanges such as Diffie-Hellman. This number is labeled MODP group 18.

OpenSSL documentation at BN_get_rfc3526_prime_8192

fn from_slice(n: &[u8]) -> Result<BigNum, ErrorStack>

Creates a new BigNum from an unsigned, big-endian encoded number of arbitrary length.

OpenSSL documentation at BN_bin2bn

let bignum = BigNum::from_slice(&[0x12, 0x00, 0x34]).unwrap();

assert_eq!(bignum, BigNum::from_u32(0x120034).unwrap());