Struct openssl::bn::BigNumRef

pub struct BigNumRef<'a>(_, _);

A borrowed, signed, arbitrary-precision integer.

Methods

impl<'a> BigNumRef<'a>

unsafe fn from_ptr(handle: *mut BIGNUM) -> BigNumRef<'a>

fn checked_sqr(&self) -> Result<BigNum, ErrorStack>

Returns the square of self.

let ref n = BigNum::new_from(10).unwrap();
let squared = BigNum::new_from(100).unwrap();

assert_eq!(n.checked_sqr().unwrap(), squared);
assert_eq!(n * n, squared);

fn checked_nnmod(&self, n: &BigNumRef) -> Result<BigNum, ErrorStack>

Returns the unsigned remainder of the division self / n.

fn checked_mod_add(&self, a: &BigNumRef, n: &BigNumRef) -> Result<BigNum, ErrorStack>

Equivalent to (self + a) mod n.

let ref s = BigNum::new_from(10).unwrap();
let ref a = BigNum::new_from(20).unwrap();
let ref n = BigNum::new_from(29).unwrap();
let result = BigNum::new_from(1).unwrap();

assert_eq!(s.checked_mod_add(a, n).unwrap(), result);

fn checked_mod_sub(&self, a: &BigNumRef, n: &BigNumRef) -> Result<BigNum, ErrorStack>

Equivalent to (self - a) mod n.

fn checked_mod_mul(&self, a: &BigNumRef, n: &BigNumRef) -> Result<BigNum, ErrorStack>

Equivalent to (self * a) mod n.

fn checked_mod_sqr(&self, n: &BigNumRef) -> Result<BigNum, ErrorStack>

Equivalent to self² mod n.

fn checked_exp(&self, p: &BigNumRef) -> Result<BigNum, ErrorStack>

Raises self to the pth power.

fn checked_mod_exp(&self, p: &BigNumRef, n: &BigNumRef) -> Result<BigNum, ErrorStack>

Equivalent to self.checked_exp(p) mod n.

fn checked_mod_inv(&self, n: &BigNumRef) -> Result<BigNum, ErrorStack>

Calculates the modular multiplicative inverse of self modulo n, that is, an integer r such that (self * r) % n == 1.

fn add_word(&mut self, w: c_ulong) -> Result<(), ErrorStack>

Add an unsigned long to self. This is more efficient than adding a BigNum.

fn sub_word(&mut self, w: c_ulong) -> Result<(), ErrorStack>

fn mul_word(&mut self, w: c_ulong) -> Result<(), ErrorStack>

fn div_word(&mut self, w: c_ulong) -> Result<c_ulong, ErrorStack>

fn mod_word(&self, w: c_ulong) -> Result<c_ulong, ErrorStack>

fn checked_gcd(&self, a: &BigNumRef) -> Result<BigNum, ErrorStack>

Computes the greatest common denominator of self and a.

fn is_prime(&self, checks: i32) -> Result<bool, ErrorStack>

Checks whether self is prime.

Performs a Miller-Rabin probabilistic primality test with checks iterations.

Return Value

Returns true if self is prime with an error probability of less than 0.25 ^ checks.

fn is_prime_fast(&self, checks: i32, do_trial_division: bool) -> Result<bool, ErrorStack>

Checks whether self is prime with optional trial division.

If do_trial_division is true, first performs trial division by a number of small primes. Then, like is_prime, performs a Miller-Rabin probabilistic primality test with checks iterations.

Return Value

Returns true if self is prime with an error probability of less than 0.25 ^ checks.

fn checked_rand_in_range(&self) -> Result<BigNum, ErrorStack>

Generates a cryptographically strong pseudo-random BigNum r in the range 0 <= r < self.

fn checked_pseudo_rand_in_range(&self) -> Result<BigNum, ErrorStack>

The cryptographically weak counterpart to checked_rand_in_range.

fn set_bit(&mut self, n: i32) -> Result<(), ErrorStack>

Sets bit n. Equivalent to self |= (1 << n).

When setting a bit outside of self, it is expanded.

fn clear_bit(&mut self, n: i32) -> Result<(), ErrorStack>

Clears bit n, setting it to 0. Equivalent to self &= ~(1 << n).

When clearing a bit outside of self, an error is returned.

fn is_bit_set(&self, n: i32) -> bool

Returns true if the nth bit of self is set to 1, false otherwise.

fn mask_bits(&mut self, n: i32) -> Result<(), ErrorStack>

Truncates self to the lowest n bits.

An error occurs if self is already shorter than n bits.

fn checked_shl1(&self) -> Result<BigNum, ErrorStack>

Returns self, shifted left by 1 bit. self may be negative.

let ref s = BigNum::new_from(0b0100).unwrap();
let result = BigNum::new_from(0b1000).unwrap();

assert_eq!(s.checked_shl1().unwrap(), result);

let ref s = -BigNum::new_from(8).unwrap();
let result = -BigNum::new_from(16).unwrap();

// (-8) << 1 == -16
assert_eq!(s.checked_shl1().unwrap(), result);

fn checked_shr1(&self) -> Result<BigNum, ErrorStack>

Returns self, shifted right by 1 bit. self may be negative.

fn checked_add(&self, a: &BigNumRef) -> Result<BigNum, ErrorStack>

fn checked_sub(&self, a: &BigNumRef) -> Result<BigNum, ErrorStack>

fn checked_mul(&self, a: &BigNumRef) -> Result<BigNum, ErrorStack>

fn checked_div(&self, a: &BigNumRef) -> Result<BigNum, ErrorStack>

fn checked_mod(&self, a: &BigNumRef) -> Result<BigNum, ErrorStack>

fn checked_shl(&self, a: &i32) -> Result<BigNum, ErrorStack>

fn checked_shr(&self, a: &i32) -> Result<BigNum, ErrorStack>

fn to_owned(&self) -> Result<BigNum, ErrorStack>

fn negate(&mut self)

Inverts the sign of self.

let mut s = BigNum::new_from(8).unwrap();

s.negate();
assert_eq!(s, -BigNum::new_from(8).unwrap());
s.negate();
assert_eq!(s, BigNum::new_from(8).unwrap());

fn abs_cmp(&self, oth: &BigNumRef) -> Ordering

Compare the absolute values of self and oth.

let s = -BigNum::new_from(8).unwrap();
let o = BigNum::new_from(8).unwrap();

assert_eq!(s.abs_cmp(&o), Ordering::Equal);

fn is_negative(&self) -> bool

fn num_bits(&self) -> i32

Returns the number of significant bits in self.

fn num_bytes(&self) -> i32

Returns the size of self in bytes.

fn as_ptr(&self) -> *mut BIGNUM

fn to_vec(&self) -> Vec<u8>

Returns a big-endian byte vector representation of the absolute value of self.

self can be recreated by using new_from_slice.

let s = -BigNum::new_from(4543).unwrap();
let r = BigNum::new_from(4543).unwrap();

let s_vec = s.to_vec();
assert_eq!(BigNum::new_from_slice(&s_vec).unwrap(), r);

fn to_dec_str(&self) -> String

Returns a decimal string representation of self.

let s = -BigNum::new_from(12345).unwrap();

assert_eq!(s.to_dec_str(), "-12345");

fn to_hex_str(&self) -> String

Returns a hexadecimal string representation of self.

let s = -BigNum::new_from(0x99ff).unwrap();

assert_eq!(s.to_hex_str(), "-99FF");