Struct openssl::cms::CmsContentInfo

pub struct CmsContentInfo(_);

High level CMS wrapper

CMS supports nesting various types of data, including signatures, certificates, encrypted data, smime messages (encrypted email), and data digest. The ContentInfo content type is the encapsulation of all those content types. RFC 5652 describes CMS and OpenSSL follows this RFC's implmentation.

Methods

impl CmsContentInfo

pub fn smime_read_cms(smime: &[u8]) -> Result<CmsContentInfo, ErrorStack>

Parses a smime formatted vec of bytes into a CmsContentInfo.

OpenSSL documentation at SMIME_read_CMS