Struct openssl::crypto::hash::Hasher

pub struct Hasher {
    // some fields omitted
}

Provides message digest (hash) computation.

Examples

Calculate a hash in one go.

use openssl::crypto::hash::{hash, Type};
let data = b"\x42\xF4\x97\xE0";
let spec = b"\x7c\x43\x0f\x17\x8a\xef\xdf\x14\x87\xfe\xe7\x14\x4e\x96\x41\xe2";
let res = hash(Type::MD5, data);
assert_eq!(res, spec);

Use the Writer trait to supply the input in chunks.

use std::old_io::Writer;
use openssl::crypto::hash::{Hasher, Type};
let data = [b"\x42\xF4", b"\x97\xE0"];
let spec = b"\x7c\x43\x0f\x17\x8a\xef\xdf\x14\x87\xfe\xe7\x14\x4e\x96\x41\xe2";
let mut h = Hasher::new(Type::MD5);
h.write_all(data[0]);
h.write_all(data[1]);
let res = h.finish();
assert_eq!(res, spec);

Warning

Don't actually use MD5 and SHA-1 hashes, they're not secure anymore.

Don't ever hash passwords, use crypto::pkcs5 or bcrypt/scrypt instead.

Methods

impl Hasher

fn new(ty: Type) -> Hasher

Creates a new Hasher with the specified hash type.

fn finish(&mut self) -> Vec<u8>

Returns the hash of the data written since creation or the last finish and resets the hasher.