Struct openssl::crypto::symm::Crypter

pub struct Crypter {
    // some fields omitted
}

Represents a symmetric cipher context.

Methods

impl Crypter

fn new(t: Type) -> Crypter

fn pad(&self, padding: bool)

Enables or disables padding. If padding is disabled, total amount of data encrypted must be a multiple of block size.

fn init(&self, mode: Mode, key: &[u8], iv: &[u8])

Initializes this crypter.

fn update(&self, data: &[u8]) -> Vec<u8>

Update this crypter with more data to encrypt or decrypt. Returns encrypted or decrypted bytes.

fn finalize(&self) -> Vec<u8>

Finish crypting. Returns the remaining partial block of output, if any.