Struct openssl::derive::Deriver

pub struct Deriver<'a>(_, _);

A type used to derive a shared secret between two keys.

Methods

impl<'a> Deriver<'a>

pub fn new<T>(key: &'a PKeyRef<T>) -> Result<Deriver<'a>, ErrorStack> where
T: HasPrivate,

Creates a new Deriver using the provided private key.

This corresponds to EVP_PKEY_derive_init.

pub fn set_peer<T>(&mut self, key: &'a PKeyRef<T>) -> Result<(), ErrorStack> where
T: HasPublic,

Sets the peer key used for secret derivation.

This corresponds to EVP_PKEY_derive_set_peer:

pub fn len(&mut self) -> Result<usize, ErrorStack>

Returns the size of the shared secret.

It can be used to size the buffer passed to Deriver::derive.

This corresponds to EVP_PKEY_derive.

pub fn derive(&mut self, buf: &mut [u8]) -> Result<usize, ErrorStack>

Derives a shared secret between the two keys, writing it into the buffer.

Returns the number of bytes written.

This corresponds to EVP_PKEY_derive.

pub fn derive_to_vec(&mut self) -> Result<Vec<u8>, ErrorStack>

A convenience function which derives a shared secret and returns it in a new buffer.

This simply wraps Deriver::len and Deriver::derive.