Struct openssl::dsa::Dsa

pub struct Dsa(_);

Methods

impl Dsa

fn generate(bits: u32) -> Result<Dsa, ErrorStack>

Generate a DSA key pair.

fn private_key_from_pem(buf: &[u8]) -> Result<Dsa, ErrorStack>

Reads a DSA private key from PEM formatted data.

fn private_key_from_pem_cb<F>(buf: &[u8], pass_cb: F) -> Result<Dsa, ErrorStack> where F: FnOnce(&mut [c_char]) -> usize

Read a private key from PEM supplying a password callback to be invoked if the private key is encrypted.

The callback will be passed the password buffer and should return the number of characters placed into the buffer.

fn public_key_from_pem(buf: &[u8]) -> Result<Dsa, ErrorStack>

Reads an DSA public key from PEM formatted data.