Module openssl::ec

Elliptic Curve

Cryptography relies on the difficulty of solving mathematical problems, such as the factor of large integers composed of two large prime numbers and the discrete logarithm of a random elliptic curve. This module provides low-level features of the latter. Elliptic Curve protocols can provide the same security with smaller keys.

There are 2 forms of elliptic curves, Fp and F2^m. These curves use irreducible trinomial or pentanomial. Being a generic interface to a wide range of algorithms, the curves are generally referenced by EcGroup. There are many built-in groups found in Nid.

OpenSSL Wiki explains the fields and curves in detail at [Elliptic Curve Cryptography].

Structs

Asn1Flag

Named Curve or Explicit

EcGroup

Describes the curve

EcGroupRef

Reference to EcGroup

EcKey

Public and optional private key on the given curve.

EcKeyRef

A reference to an EcKey.

EcPoint

Represents a point on the curve

EcPointRef

A reference a borrowed EcPoint.

PointConversionForm

Compressed or Uncompressed conversion