Struct openssl::ec::EcKey

pub struct EcKey<T>(_, _);

Public and optional Private key on the given curve

OpenSSL documentation at EC_KEY_new

Implementations

impl EcKey<Params>

pub fn from_curve_name(nid: Nid) -> Result<EcKey<Params>, ErrorStack>

Constructs an EcKey corresponding to a known curve.

It will not have an associated public or private key. This kind of key is primarily useful to be provided to the set_tmp_ecdh methods on Ssl and SslContextBuilder.

OpenSSL documentation at EC_KEY_new_by_curve_name

pub fn from_group(group: &EcGroupRef) -> Result<EcKey<Params>, ErrorStack>

Constructs an EcKey corresponding to a curve.

This corresponds to EC_KEY_set_group.

impl EcKey<Public>

pub fn from_public_key(
group: &EcGroupRef,
public_key: &EcPointRef
) -> Result<EcKey<Public>, ErrorStack>

Constructs an EcKey from the specified group with the associated EcPoint, public_key.

This will only have the associated public_key.

Example

use openssl::bn::BigNumContext;
use openssl::ec::*;
use openssl::nid::Nid;
use openssl::pkey::PKey;

// get bytes from somewhere, i.e. this will not produce a valid key
let public_key: Vec<u8> = vec![];

// create an EcKey from the binary form of a EcPoint
let group = EcGroup::from_curve_name(Nid::SECP256K1).unwrap();
let mut ctx = BigNumContext::new().unwrap();
let point = EcPoint::from_bytes(&group, &public_key, &mut ctx).unwrap();
let key = EcKey::from_public_key(&group, &point);

pub fn from_public_key_affine_coordinates(
group: &EcGroupRef,
x: &BigNumRef,
y: &BigNumRef
) -> Result<EcKey<Public>, ErrorStack>

Constructs a public key from its affine coordinates.

pub fn public_key_from_pem(pem: &[u8]) -> Result<EcKey<Public>, ErrorStack>

Decodes a PEM-encoded SubjectPublicKeyInfo structure containing a EC key.

The input should have a header of -----BEGIN PUBLIC KEY-----.

This corresponds to PEM_read_bio_EC_PUBKEY.

pub fn public_key_from_der(der: &[u8]) -> Result<EcKey<Public>, ErrorStack>

Decodes a DER-encoded SubjectPublicKeyInfo structure containing a EC key.

This corresponds to d2i_EC_PUBKEY.

impl EcKey<Private>

pub fn generate(group: &EcGroupRef) -> Result<EcKey<Private>, ErrorStack>

Generates a new public/private key pair on the specified curve.

pub fn from_private_components(
group: &EcGroupRef,
private_number: &BigNumRef,
public_key: &EcPointRef
) -> Result<EcKey<Private>, ErrorStack>

Constructs an public/private key pair given a curve, a private key and a public key point.

pub fn private_key_from_pem(pem: &[u8]) -> Result<EcKey<Private>, ErrorStack>

Deserializes a private key from a PEM-encoded ECPrivateKey structure.

The input should have a header of -----BEGIN EC PRIVATE KEY-----.

This corresponds to PEM_read_bio_ECPrivateKey.

pub fn private_key_from_pem_passphrase(
pem: &[u8],
passphrase: &[u8]
) -> Result<EcKey<Private>, ErrorStack>

Deserializes a private key from a PEM-encoded encrypted ECPrivateKey structure.

The input should have a header of -----BEGIN EC PRIVATE KEY-----.

This corresponds to PEM_read_bio_ECPrivateKey.

pub fn private_key_from_pem_callback<F>(
pem: &[u8],
callback: F
) -> Result<EcKey<Private>, ErrorStack> where
F: FnOnce(&mut [u8]) -> Result<usize, ErrorStack>,

Deserializes a private key from a PEM-encoded encrypted ECPrivateKey structure.

The callback should fill the password into the provided buffer and return its length.

The input should have a header of -----BEGIN EC PRIVATE KEY-----.

This corresponds to PEM_read_bio_ECPrivateKey.

pub fn private_key_from_der(der: &[u8]) -> Result<EcKey<Private>, ErrorStack>

Decodes a DER-encoded elliptic curve private key structure.

This corresponds to d2i_ECPrivateKey.