Struct openssl::ec::EcPointRef

pub struct EcPointRef(_);

Reference to EcPoint

Methods

impl EcPointRef

pub fn add(
&mut self,
group: &EcGroupRef,
a: &EcPointRef,
b: &EcPointRef,
ctx: &mut BigNumContextRef
) -> Result<(), ErrorStack>

Computes a + b, storing the result in self.

OpenSSL documentation at EC_POINT_add

pub fn mul(
&mut self,
group: &EcGroupRef,
q: &EcPointRef,
m: &BigNumRef,
ctx: &BigNumContextRef
) -> Result<(), ErrorStack>

Computes q * m, storing the result in self.

OpenSSL documentation at EC_POINT_mul

pub fn mul_generator(
&mut self,
group: &EcGroupRef,
n: &BigNumRef,
ctx: &BigNumContextRef
) -> Result<(), ErrorStack>

Computes generator * n, storing the result ing self.

pub fn mul_full(
&mut self,
group: &EcGroupRef,
n: &BigNumRef,
q: &EcPointRef,
m: &BigNumRef,
ctx: &mut BigNumContextRef
) -> Result<(), ErrorStack>

Computes generator * n + q * m, storing the result in self.

pub fn invert(
&mut self,
group: &EcGroupRef,
ctx: &BigNumContextRef
) -> Result<(), ErrorStack>

Inverts self.

OpenSSL documentation at EC_POINT_invert

pub fn to_bytes(
&self,
group: &EcGroupRef,
form: PointConversionForm,
ctx: &mut BigNumContextRef
) -> Result<Vec<u8>, ErrorStack>

Serializes the point to a binary representation.

OpenSSL documentation at EC_POINT_point2oct

pub fn eq(
&self,
group: &EcGroupRef,
other: &EcPointRef,
ctx: &mut BigNumContextRef
) -> Result<bool, ErrorStack>

Determines if this point is equal to another.

OpenSSL doucmentation at EC_POINT_cmp

pub fn affine_coordinates_gfp(
&self,
group: &EcGroupRef,
x: &mut BigNumRef,
y: &mut BigNumRef,
ctx: &mut BigNumContextRef
) -> Result<(), ErrorStack>

Place affine coordinates of a curve over a prime field in the provided x and y BigNums

OpenSSL documentation at EC_POINT_get_affine_coordinates_GFp

pub fn affine_coordinates_gf2m(
&self,
group: &EcGroupRef,
x: &mut BigNumRef,
y: &mut BigNumRef,
ctx: &mut BigNumContextRef
) -> Result<(), ErrorStack>

Place affine coordinates of a curve over a binary field in the provided x and y BigNums

OpenSSL documentation at EC_POINT_get_affine_coordinates_GF2m