Struct openssl::nid::Nid

pub struct Nid(_);

A numerical identifier for an OpenSSL object.

Objects in OpenSSL can have a short name, a long name, and a numerical identifier (NID). For convenience, objects are usually represented in source code using these numeric identifiers.

Users should generally not need to create new Nids.

Examples

To view the integer representation of a Nid:

use openssl::nid::Nid;

assert!(Nid::AES_256_GCM.as_raw() == 901);

External Documentation

The following documentation provides context about Nids and their usage in OpenSSL.

Obj_nid2obj

Implementations

impl Nid

pub fn from_raw(raw: c_int) -> Nid

Create a Nid from an integer representation.

pub fn as_raw(&self) -> c_int

Return the integer representation of a Nid.

pub fn signature_algorithms(&self) -> Option<SignatureAlgorithms>

Returns the Nids of the digest and public key algorithms associated with a signature ID.

This corresponds to OBJ_find_sigid_algs.

pub fn long_name(&self) -> Result<&'static str, ErrorStack>

Return the string representation of a Nid (long) This corresponds to OBJ_nid2ln

pub fn short_name(&self) -> Result<&'static str, ErrorStack>

Return the string representation of a Nid (short) This corresponds to OBJ_nid2sn

pub const UNDEF: Nid

pub const ITU_T: Nid

pub const CCITT: Nid

pub const ISO: Nid

pub const JOINT_ISO_ITU_T: Nid

pub const JOINT_ISO_CCITT: Nid

pub const MEMBER_BODY: Nid

pub const IDENTIFIED_ORGANIZATION: Nid

pub const HMAC_MD5: Nid

pub const HMAC_SHA1: Nid

pub const CERTICOM_ARC: Nid

pub const INTERNATIONAL_ORGANIZATIONS: Nid

pub const WAP: Nid

pub const WAP_WSG: Nid

pub const SELECTED_ATTRIBUTE_TYPES: Nid

pub const CLEARANCE: Nid

pub const ISO_US: Nid

pub const X9_57: Nid

pub const X9CM: Nid

pub const DSA: Nid

pub const DSAWITHSHA1: Nid

pub const ANSI_X9_62: Nid

pub const X9_62_PRIME_FIELD: Nid

pub const X9_62_CHARACTERISTIC_TWO_FIELD: Nid

pub const X9_62_ID_CHARACTERISTIC_TWO_BASIS: Nid

pub const X9_62_ONBASIS: Nid

pub const X9_62_TPBASIS: Nid

pub const X9_62_PPBASIS: Nid

pub const X9_62_ID_ECPUBLICKEY: Nid

pub const X9_62_C2PNB163V1: Nid

pub const X9_62_C2PNB163V2: Nid

pub const X9_62_C2PNB163V3: Nid

pub const X9_62_C2PNB176V1: Nid

pub const X9_62_C2TNB191V1: Nid

pub const X9_62_C2TNB191V2: Nid

pub const X9_62_C2TNB191V3: Nid

pub const X9_62_C2ONB191V4: Nid

pub const X9_62_C2ONB191V5: Nid

pub const X9_62_C2PNB208W1: Nid

pub const X9_62_C2TNB239V1: Nid

pub const X9_62_C2TNB239V2: Nid

pub const X9_62_C2TNB239V3: Nid

pub const X9_62_C2ONB239V4: Nid

pub const X9_62_C2ONB239V5: Nid

pub const X9_62_C2PNB272W1: Nid

pub const X9_62_C2PNB304W1: Nid

pub const X9_62_C2TNB359V1: Nid

pub const X9_62_C2PNB368W1: Nid

pub const X9_62_C2TNB431R1: Nid

pub const X9_62_PRIME192V1: Nid

pub const X9_62_PRIME192V2: Nid

pub const X9_62_PRIME192V3: Nid

pub const X9_62_PRIME239V1: Nid

pub const X9_62_PRIME239V2: Nid

pub const X9_62_PRIME239V3: Nid

pub const X9_62_PRIME256V1: Nid

pub const ECDSA_WITH_SHA1: Nid

pub const ECDSA_WITH_RECOMMENDED: Nid

pub const ECDSA_WITH_SPECIFIED: Nid

pub const ECDSA_WITH_SHA224: Nid

pub const ECDSA_WITH_SHA256: Nid

pub const ECDSA_WITH_SHA384: Nid

pub const ECDSA_WITH_SHA512: Nid

pub const SECP112R1: Nid

pub const SECP112R2: Nid

pub const SECP128R1: Nid

pub const SECP128R2: Nid

pub const SECP160K1: Nid

pub const SECP160R1: Nid

pub const SECP160R2: Nid

pub const SECP192K1: Nid

pub const SECP224K1: Nid

pub const SECP224R1: Nid

pub const SECP256K1: Nid

pub const SECP384R1: Nid

pub const SECP521R1: Nid

pub const SECT113R1: Nid

pub const SECT113R2: Nid

pub const SECT131R1: Nid

pub const SECT131R2: Nid

pub const SECT163K1: Nid

pub const SECT163R1: Nid

pub const SECT163R2: Nid

pub const SECT193R1: Nid

pub const SECT193R2: Nid

pub const SECT233K1: Nid

pub const SECT233R1: Nid

pub const SECT239K1: Nid

pub const SECT283K1: Nid

pub const SECT283R1: Nid

pub const SECT409K1: Nid

pub const SECT409R1: Nid

pub const SECT571K1: Nid

pub const SECT571R1: Nid

pub const WAP_WSG_IDM_ECID_WTLS1: Nid

pub const WAP_WSG_IDM_ECID_WTLS3: Nid

pub const WAP_WSG_IDM_ECID_WTLS4: Nid

pub const WAP_WSG_IDM_ECID_WTLS5: Nid

pub const WAP_WSG_IDM_ECID_WTLS6: Nid

pub const WAP_WSG_IDM_ECID_WTLS7: Nid

pub const WAP_WSG_IDM_ECID_WTLS8: Nid

pub const WAP_WSG_IDM_ECID_WTLS9: Nid

pub const WAP_WSG_IDM_ECID_WTLS10: Nid

pub const WAP_WSG_IDM_ECID_WTLS11: Nid

pub const WAP_WSG_IDM_ECID_WTLS12: Nid

pub const CAST5_CBC: Nid

pub const CAST5_ECB: Nid

pub const CAST5_CFB64: Nid

pub const CAST5_OFB64: Nid

pub const PBEWITHMD5ANDCAST5_CBC: Nid

pub const ID_PASSWORDBASEDMAC: Nid

pub const ID_DHBASEDMAC: Nid

pub const RSADSI: Nid

pub const PKCS: Nid

pub const PKCS1: Nid

pub const RSAENCRYPTION: Nid

pub const MD2WITHRSAENCRYPTION: Nid

pub const MD4WITHRSAENCRYPTION: Nid

pub const MD5WITHRSAENCRYPTION: Nid

pub const SHA1WITHRSAENCRYPTION: Nid

pub const RSAESOAEP: Nid

pub const MGF1: Nid

pub const RSASSAPSS: Nid

pub const SHA256WITHRSAENCRYPTION: Nid

pub const SHA384WITHRSAENCRYPTION: Nid

pub const SHA512WITHRSAENCRYPTION: Nid

pub const SHA224WITHRSAENCRYPTION: Nid

pub const PKCS3: Nid

pub const DHKEYAGREEMENT: Nid

pub const PKCS5: Nid

pub const PBEWITHMD2ANDDES_CBC: Nid

pub const PBEWITHMD5ANDDES_CBC: Nid

pub const PBEWITHMD2ANDRC2_CBC: Nid

pub const PBEWITHMD5ANDRC2_CBC: Nid

pub const PBEWITHSHA1ANDDES_CBC: Nid

pub const PBEWITHSHA1ANDRC2_CBC: Nid

pub const ID_PBKDF2: Nid

pub const PBES2: Nid

pub const PBMAC1: Nid

pub const PKCS7: Nid

pub const PKCS7_DATA: Nid

pub const PKCS7_SIGNED: Nid

pub const PKCS7_ENVELOPED: Nid

pub const PKCS7_SIGNEDANDENVELOPED: Nid

pub const PKCS7_DIGEST: Nid

pub const PKCS7_ENCRYPTED: Nid

pub const PKCS9: Nid

pub const PKCS9_EMAILADDRESS: Nid

pub const PKCS9_UNSTRUCTUREDNAME: Nid

pub const PKCS9_CONTENTTYPE: Nid

pub const PKCS9_MESSAGEDIGEST: Nid

pub const PKCS9_SIGNINGTIME: Nid

pub const PKCS9_COUNTERSIGNATURE: Nid

pub const PKCS9_CHALLENGEPASSWORD: Nid

pub const PKCS9_UNSTRUCTUREDADDRESS: Nid

pub const PKCS9_EXTCERTATTRIBUTES: Nid

pub const EXT_REQ: Nid

pub const SMIMECAPABILITIES: Nid

pub const SMIME: Nid

pub const ID_SMIME_MOD: Nid

pub const ID_SMIME_CT: Nid

pub const ID_SMIME_AA: Nid

pub const ID_SMIME_ALG: Nid

pub const ID_SMIME_CD: Nid

pub const ID_SMIME_SPQ: Nid

pub const ID_SMIME_CTI: Nid

pub const ID_SMIME_MOD_CMS: Nid

pub const ID_SMIME_MOD_ESS: Nid

pub const ID_SMIME_MOD_OID: Nid

pub const ID_SMIME_MOD_MSG_V3: Nid

pub const ID_SMIME_MOD_ETS_ESIGNATURE_88: Nid

pub const ID_SMIME_MOD_ETS_ESIGNATURE_97: Nid

pub const ID_SMIME_MOD_ETS_ESIGPOLICY_88: Nid

pub const ID_SMIME_MOD_ETS_ESIGPOLICY_97: Nid

pub const ID_SMIME_CT_RECEIPT: Nid

pub const ID_SMIME_CT_AUTHDATA: Nid

pub const ID_SMIME_CT_PUBLISHCERT: Nid

pub const ID_SMIME_CT_TSTINFO: Nid

pub const ID_SMIME_CT_TDTINFO: Nid

pub const ID_SMIME_CT_CONTENTINFO: Nid

pub const ID_SMIME_CT_DVCSREQUESTDATA: Nid

pub const ID_SMIME_CT_DVCSRESPONSEDATA: Nid

pub const ID_SMIME_CT_COMPRESSEDDATA: Nid

pub const ID_CT_ASCIITEXTWITHCRLF: Nid

pub const ID_SMIME_AA_RECEIPTREQUEST: Nid

pub const ID_SMIME_AA_SECURITYLABEL: Nid

pub const ID_SMIME_AA_MLEXPANDHISTORY: Nid

pub const ID_SMIME_AA_CONTENTHINT: Nid

pub const ID_SMIME_AA_MSGSIGDIGEST: Nid

pub const ID_SMIME_AA_ENCAPCONTENTTYPE: Nid

pub const ID_SMIME_AA_CONTENTIDENTIFIER: Nid

pub const ID_SMIME_AA_MACVALUE: Nid

pub const ID_SMIME_AA_EQUIVALENTLABELS: Nid

pub const ID_SMIME_AA_CONTENTREFERENCE: Nid

pub const ID_SMIME_AA_ENCRYPKEYPREF: Nid

pub const ID_SMIME_AA_SIGNINGCERTIFICATE: Nid

pub const ID_SMIME_AA_SMIMEENCRYPTCERTS: Nid

pub const ID_SMIME_AA_TIMESTAMPTOKEN: Nid

pub const ID_SMIME_AA_ETS_SIGPOLICYID: Nid

pub const ID_SMIME_AA_ETS_COMMITMENTTYPE: Nid

pub const ID_SMIME_AA_ETS_SIGNERLOCATION: Nid

pub const ID_SMIME_AA_ETS_SIGNERATTR: Nid

pub const ID_SMIME_AA_ETS_OTHERSIGCERT: Nid

pub const ID_SMIME_AA_ETS_CONTENTTIMESTAMP: Nid

pub const ID_SMIME_AA_ETS_CERTIFICATEREFS: Nid

pub const ID_SMIME_AA_ETS_REVOCATIONREFS: Nid

pub const ID_SMIME_AA_ETS_CERTVALUES: Nid

pub const ID_SMIME_AA_ETS_REVOCATIONVALUES: Nid

pub const ID_SMIME_AA_ETS_ESCTIMESTAMP: Nid

pub const ID_SMIME_AA_ETS_CERTCRLTIMESTAMP: Nid

pub const ID_SMIME_AA_ETS_ARCHIVETIMESTAMP: Nid

pub const ID_SMIME_AA_SIGNATURETYPE: Nid

pub const ID_SMIME_AA_DVCS_DVC: Nid

pub const ID_SMIME_ALG_ESDHWITH3DES: Nid

pub const ID_SMIME_ALG_ESDHWITHRC2: Nid

pub const ID_SMIME_ALG_3DESWRAP: Nid

pub const ID_SMIME_ALG_RC2WRAP: Nid

pub const ID_SMIME_ALG_ESDH: Nid

pub const ID_SMIME_ALG_CMS3DESWRAP: Nid

pub const ID_SMIME_ALG_CMSRC2WRAP: Nid

pub const ID_ALG_PWRI_KEK: Nid

pub const ID_SMIME_CD_LDAP: Nid

pub const ID_SMIME_SPQ_ETS_SQT_URI: Nid

pub const ID_SMIME_SPQ_ETS_SQT_UNOTICE: Nid

pub const ID_SMIME_CTI_ETS_PROOFOFORIGIN: Nid

pub const ID_SMIME_CTI_ETS_PROOFOFRECEIPT: Nid

pub const ID_SMIME_CTI_ETS_PROOFOFDELIVERY: Nid

pub const ID_SMIME_CTI_ETS_PROOFOFSENDER: Nid

pub const ID_SMIME_CTI_ETS_PROOFOFAPPROVAL: Nid

pub const ID_SMIME_CTI_ETS_PROOFOFCREATION: Nid

pub const FRIENDLYNAME: Nid

pub const LOCALKEYID: Nid

pub const MS_CSP_NAME: Nid

pub const LOCALKEYSET: Nid

pub const X509CERTIFICATE: Nid

pub const SDSICERTIFICATE: Nid

pub const X509CRL: Nid

pub const PBE_WITHSHA1AND128BITRC4: Nid

pub const PBE_WITHSHA1AND40BITRC4: Nid

pub const PBE_WITHSHA1AND3_KEY_TRIPLEDES_CBC: Nid

pub const PBE_WITHSHA1AND2_KEY_TRIPLEDES_CBC: Nid

pub const PBE_WITHSHA1AND128BITRC2_CBC: Nid

pub const PBE_WITHSHA1AND40BITRC2_CBC: Nid

pub const KEYBAG: Nid

pub const PKCS8SHROUDEDKEYBAG: Nid

pub const CERTBAG: Nid

pub const CRLBAG: Nid

pub const SECRETBAG: Nid

pub const SAFECONTENTSBAG: Nid

pub const MD2: Nid

pub const MD4: Nid

pub const MD5: Nid

pub const MD5_SHA1: Nid

pub const HMACWITHMD5: Nid

pub const HMACWITHSHA1: Nid

pub const HMACWITHSHA224: Nid

pub const HMACWITHSHA256: Nid

pub const HMACWITHSHA384: Nid

pub const HMACWITHSHA512: Nid

pub const RC2_CBC: Nid

pub const RC2_ECB: Nid

pub const RC2_CFB64: Nid

pub const RC2_OFB64: Nid

pub const RC2_40_CBC: Nid

pub const RC2_64_CBC: Nid

pub const RC4: Nid

pub const RC4_40: Nid

pub const DES_EDE3_CBC: Nid

pub const RC5_CBC: Nid

pub const RC5_ECB: Nid

pub const RC5_CFB64: Nid

pub const RC5_OFB64: Nid

pub const MS_EXT_REQ: Nid

pub const MS_CODE_IND: Nid

pub const MS_CODE_COM: Nid

pub const MS_CTL_SIGN: Nid

pub const MS_SGC: Nid

pub const MS_EFS: Nid

pub const MS_SMARTCARD_LOGIN: Nid

pub const MS_UPN: Nid

pub const IDEA_CBC: Nid

pub const IDEA_ECB: Nid

pub const IDEA_CFB64: Nid

pub const IDEA_OFB64: Nid

pub const BF_CBC: Nid

pub const BF_ECB: Nid

pub const BF_CFB64: Nid

pub const BF_OFB64: Nid

pub const ID_PKIX: Nid

pub const ID_PKIX_MOD: Nid

pub const ID_PE: Nid

pub const ID_QT: Nid

pub const ID_KP: Nid

pub const ID_IT: Nid

pub const ID_PKIP: Nid

pub const ID_ALG: Nid

pub const ID_CMC: Nid

pub const ID_ON: Nid

pub const ID_PDA: Nid

pub const ID_ACA: Nid

pub const ID_QCS: Nid

pub const ID_CCT: Nid

pub const ID_PPL: Nid

pub const ID_AD: Nid

pub const ID_PKIX1_EXPLICIT_88: Nid

pub const ID_PKIX1_IMPLICIT_88: Nid

pub const ID_PKIX1_EXPLICIT_93: Nid

pub const ID_PKIX1_IMPLICIT_93: Nid

pub const ID_MOD_CRMF: Nid

pub const ID_MOD_CMC: Nid

pub const ID_MOD_KEA_PROFILE_88: Nid

pub const ID_MOD_KEA_PROFILE_93: Nid

pub const ID_MOD_CMP: Nid

pub const ID_MOD_QUALIFIED_CERT_88: Nid

pub const ID_MOD_QUALIFIED_CERT_93: Nid

pub const ID_MOD_ATTRIBUTE_CERT: Nid

pub const ID_MOD_TIMESTAMP_PROTOCOL: Nid

pub const ID_MOD_OCSP: Nid

pub const ID_MOD_DVCS: Nid

pub const ID_MOD_CMP2000: Nid

pub const INFO_ACCESS: Nid

pub const BIOMETRICINFO: Nid

pub const QCSTATEMENTS: Nid

pub const AC_AUDITENTITY: Nid

pub const AC_TARGETING: Nid

pub const AACONTROLS: Nid

pub const SBGP_IPADDRBLOCK: Nid

pub const SBGP_AUTONOMOUSSYSNUM: Nid

pub const SBGP_ROUTERIDENTIFIER: Nid

pub const AC_PROXYING: Nid

pub const SINFO_ACCESS: Nid

pub const PROXYCERTINFO: Nid

pub const ID_QT_CPS: Nid

pub const ID_QT_UNOTICE: Nid

pub const TEXTNOTICE: Nid

pub const SERVER_AUTH: Nid

pub const CLIENT_AUTH: Nid

pub const CODE_SIGN: Nid

pub const EMAIL_PROTECT: Nid

pub const IPSECENDSYSTEM: Nid

pub const IPSECTUNNEL: Nid

pub const IPSECUSER: Nid

pub const TIME_STAMP: Nid

pub const OCSP_SIGN: Nid

pub const DVCS: Nid

pub const ID_IT_CAPROTENCCERT: Nid

pub const ID_IT_SIGNKEYPAIRTYPES: Nid

pub const ID_IT_ENCKEYPAIRTYPES: Nid

pub const ID_IT_PREFERREDSYMMALG: Nid

pub const ID_IT_CAKEYUPDATEINFO: Nid

pub const ID_IT_CURRENTCRL: Nid

pub const ID_IT_UNSUPPORTEDOIDS: Nid

pub const ID_IT_SUBSCRIPTIONREQUEST: Nid

pub const ID_IT_SUBSCRIPTIONRESPONSE: Nid

pub const ID_IT_KEYPAIRPARAMREQ: Nid

pub const ID_IT_KEYPAIRPARAMREP: Nid

pub const ID_IT_REVPASSPHRASE: Nid

pub const ID_IT_IMPLICITCONFIRM: Nid

pub const ID_IT_CONFIRMWAITTIME: Nid

pub const ID_IT_ORIGPKIMESSAGE: Nid

pub const ID_IT_SUPPLANGTAGS: Nid

pub const ID_REGCTRL: Nid

pub const ID_REGINFO: Nid

pub const ID_REGCTRL_REGTOKEN: Nid

pub const ID_REGCTRL_AUTHENTICATOR: Nid

pub const ID_REGCTRL_PKIPUBLICATIONINFO: Nid

pub const ID_REGCTRL_PKIARCHIVEOPTIONS: Nid

pub const ID_REGCTRL_OLDCERTID: Nid

pub const ID_REGCTRL_PROTOCOLENCRKEY: Nid

pub const ID_REGINFO_UTF8PAIRS: Nid

pub const ID_REGINFO_CERTREQ: Nid

pub const ID_ALG_DES40: Nid

pub const ID_ALG_NOSIGNATURE: Nid

pub const ID_ALG_DH_SIG_HMAC_SHA1: Nid

pub const ID_ALG_DH_POP: Nid

pub const ID_CMC_STATUSINFO: Nid

pub const ID_CMC_IDENTIFICATION: Nid

pub const ID_CMC_IDENTITYPROOF: Nid

pub const ID_CMC_DATARETURN: Nid

pub const ID_CMC_TRANSACTIONID: Nid

pub const ID_CMC_SENDERNONCE: Nid

pub const ID_CMC_RECIPIENTNONCE: Nid

pub const ID_CMC_ADDEXTENSIONS: Nid

pub const ID_CMC_ENCRYPTEDPOP: Nid

pub const ID_CMC_DECRYPTEDPOP: Nid

pub const ID_CMC_LRAPOPWITNESS: Nid

pub const ID_CMC_GETCERT: Nid

pub const ID_CMC_GETCRL: Nid

pub const ID_CMC_REVOKEREQUEST: Nid

pub const ID_CMC_REGINFO: Nid

pub const ID_CMC_RESPONSEINFO: Nid

pub const ID_CMC_QUERYPENDING: Nid

pub const ID_CMC_POPLINKRANDOM: Nid

pub const ID_CMC_POPLINKWITNESS: Nid

pub const ID_CMC_CONFIRMCERTACCEPTANCE: Nid

pub const ID_ON_PERSONALDATA: Nid

pub const ID_ON_PERMANENTIDENTIFIER: Nid

pub const ID_PDA_DATEOFBIRTH: Nid

pub const ID_PDA_PLACEOFBIRTH: Nid

pub const ID_PDA_GENDER: Nid

pub const ID_PDA_COUNTRYOFCITIZENSHIP: Nid

pub const ID_PDA_COUNTRYOFRESIDENCE: Nid

pub const ID_ACA_AUTHENTICATIONINFO: Nid

pub const ID_ACA_ACCESSIDENTITY: Nid

pub const ID_ACA_CHARGINGIDENTITY: Nid

pub const ID_ACA_GROUP: Nid

pub const ID_ACA_ROLE: Nid

pub const ID_ACA_ENCATTRS: Nid

pub const ID_QCS_PKIXQCSYNTAX_V1: Nid

pub const ID_CCT_CRS: Nid

pub const ID_CCT_PKIDATA: Nid

pub const ID_CCT_PKIRESPONSE: Nid

pub const ID_PPL_ANYLANGUAGE: Nid

pub const ID_PPL_INHERITALL: Nid

pub const INDEPENDENT: Nid

pub const AD_OCSP: Nid

pub const AD_CA_ISSUERS: Nid

pub const AD_TIMESTAMPING: Nid

pub const AD_DVCS: Nid

pub const CAREPOSITORY: Nid

pub const ID_PKIX_OCSP_BASIC: Nid

pub const ID_PKIX_OCSP_NONCE: Nid

pub const ID_PKIX_OCSP_CRLID: Nid

pub const ID_PKIX_OCSP_ACCEPTABLERESPONSES: Nid

pub const ID_PKIX_OCSP_NOCHECK: Nid

pub const ID_PKIX_OCSP_ARCHIVECUTOFF: Nid

pub const ID_PKIX_OCSP_SERVICELOCATOR: Nid

pub const ID_PKIX_OCSP_EXTENDEDSTATUS: Nid

pub const ID_PKIX_OCSP_VALID: Nid

pub const ID_PKIX_OCSP_PATH: Nid

pub const ID_PKIX_OCSP_TRUSTROOT: Nid

pub const ALGORITHM: Nid

pub const MD5WITHRSA: Nid

pub const DES_ECB: Nid

pub const DES_CBC: Nid

pub const DES_OFB64: Nid

pub const DES_CFB64: Nid

pub const RSASIGNATURE: Nid

pub const DSA_2: Nid

pub const DSAWITHSHA: Nid

pub const SHAWITHRSAENCRYPTION: Nid

pub const DES_EDE_ECB: Nid

pub const DES_EDE3_ECB: Nid

pub const DES_EDE_CBC: Nid

pub const DES_EDE_CFB64: Nid

pub const DES_EDE3_CFB64: Nid

pub const DES_EDE_OFB64: Nid

pub const DES_EDE3_OFB64: Nid

pub const DESX_CBC: Nid

pub const SHA: Nid

pub const SHA1: Nid

pub const DSAWITHSHA1_2: Nid

pub const SHA1WITHRSA: Nid

pub const RIPEMD160: Nid

pub const RIPEMD160WITHRSA: Nid

pub const SXNET: Nid

pub const X500: Nid

pub const X509: Nid

pub const COMMONNAME: Nid

pub const SURNAME: Nid

pub const SERIALNUMBER: Nid

pub const COUNTRYNAME: Nid

pub const LOCALITYNAME: Nid

pub const STATEORPROVINCENAME: Nid

pub const STREETADDRESS: Nid

pub const ORGANIZATIONNAME: Nid

pub const ORGANIZATIONALUNITNAME: Nid

pub const TITLE: Nid

pub const DESCRIPTION: Nid

pub const SEARCHGUIDE: Nid

pub const BUSINESSCATEGORY: Nid

pub const POSTALADDRESS: Nid

pub const POSTALCODE: Nid

pub const POSTOFFICEBOX: Nid

pub const PHYSICALDELIVERYOFFICENAME: Nid

pub const TELEPHONENUMBER: Nid

pub const TELEXNUMBER: Nid

pub const TELETEXTERMINALIDENTIFIER: Nid

pub const FACSIMILETELEPHONENUMBER: Nid

pub const X121ADDRESS: Nid

pub const INTERNATIONALISDNNUMBER: Nid

pub const REGISTEREDADDRESS: Nid

pub const DESTINATIONINDICATOR: Nid

pub const PREFERREDDELIVERYMETHOD: Nid

pub const PRESENTATIONADDRESS: Nid

pub const SUPPORTEDAPPLICATIONCONTEXT: Nid

pub const MEMBER: Nid

pub const OWNER: Nid

pub const ROLEOCCUPANT: Nid

pub const SEEALSO: Nid

pub const USERPASSWORD: Nid

pub const USERCERTIFICATE: Nid

pub const CACERTIFICATE: Nid

pub const AUTHORITYREVOCATIONLIST: Nid

pub const CERTIFICATEREVOCATIONLIST: Nid

pub const CROSSCERTIFICATEPAIR: Nid

pub const NAME: Nid

pub const GIVENNAME: Nid

pub const INITIALS: Nid

pub const GENERATIONQUALIFIER: Nid

pub const X500UNIQUEIDENTIFIER: Nid

pub const DNQUALIFIER: Nid

pub const ENHANCEDSEARCHGUIDE: Nid

pub const PROTOCOLINFORMATION: Nid

pub const DISTINGUISHEDNAME: Nid

pub const UNIQUEMEMBER: Nid

pub const HOUSEIDENTIFIER: Nid

pub const SUPPORTEDALGORITHMS: Nid

pub const DELTAREVOCATIONLIST: Nid

pub const DMDNAME: Nid

pub const PSEUDONYM: Nid

pub const ROLE: Nid

pub const X500ALGORITHMS: Nid

pub const RSA: Nid

pub const MDC2WITHRSA: Nid

pub const MDC2: Nid

pub const ID_CE: Nid

pub const SUBJECT_DIRECTORY_ATTRIBUTES: Nid

pub const SUBJECT_KEY_IDENTIFIER: Nid

pub const KEY_USAGE: Nid

pub const PRIVATE_KEY_USAGE_PERIOD: Nid

pub const SUBJECT_ALT_NAME: Nid

pub const ISSUER_ALT_NAME: Nid

pub const BASIC_CONSTRAINTS: Nid

pub const CRL_NUMBER: Nid

pub const CRL_REASON: Nid

pub const INVALIDITY_DATE: Nid

pub const DELTA_CRL: Nid

pub const ISSUING_DISTRIBUTION_POINT: Nid

pub const CERTIFICATE_ISSUER: Nid

pub const NAME_CONSTRAINTS: Nid

pub const CRL_DISTRIBUTION_POINTS: Nid

pub const CERTIFICATE_POLICIES: Nid

pub const ANY_POLICY: Nid

pub const POLICY_MAPPINGS: Nid

pub const AUTHORITY_KEY_IDENTIFIER: Nid

pub const POLICY_CONSTRAINTS: Nid

pub const EXT_KEY_USAGE: Nid

pub const FRESHEST_CRL: Nid

pub const INHIBIT_ANY_POLICY: Nid

pub const TARGET_INFORMATION: Nid

pub const NO_REV_AVAIL: Nid

pub const ANYEXTENDEDKEYUSAGE: Nid

pub const NETSCAPE: Nid

pub const NETSCAPE_CERT_EXTENSION: Nid

pub const NETSCAPE_DATA_TYPE: Nid

pub const NETSCAPE_CERT_TYPE: Nid

pub const NETSCAPE_BASE_URL: Nid

pub const NETSCAPE_REVOCATION_URL: Nid

pub const NETSCAPE_CA_REVOCATION_URL: Nid

pub const NETSCAPE_RENEWAL_URL: Nid

pub const NETSCAPE_CA_POLICY_URL: Nid

pub const NETSCAPE_SSL_SERVER_NAME: Nid

pub const NETSCAPE_COMMENT: Nid

pub const NETSCAPE_CERT_SEQUENCE: Nid

pub const NS_SGC: Nid

pub const ORG: Nid

pub const DOD: Nid

pub const IANA: Nid

pub const DIRECTORY: Nid

pub const MANAGEMENT: Nid

pub const EXPERIMENTAL: Nid

pub const PRIVATE: Nid

pub const SECURITY: Nid

pub const SNMPV2: Nid

pub const MAIL: Nid

pub const ENTERPRISES: Nid

pub const DCOBJECT: Nid

pub const MIME_MHS: Nid

pub const MIME_MHS_HEADINGS: Nid

pub const MIME_MHS_BODIES: Nid

pub const ID_HEX_PARTIAL_MESSAGE: Nid

pub const ID_HEX_MULTIPART_MESSAGE: Nid

pub const ZLIB_COMPRESSION: Nid

pub const AES_128_ECB: Nid

pub const AES_128_CBC: Nid

pub const AES_128_OFB128: Nid

pub const AES_128_CFB128: Nid

pub const ID_AES128_WRAP: Nid

pub const AES_128_GCM: Nid

pub const AES_128_CCM: Nid

pub const ID_AES128_WRAP_PAD: Nid

pub const AES_192_ECB: Nid

pub const AES_192_CBC: Nid

pub const AES_192_OFB128: Nid

pub const AES_192_CFB128: Nid

pub const ID_AES192_WRAP: Nid

pub const AES_192_GCM: Nid

pub const AES_192_CCM: Nid

pub const ID_AES192_WRAP_PAD: Nid

pub const AES_256_ECB: Nid

pub const AES_256_CBC: Nid

pub const AES_256_OFB128: Nid

pub const AES_256_CFB128: Nid

pub const ID_AES256_WRAP: Nid

pub const AES_256_GCM: Nid

pub const AES_256_CCM: Nid

pub const ID_AES256_WRAP_PAD: Nid

pub const AES_128_CFB1: Nid

pub const AES_192_CFB1: Nid

pub const AES_256_CFB1: Nid

pub const AES_128_CFB8: Nid

pub const AES_192_CFB8: Nid

pub const AES_256_CFB8: Nid

pub const AES_128_CTR: Nid

pub const AES_192_CTR: Nid

pub const AES_256_CTR: Nid

pub const AES_128_XTS: Nid

pub const AES_256_XTS: Nid

pub const DES_CFB1: Nid

pub const DES_CFB8: Nid

pub const DES_EDE3_CFB1: Nid

pub const DES_EDE3_CFB8: Nid

pub const SHA256: Nid

pub const SHA384: Nid

pub const SHA512: Nid

pub const SHA224: Nid

pub const DSA_WITH_SHA224: Nid

pub const DSA_WITH_SHA256: Nid

pub const HOLD_INSTRUCTION_CODE: Nid

pub const HOLD_INSTRUCTION_NONE: Nid

pub const HOLD_INSTRUCTION_CALL_ISSUER: Nid

pub const HOLD_INSTRUCTION_REJECT: Nid

pub const DATA: Nid

pub const PSS: Nid

pub const UCL: Nid

pub const PILOT: Nid

pub const PILOTATTRIBUTETYPE: Nid

pub const PILOTATTRIBUTESYNTAX: Nid

pub const PILOTOBJECTCLASS: Nid

pub const PILOTGROUPS: Nid

pub const IA5STRINGSYNTAX: Nid

pub const CASEIGNOREIA5STRINGSYNTAX: Nid

pub const PILOTOBJECT: Nid

pub const PILOTPERSON: Nid

pub const ACCOUNT: Nid

pub const DOCUMENT: Nid

pub const ROOM: Nid

pub const DOCUMENTSERIES: Nid

pub const DOMAIN: Nid

pub const RFC822LOCALPART: Nid

pub const DNSDOMAIN: Nid

pub const DOMAINRELATEDOBJECT: Nid

pub const FRIENDLYCOUNTRY: Nid

pub const SIMPLESECURITYOBJECT: Nid

pub const PILOTORGANIZATION: Nid

pub const PILOTDSA: Nid

pub const QUALITYLABELLEDDATA: Nid

pub const USERID: Nid

pub const TEXTENCODEDORADDRESS: Nid

pub const RFC822MAILBOX: Nid

pub const INFO: Nid

pub const FAVOURITEDRINK: Nid

pub const ROOMNUMBER: Nid

pub const PHOTO: Nid

pub const USERCLASS: Nid

pub const HOST: Nid

pub const MANAGER: Nid

pub const DOCUMENTIDENTIFIER: Nid

pub const DOCUMENTTITLE: Nid

pub const DOCUMENTVERSION: Nid

pub const DOCUMENTAUTHOR: Nid

pub const DOCUMENTLOCATION: Nid

pub const HOMETELEPHONENUMBER: Nid

pub const SECRETARY: Nid

pub const OTHERMAILBOX: Nid

pub const LASTMODIFIEDTIME: Nid

pub const LASTMODIFIEDBY: Nid

pub const DOMAINCOMPONENT: Nid

pub const ARECORD: Nid

pub const PILOTATTRIBUTETYPE27: Nid

pub const MXRECORD: Nid

pub const NSRECORD: Nid

pub const SOARECORD: Nid

pub const CNAMERECORD: Nid

pub const ASSOCIATEDDOMAIN: Nid

pub const ASSOCIATEDNAME: Nid

pub const HOMEPOSTALADDRESS: Nid

pub const PERSONALTITLE: Nid

pub const MOBILETELEPHONENUMBER: Nid

pub const PAGERTELEPHONENUMBER: Nid

pub const FRIENDLYCOUNTRYNAME: Nid

pub const ORGANIZATIONALSTATUS: Nid

pub const JANETMAILBOX: Nid

pub const MAILPREFERENCEOPTION: Nid

pub const BUILDINGNAME: Nid

pub const DSAQUALITY: Nid

pub const SINGLELEVELQUALITY: Nid

pub const SUBTREEMINIMUMQUALITY: Nid

pub const SUBTREEMAXIMUMQUALITY: Nid

pub const PERSONALSIGNATURE: Nid

pub const DITREDIRECT: Nid

pub const AUDIO: Nid

pub const DOCUMENTPUBLISHER: Nid

pub const ID_SET: Nid

pub const SET_CTYPE: Nid

pub const SET_MSGEXT: Nid

pub const SET_ATTR: Nid

pub const SET_POLICY: Nid

pub const SET_CERTEXT: Nid

pub const SET_BRAND: Nid

pub const SETCT_PANDATA: Nid

pub const SETCT_PANTOKEN: Nid

pub const SETCT_PANONLY: Nid

pub const SETCT_OIDATA: Nid

pub const SETCT_PI: Nid

pub const SETCT_PIDATA: Nid

pub const SETCT_PIDATAUNSIGNED: Nid

pub const SETCT_HODINPUT: Nid

pub const SETCT_AUTHRESBAGGAGE: Nid

pub const SETCT_AUTHREVREQBAGGAGE: Nid

pub const SETCT_AUTHREVRESBAGGAGE: Nid

pub const SETCT_CAPTOKENSEQ: Nid

pub const SETCT_PINITRESDATA: Nid

pub const SETCT_PI_TBS: Nid

pub const SETCT_PRESDATA: Nid

pub const SETCT_AUTHREQTBS: Nid

pub const SETCT_AUTHRESTBS: Nid

pub const SETCT_AUTHRESTBSX: Nid

pub const SETCT_AUTHTOKENTBS: Nid

pub const SETCT_CAPTOKENDATA: Nid

pub const SETCT_CAPTOKENTBS: Nid

pub const SETCT_ACQCARDCODEMSG: Nid

pub const SETCT_AUTHREVREQTBS: Nid

pub const SETCT_AUTHREVRESDATA: Nid

pub const SETCT_AUTHREVRESTBS: Nid

pub const SETCT_CAPREQTBS: Nid

pub const SETCT_CAPREQTBSX: Nid

pub const SETCT_CAPRESDATA: Nid

pub const SETCT_CAPREVREQTBS: Nid

pub const SETCT_CAPREVREQTBSX: Nid

pub const SETCT_CAPREVRESDATA: Nid

pub const SETCT_CREDREQTBS: Nid

pub const SETCT_CREDREQTBSX: Nid

pub const SETCT_CREDRESDATA: Nid

pub const SETCT_CREDREVREQTBS: Nid

pub const SETCT_CREDREVREQTBSX: Nid

pub const SETCT_CREDREVRESDATA: Nid

pub const SETCT_PCERTREQDATA: Nid

pub const SETCT_PCERTRESTBS: Nid

pub const SETCT_BATCHADMINREQDATA: Nid

pub const SETCT_BATCHADMINRESDATA: Nid

pub const SETCT_CARDCINITRESTBS: Nid

pub const SETCT_MEAQCINITRESTBS: Nid

pub const SETCT_REGFORMRESTBS: Nid

pub const SETCT_CERTREQDATA: Nid

pub const SETCT_CERTREQTBS: Nid

pub const SETCT_CERTRESDATA: Nid

pub const SETCT_CERTINQREQTBS: Nid

pub const SETCT_ERRORTBS: Nid

pub const SETCT_PIDUALSIGNEDTBE: Nid

pub const SETCT_PIUNSIGNEDTBE: Nid

pub const SETCT_AUTHREQTBE: Nid

pub const SETCT_AUTHRESTBE: Nid

pub const SETCT_AUTHRESTBEX: Nid

pub const SETCT_AUTHTOKENTBE: Nid

pub const SETCT_CAPTOKENTBE: Nid

pub const SETCT_CAPTOKENTBEX: Nid

pub const SETCT_ACQCARDCODEMSGTBE: Nid

pub const SETCT_AUTHREVREQTBE: Nid

pub const SETCT_AUTHREVRESTBE: Nid

pub const SETCT_AUTHREVRESTBEB: Nid

pub const SETCT_CAPREQTBE: Nid

pub const SETCT_CAPREQTBEX: Nid

pub const SETCT_CAPRESTBE: Nid

pub const SETCT_CAPREVREQTBE: Nid

pub const SETCT_CAPREVREQTBEX: Nid

pub const SETCT_CAPREVRESTBE: Nid

pub const SETCT_CREDREQTBE: Nid

pub const SETCT_CREDREQTBEX: Nid

pub const SETCT_CREDRESTBE: Nid

pub const SETCT_CREDREVREQTBE: Nid

pub const SETCT_CREDREVREQTBEX: Nid

pub const SETCT_CREDREVRESTBE: Nid

pub const SETCT_BATCHADMINREQTBE: Nid

pub const SETCT_BATCHADMINRESTBE: Nid

pub const SETCT_REGFORMREQTBE: Nid

pub const SETCT_CERTREQTBE: Nid

pub const SETCT_CERTREQTBEX: Nid

pub const SETCT_CERTRESTBE: Nid

pub const SETCT_CRLNOTIFICATIONTBS: Nid

pub const SETCT_CRLNOTIFICATIONRESTBS: Nid

pub const SETCT_BCIDISTRIBUTIONTBS: Nid

pub const SETEXT_GENCRYPT: Nid

pub const SETEXT_MIAUTH: Nid

pub const SETEXT_PINSECURE: Nid

pub const SETEXT_PINANY: Nid

pub const SETEXT_TRACK2: Nid

pub const SETEXT_CV: Nid

pub const SET_POLICY_ROOT: Nid

pub const SETCEXT_HASHEDROOT: Nid

pub const SETCEXT_CERTTYPE: Nid

pub const SETCEXT_MERCHDATA: Nid

pub const SETCEXT_CCERTREQUIRED: Nid

pub const SETCEXT_TUNNELING: Nid

pub const SETCEXT_SETEXT: Nid

pub const SETCEXT_SETQUALF: Nid

pub const SETCEXT_PGWYCAPABILITIES: Nid

pub const SETCEXT_TOKENIDENTIFIER: Nid

pub const SETCEXT_TRACK2DATA: Nid

pub const SETCEXT_TOKENTYPE: Nid

pub const SETCEXT_ISSUERCAPABILITIES: Nid

pub const SETATTR_CERT: Nid

pub const SETATTR_PGWYCAP: Nid

pub const SETATTR_TOKENTYPE: Nid

pub const SETATTR_ISSCAP: Nid

pub const SET_ROOTKEYTHUMB: Nid

pub const SET_ADDPOLICY: Nid

pub const SETATTR_TOKEN_EMV: Nid

pub const SETATTR_TOKEN_B0PRIME: Nid

pub const SETATTR_ISSCAP_CVM: Nid

pub const SETATTR_ISSCAP_T2: Nid

pub const SETATTR_ISSCAP_SIG: Nid

pub const SETATTR_GENCRYPTGRM: Nid

pub const SETATTR_T2ENC: Nid

pub const SETATTR_T2CLEARTXT: Nid

pub const SETATTR_TOKICCSIG: Nid

pub const SETATTR_SECDEVSIG: Nid

pub const SET_BRAND_IATA_ATA: Nid

pub const SET_BRAND_DINERS: Nid

pub const SET_BRAND_AMERICANEXPRESS: Nid

pub const SET_BRAND_JCB: Nid

pub const SET_BRAND_VISA: Nid

pub const SET_BRAND_MASTERCARD: Nid

pub const SET_BRAND_NOVUS: Nid

pub const DES_CDMF: Nid

pub const RSAOAEPENCRYPTIONSET: Nid

pub const IPSEC3: Nid

pub const IPSEC4: Nid

pub const WHIRLPOOL: Nid

pub const CRYPTOPRO: Nid

pub const CRYPTOCOM: Nid

pub const ID_GOSTR3411_94_WITH_GOSTR3410_2001: Nid

pub const ID_GOSTR3411_94_WITH_GOSTR3410_94: Nid

pub const ID_GOSTR3411_94: Nid

pub const ID_HMACGOSTR3411_94: Nid

pub const ID_GOSTR3410_2001: Nid

pub const ID_GOSTR3410_94: Nid

pub const ID_GOST28147_89: Nid

pub const GOST89_CNT: Nid

pub const ID_GOST28147_89_MAC: Nid

pub const ID_GOSTR3411_94_PRF: Nid

pub const ID_GOSTR3410_2001DH: Nid

pub const ID_GOSTR3410_94DH: Nid

pub const ID_GOST28147_89_CRYPTOPRO_KEYMESHING: Nid

pub const ID_GOST28147_89_NONE_KEYMESHING: Nid

pub const ID_GOSTR3411_94_TESTPARAMSET: Nid

pub const ID_GOSTR3411_94_CRYPTOPROPARAMSET: Nid

pub const ID_GOST28147_89_TESTPARAMSET: Nid

pub const ID_GOST28147_89_CRYPTOPRO_A_PARAMSET: Nid

pub const ID_GOST28147_89_CRYPTOPRO_B_PARAMSET: Nid

pub const ID_GOST28147_89_CRYPTOPRO_C_PARAMSET: Nid

pub const ID_GOST28147_89_CRYPTOPRO_D_PARAMSET: Nid

pub const ID_GOST28147_89_CRYPTOPRO_OSCAR_1_1_PARAMSET: Nid

pub const ID_GOST28147_89_CRYPTOPRO_OSCAR_1_0_PARAMSET: Nid

pub const ID_GOST28147_89_CRYPTOPRO_RIC_1_PARAMSET: Nid

pub const ID_GOSTR3410_94_TESTPARAMSET: Nid

pub const ID_GOSTR3410_94_CRYPTOPRO_A_PARAMSET: Nid

pub const ID_GOSTR3410_94_CRYPTOPRO_B_PARAMSET: Nid

pub const ID_GOSTR3410_94_CRYPTOPRO_C_PARAMSET: Nid

pub const ID_GOSTR3410_94_CRYPTOPRO_D_PARAMSET: Nid

pub const ID_GOSTR3410_94_CRYPTOPRO_XCHA_PARAMSET: Nid

pub const ID_GOSTR3410_94_CRYPTOPRO_XCHB_PARAMSET: Nid

pub const ID_GOSTR3410_94_CRYPTOPRO_XCHC_PARAMSET: Nid

pub const ID_GOSTR3410_2001_TESTPARAMSET: Nid

pub const ID_GOSTR3410_2001_CRYPTOPRO_A_PARAMSET: Nid

pub const ID_GOSTR3410_2001_CRYPTOPRO_B_PARAMSET: Nid

pub const ID_GOSTR3410_2001_CRYPTOPRO_C_PARAMSET: Nid

pub const ID_GOSTR3410_2001_CRYPTOPRO_XCHA_PARAMSET: Nid

pub const ID_GOSTR3410_2001_CRYPTOPRO_XCHB_PARAMSET: Nid

pub const ID_GOSTR3410_94_A: Nid

pub const ID_GOSTR3410_94_ABIS: Nid

pub const ID_GOSTR3410_94_B: Nid

pub const ID_GOSTR3410_94_BBIS: Nid

pub const ID_GOST28147_89_CC: Nid

pub const ID_GOSTR3410_94_CC: Nid

pub const ID_GOSTR3410_2001_CC: Nid

pub const ID_GOSTR3411_94_WITH_GOSTR3410_94_CC: Nid

pub const ID_GOSTR3411_94_WITH_GOSTR3410_2001_CC: Nid

pub const ID_GOSTR3410_2001_PARAMSET_CC: Nid

pub const CAMELLIA_128_CBC: Nid

pub const CAMELLIA_192_CBC: Nid

pub const CAMELLIA_256_CBC: Nid

pub const ID_CAMELLIA128_WRAP: Nid

pub const ID_CAMELLIA192_WRAP: Nid

pub const ID_CAMELLIA256_WRAP: Nid

pub const CAMELLIA_128_ECB: Nid

pub const CAMELLIA_128_OFB128: Nid

pub const CAMELLIA_128_CFB128: Nid

pub const CAMELLIA_192_ECB: Nid

pub const CAMELLIA_192_OFB128: Nid

pub const CAMELLIA_192_CFB128: Nid

pub const CAMELLIA_256_ECB: Nid

pub const CAMELLIA_256_OFB128: Nid

pub const CAMELLIA_256_CFB128: Nid

pub const CAMELLIA_128_CFB1: Nid

pub const CAMELLIA_192_CFB1: Nid

pub const CAMELLIA_256_CFB1: Nid

pub const CAMELLIA_128_CFB8: Nid

pub const CAMELLIA_192_CFB8: Nid

pub const CAMELLIA_256_CFB8: Nid

pub const KISA: Nid

pub const SEED_ECB: Nid

pub const SEED_CBC: Nid

pub const SEED_CFB128: Nid

pub const SEED_OFB128: Nid

pub const HMAC: Nid

pub const CMAC: Nid

pub const RC4_HMAC_MD5: Nid

pub const AES_128_CBC_HMAC_SHA1: Nid

pub const AES_192_CBC_HMAC_SHA1: Nid

pub const AES_256_CBC_HMAC_SHA1: Nid