Struct openssl::nid::Nid

pub struct Nid(_);

A numerical identifier for an OpenSSL object.

Objects in OpenSSL can have a short name, a long name, and a numerical identifier (NID). For convenience, objects are usually represented in source code using these numeric identifiers.

Users should generally not need to create new Nids.

Examples

To view the integer representation of a Nid:

use openssl::nid::Nid;

assert!(Nid::AES_256_GCM.as_raw() == 901);

External Documentation

The following documentation provides context about Nids and their usage in OpenSSL.

Obj_nid2obj

Implementations

impl Nid

pub fn from_raw(raw: c_int) -> Nid

Create a Nid from an integer representation.

pub fn as_raw(&self) -> c_int

Return the integer representation of a Nid.

pub fn create(oid: &str, sn: &str, ln: &str) -> Result<Nid, ErrorStack>

Creates a new Nid for the oid with short name sn and long name ln.

This corresponds to OBJ_create.

pub fn signature_algorithms(&self) -> Option<SignatureAlgorithms>

Returns the Nids of the digest and public key algorithms associated with a signature ID.

This corresponds to OBJ_find_sigid_algs.

This corresponds to OBJ_find_sigid_algs.

pub fn long_name(&self) -> Result<&'static str, ErrorStack>

Returns the string representation of a Nid (long).

This corresponds to OBJ_nid2ln.

pub fn short_name(&self) -> Result<&'static str, ErrorStack>

Returns the string representation of a Nid (short).

This corresponds to OBJ_nid2sn.

pub const UNDEF: Nid = _

pub const ITU_T: Nid = _

pub const CCITT: Nid = _

pub const ISO: Nid = _

pub const JOINT_ISO_ITU_T: Nid = _

pub const JOINT_ISO_CCITT: Nid = _

pub const MEMBER_BODY: Nid = _

pub const IDENTIFIED_ORGANIZATION: Nid = _

pub const HMAC_MD5: Nid = _

pub const HMAC_SHA1: Nid = _

pub const CERTICOM_ARC: Nid = _

pub const INTERNATIONAL_ORGANIZATIONS: Nid = _

pub const WAP: Nid = _

pub const WAP_WSG: Nid = _

pub const SELECTED_ATTRIBUTE_TYPES: Nid = _

pub const CLEARANCE: Nid = _

pub const ISO_US: Nid = _

pub const X9_57: Nid = _

pub const X9CM: Nid = _

pub const DSA: Nid = _

pub const DSAWITHSHA1: Nid = _

pub const ANSI_X9_62: Nid = _

pub const X9_62_PRIME_FIELD: Nid = _

pub const X9_62_CHARACTERISTIC_TWO_FIELD: Nid = _

pub const X9_62_ID_CHARACTERISTIC_TWO_BASIS: Nid = _

pub const X9_62_ONBASIS: Nid = _

pub const X9_62_TPBASIS: Nid = _

pub const X9_62_PPBASIS: Nid = _

pub const X9_62_ID_ECPUBLICKEY: Nid = _

pub const X9_62_C2PNB163V1: Nid = _

pub const X9_62_C2PNB163V2: Nid = _

pub const X9_62_C2PNB163V3: Nid = _

pub const X9_62_C2PNB176V1: Nid = _

pub const X9_62_C2TNB191V1: Nid = _

pub const X9_62_C2TNB191V2: Nid = _

pub const X9_62_C2TNB191V3: Nid = _

pub const X9_62_C2ONB191V4: Nid = _

pub const X9_62_C2ONB191V5: Nid = _

pub const X9_62_C2PNB208W1: Nid = _

pub const X9_62_C2TNB239V1: Nid = _

pub const X9_62_C2TNB239V2: Nid = _

pub const X9_62_C2TNB239V3: Nid = _

pub const X9_62_C2ONB239V4: Nid = _

pub const X9_62_C2ONB239V5: Nid = _

pub const X9_62_C2PNB272W1: Nid = _

pub const X9_62_C2PNB304W1: Nid = _

pub const X9_62_C2TNB359V1: Nid = _

pub const X9_62_C2PNB368W1: Nid = _

pub const X9_62_C2TNB431R1: Nid = _

pub const X9_62_PRIME192V1: Nid = _

pub const X9_62_PRIME192V2: Nid = _

pub const X9_62_PRIME192V3: Nid = _

pub const X9_62_PRIME239V1: Nid = _

pub const X9_62_PRIME239V2: Nid = _

pub const X9_62_PRIME239V3: Nid = _

pub const X9_62_PRIME256V1: Nid = _

pub const ECDSA_WITH_SHA1: Nid = _

pub const ECDSA_WITH_RECOMMENDED: Nid = _

pub const ECDSA_WITH_SPECIFIED: Nid = _

pub const ECDSA_WITH_SHA224: Nid = _

pub const ECDSA_WITH_SHA256: Nid = _

pub const ECDSA_WITH_SHA384: Nid = _

pub const ECDSA_WITH_SHA512: Nid = _

pub const SECP112R1: Nid = _

pub const SECP112R2: Nid = _

pub const SECP128R1: Nid = _

pub const SECP128R2: Nid = _

pub const SECP160K1: Nid = _

pub const SECP160R1: Nid = _

pub const SECP160R2: Nid = _

pub const SECP192K1: Nid = _

pub const SECP224K1: Nid = _

pub const SECP224R1: Nid = _

pub const SECP256K1: Nid = _

pub const SECP384R1: Nid = _

pub const SECP521R1: Nid = _

pub const SECT113R1: Nid = _

pub const SECT113R2: Nid = _

pub const SECT131R1: Nid = _

pub const SECT131R2: Nid = _

pub const SECT163K1: Nid = _

pub const SECT163R1: Nid = _

pub const SECT163R2: Nid = _

pub const SECT193R1: Nid = _

pub const SECT193R2: Nid = _

pub const SECT233K1: Nid = _

pub const SECT233R1: Nid = _

pub const SECT239K1: Nid = _

pub const SECT283K1: Nid = _

pub const SECT283R1: Nid = _

pub const SECT409K1: Nid = _

pub const SECT409R1: Nid = _

pub const SECT571K1: Nid = _

pub const SECT571R1: Nid = _

pub const BRAINPOOL_P256R1: Nid = _

pub const BRAINPOOL_P384R1: Nid = _

pub const BRAINPOOL_P512R1: Nid = _

pub const WAP_WSG_IDM_ECID_WTLS1: Nid = _

pub const WAP_WSG_IDM_ECID_WTLS3: Nid = _

pub const WAP_WSG_IDM_ECID_WTLS4: Nid = _

pub const WAP_WSG_IDM_ECID_WTLS5: Nid = _

pub const WAP_WSG_IDM_ECID_WTLS6: Nid = _

pub const WAP_WSG_IDM_ECID_WTLS7: Nid = _

pub const WAP_WSG_IDM_ECID_WTLS8: Nid = _

pub const WAP_WSG_IDM_ECID_WTLS9: Nid = _

pub const WAP_WSG_IDM_ECID_WTLS10: Nid = _

pub const WAP_WSG_IDM_ECID_WTLS11: Nid = _

pub const WAP_WSG_IDM_ECID_WTLS12: Nid = _

pub const CAST5_CBC: Nid = _

pub const CAST5_ECB: Nid = _

pub const CAST5_CFB64: Nid = _

pub const CAST5_OFB64: Nid = _

pub const PBEWITHMD5ANDCAST5_CBC: Nid = _

pub const ID_PASSWORDBASEDMAC: Nid = _

pub const ID_DHBASEDMAC: Nid = _

pub const RSADSI: Nid = _

pub const PKCS: Nid = _

pub const PKCS1: Nid = _

pub const RSAENCRYPTION: Nid = _

pub const MD2WITHRSAENCRYPTION: Nid = _

pub const MD4WITHRSAENCRYPTION: Nid = _

pub const MD5WITHRSAENCRYPTION: Nid = _

pub const SHA1WITHRSAENCRYPTION: Nid = _

pub const RSAESOAEP: Nid = _

pub const MGF1: Nid = _

pub const RSASSAPSS: Nid = _

pub const SHA256WITHRSAENCRYPTION: Nid = _

pub const SHA384WITHRSAENCRYPTION: Nid = _

pub const SHA512WITHRSAENCRYPTION: Nid = _

pub const SHA224WITHRSAENCRYPTION: Nid = _

pub const PKCS3: Nid = _

pub const DHKEYAGREEMENT: Nid = _

pub const PKCS5: Nid = _

pub const PBEWITHMD2ANDDES_CBC: Nid = _

pub const PBEWITHMD5ANDDES_CBC: Nid = _

pub const PBEWITHMD2ANDRC2_CBC: Nid = _

pub const PBEWITHMD5ANDRC2_CBC: Nid = _

pub const PBEWITHSHA1ANDDES_CBC: Nid = _

pub const PBEWITHSHA1ANDRC2_CBC: Nid = _

pub const ID_PBKDF2: Nid = _

pub const PBES2: Nid = _

pub const PBMAC1: Nid = _

pub const PKCS7: Nid = _

pub const PKCS7_DATA: Nid = _

pub const PKCS7_SIGNED: Nid = _

pub const PKCS7_ENVELOPED: Nid = _

pub const PKCS7_SIGNEDANDENVELOPED: Nid = _

pub const PKCS7_DIGEST: Nid = _

pub const PKCS7_ENCRYPTED: Nid = _

pub const PKCS9: Nid = _

pub const PKCS9_EMAILADDRESS: Nid = _

pub const PKCS9_UNSTRUCTUREDNAME: Nid = _

pub const PKCS9_CONTENTTYPE: Nid = _

pub const PKCS9_MESSAGEDIGEST: Nid = _

pub const PKCS9_SIGNINGTIME: Nid = _

pub const PKCS9_COUNTERSIGNATURE: Nid = _

pub const PKCS9_CHALLENGEPASSWORD: Nid = _

pub const PKCS9_UNSTRUCTUREDADDRESS: Nid = _

pub const PKCS9_EXTCERTATTRIBUTES: Nid = _

pub const EXT_REQ: Nid = _

pub const SMIMECAPABILITIES: Nid = _

pub const SMIME: Nid = _

pub const ID_SMIME_MOD: Nid = _

pub const ID_SMIME_CT: Nid = _

pub const ID_SMIME_AA: Nid = _

pub const ID_SMIME_ALG: Nid = _

pub const ID_SMIME_CD: Nid = _

pub const ID_SMIME_SPQ: Nid = _

pub const ID_SMIME_CTI: Nid = _

pub const ID_SMIME_MOD_CMS: Nid = _

pub const ID_SMIME_MOD_ESS: Nid = _

pub const ID_SMIME_MOD_OID: Nid = _

pub const ID_SMIME_MOD_MSG_V3: Nid = _

pub const ID_SMIME_MOD_ETS_ESIGNATURE_88: Nid = _

pub const ID_SMIME_MOD_ETS_ESIGNATURE_97: Nid = _

pub const ID_SMIME_MOD_ETS_ESIGPOLICY_88: Nid = _

pub const ID_SMIME_MOD_ETS_ESIGPOLICY_97: Nid = _

pub const ID_SMIME_CT_RECEIPT: Nid = _

pub const ID_SMIME_CT_AUTHDATA: Nid = _

pub const ID_SMIME_CT_PUBLISHCERT: Nid = _

pub const ID_SMIME_CT_TSTINFO: Nid = _

pub const ID_SMIME_CT_TDTINFO: Nid = _

pub const ID_SMIME_CT_CONTENTINFO: Nid = _

pub const ID_SMIME_CT_DVCSREQUESTDATA: Nid = _

pub const ID_SMIME_CT_DVCSRESPONSEDATA: Nid = _

pub const ID_SMIME_CT_COMPRESSEDDATA: Nid = _

pub const ID_CT_ASCIITEXTWITHCRLF: Nid = _

pub const ID_SMIME_AA_RECEIPTREQUEST: Nid = _

pub const ID_SMIME_AA_SECURITYLABEL: Nid = _

pub const ID_SMIME_AA_MLEXPANDHISTORY: Nid = _

pub const ID_SMIME_AA_CONTENTHINT: Nid = _

pub const ID_SMIME_AA_MSGSIGDIGEST: Nid = _

pub const ID_SMIME_AA_ENCAPCONTENTTYPE: Nid = _

pub const ID_SMIME_AA_CONTENTIDENTIFIER: Nid = _

pub const ID_SMIME_AA_MACVALUE: Nid = _

pub const ID_SMIME_AA_EQUIVALENTLABELS: Nid = _

pub const ID_SMIME_AA_CONTENTREFERENCE: Nid = _

pub const ID_SMIME_AA_ENCRYPKEYPREF: Nid = _

pub const ID_SMIME_AA_SIGNINGCERTIFICATE: Nid = _

pub const ID_SMIME_AA_SMIMEENCRYPTCERTS: Nid = _

pub const ID_SMIME_AA_TIMESTAMPTOKEN: Nid = _

pub const ID_SMIME_AA_ETS_SIGPOLICYID: Nid = _

pub const ID_SMIME_AA_ETS_COMMITMENTTYPE: Nid = _

pub const ID_SMIME_AA_ETS_SIGNERLOCATION: Nid = _

pub const ID_SMIME_AA_ETS_SIGNERATTR: Nid = _

pub const ID_SMIME_AA_ETS_OTHERSIGCERT: Nid = _

pub const ID_SMIME_AA_ETS_CONTENTTIMESTAMP: Nid = _

pub const ID_SMIME_AA_ETS_CERTIFICATEREFS: Nid = _

pub const ID_SMIME_AA_ETS_REVOCATIONREFS: Nid = _

pub const ID_SMIME_AA_ETS_CERTVALUES: Nid = _

pub const ID_SMIME_AA_ETS_REVOCATIONVALUES: Nid = _

pub const ID_SMIME_AA_ETS_ESCTIMESTAMP: Nid = _

pub const ID_SMIME_AA_ETS_CERTCRLTIMESTAMP: Nid = _

pub const ID_SMIME_AA_ETS_ARCHIVETIMESTAMP: Nid = _

pub const ID_SMIME_AA_SIGNATURETYPE: Nid = _

pub const ID_SMIME_AA_DVCS_DVC: Nid = _

pub const ID_SMIME_ALG_ESDHWITH3DES: Nid = _

pub const ID_SMIME_ALG_ESDHWITHRC2: Nid = _

pub const ID_SMIME_ALG_3DESWRAP: Nid = _

pub const ID_SMIME_ALG_RC2WRAP: Nid = _

pub const ID_SMIME_ALG_ESDH: Nid = _

pub const ID_SMIME_ALG_CMS3DESWRAP: Nid = _

pub const ID_SMIME_ALG_CMSRC2WRAP: Nid = _

pub const ID_ALG_PWRI_KEK: Nid = _

pub const ID_SMIME_CD_LDAP: Nid = _

pub const ID_SMIME_SPQ_ETS_SQT_URI: Nid = _

pub const ID_SMIME_SPQ_ETS_SQT_UNOTICE: Nid = _

pub const ID_SMIME_CTI_ETS_PROOFOFORIGIN: Nid = _

pub const ID_SMIME_CTI_ETS_PROOFOFRECEIPT: Nid = _

pub const ID_SMIME_CTI_ETS_PROOFOFDELIVERY: Nid = _

pub const ID_SMIME_CTI_ETS_PROOFOFSENDER: Nid = _

pub const ID_SMIME_CTI_ETS_PROOFOFAPPROVAL: Nid = _

pub const ID_SMIME_CTI_ETS_PROOFOFCREATION: Nid = _

pub const FRIENDLYNAME: Nid = _

pub const LOCALKEYID: Nid = _

pub const MS_CSP_NAME: Nid = _

pub const LOCALKEYSET: Nid = _

pub const X509CERTIFICATE: Nid = _

pub const SDSICERTIFICATE: Nid = _

pub const X509CRL: Nid = _

pub const PBE_WITHSHA1AND128BITRC4: Nid = _

pub const PBE_WITHSHA1AND40BITRC4: Nid = _

pub const PBE_WITHSHA1AND3_KEY_TRIPLEDES_CBC: Nid = _

pub const PBE_WITHSHA1AND2_KEY_TRIPLEDES_CBC: Nid = _

pub const PBE_WITHSHA1AND128BITRC2_CBC: Nid = _

pub const PBE_WITHSHA1AND40BITRC2_CBC: Nid = _

pub const KEYBAG: Nid = _

pub const PKCS8SHROUDEDKEYBAG: Nid = _

pub const CERTBAG: Nid = _

pub const CRLBAG: Nid = _

pub const SECRETBAG: Nid = _

pub const SAFECONTENTSBAG: Nid = _

pub const MD2: Nid = _

pub const MD4: Nid = _

pub const MD5: Nid = _

pub const MD5_SHA1: Nid = _

pub const HMACWITHMD5: Nid = _

pub const HMACWITHSHA1: Nid = _

pub const HMACWITHSHA224: Nid = _

pub const HMACWITHSHA256: Nid = _

pub const HMACWITHSHA384: Nid = _

pub const HMACWITHSHA512: Nid = _

pub const RC2_CBC: Nid = _

pub const RC2_ECB: Nid = _

pub const RC2_CFB64: Nid = _

pub const RC2_OFB64: Nid = _

pub const RC2_40_CBC: Nid = _

pub const RC2_64_CBC: Nid = _

pub const RC4: Nid = _

pub const RC4_40: Nid = _

pub const DES_EDE3_CBC: Nid = _

pub const RC5_CBC: Nid = _

pub const RC5_ECB: Nid = _

pub const RC5_CFB64: Nid = _

pub const RC5_OFB64: Nid = _

pub const MS_EXT_REQ: Nid = _

pub const MS_CODE_IND: Nid = _

pub const MS_CODE_COM: Nid = _

pub const MS_CTL_SIGN: Nid = _

pub const MS_SGC: Nid = _

pub const MS_EFS: Nid = _

pub const MS_SMARTCARD_LOGIN: Nid = _

pub const MS_UPN: Nid = _

pub const IDEA_CBC: Nid = _

pub const IDEA_ECB: Nid = _

pub const IDEA_CFB64: Nid = _

pub const IDEA_OFB64: Nid = _

pub const BF_CBC: Nid = _

pub const BF_ECB: Nid = _

pub const BF_CFB64: Nid = _

pub const BF_OFB64: Nid = _

pub const ID_PKIX: Nid = _

pub const ID_PKIX_MOD: Nid = _

pub const ID_PE: Nid = _

pub const ID_QT: Nid = _

pub const ID_KP: Nid = _

pub const ID_IT: Nid = _

pub const ID_PKIP: Nid = _

pub const ID_ALG: Nid = _

pub const ID_CMC: Nid = _

pub const ID_ON: Nid = _

pub const ID_PDA: Nid = _

pub const ID_ACA: Nid = _

pub const ID_QCS: Nid = _

pub const ID_CCT: Nid = _

pub const ID_PPL: Nid = _

pub const ID_AD: Nid = _

pub const ID_PKIX1_EXPLICIT_88: Nid = _

pub const ID_PKIX1_IMPLICIT_88: Nid = _

pub const ID_PKIX1_EXPLICIT_93: Nid = _

pub const ID_PKIX1_IMPLICIT_93: Nid = _

pub const ID_MOD_CRMF: Nid = _

pub const ID_MOD_CMC: Nid = _

pub const ID_MOD_KEA_PROFILE_88: Nid = _

pub const ID_MOD_KEA_PROFILE_93: Nid = _

pub const ID_MOD_CMP: Nid = _

pub const ID_MOD_QUALIFIED_CERT_88: Nid = _

pub const ID_MOD_QUALIFIED_CERT_93: Nid = _

pub const ID_MOD_ATTRIBUTE_CERT: Nid = _

pub const ID_MOD_TIMESTAMP_PROTOCOL: Nid = _

pub const ID_MOD_OCSP: Nid = _

pub const ID_MOD_DVCS: Nid = _

pub const ID_MOD_CMP2000: Nid = _

pub const INFO_ACCESS: Nid = _

pub const BIOMETRICINFO: Nid = _

pub const QCSTATEMENTS: Nid = _

pub const AC_AUDITENTITY: Nid = _

pub const AC_TARGETING: Nid = _

pub const AACONTROLS: Nid = _

pub const SBGP_IPADDRBLOCK: Nid = _

pub const SBGP_AUTONOMOUSSYSNUM: Nid = _

pub const SBGP_ROUTERIDENTIFIER: Nid = _

pub const AC_PROXYING: Nid = _

pub const SINFO_ACCESS: Nid = _

pub const PROXYCERTINFO: Nid = _

pub const ID_QT_CPS: Nid = _

pub const ID_QT_UNOTICE: Nid = _

pub const TEXTNOTICE: Nid = _

pub const SERVER_AUTH: Nid = _

pub const CLIENT_AUTH: Nid = _

pub const CODE_SIGN: Nid = _

pub const EMAIL_PROTECT: Nid = _

pub const IPSECENDSYSTEM: Nid = _

pub const IPSECTUNNEL: Nid = _

pub const IPSECUSER: Nid = _

pub const TIME_STAMP: Nid = _

pub const OCSP_SIGN: Nid = _

pub const DVCS: Nid = _

pub const ID_IT_CAPROTENCCERT: Nid = _

pub const ID_IT_SIGNKEYPAIRTYPES: Nid = _

pub const ID_IT_ENCKEYPAIRTYPES: Nid = _

pub const ID_IT_PREFERREDSYMMALG: Nid = _

pub const ID_IT_CAKEYUPDATEINFO: Nid = _

pub const ID_IT_CURRENTCRL: Nid = _

pub const ID_IT_UNSUPPORTEDOIDS: Nid = _

pub const ID_IT_SUBSCRIPTIONREQUEST: Nid = _

pub const ID_IT_SUBSCRIPTIONRESPONSE: Nid = _

pub const ID_IT_KEYPAIRPARAMREQ: Nid = _

pub const ID_IT_KEYPAIRPARAMREP: Nid = _

pub const ID_IT_REVPASSPHRASE: Nid = _

pub const ID_IT_IMPLICITCONFIRM: Nid = _

pub const ID_IT_CONFIRMWAITTIME: Nid = _

pub const ID_IT_ORIGPKIMESSAGE: Nid = _

pub const ID_IT_SUPPLANGTAGS: Nid = _

pub const ID_REGCTRL: Nid = _

pub const ID_REGINFO: Nid = _

pub const ID_REGCTRL_REGTOKEN: Nid = _

pub const ID_REGCTRL_AUTHENTICATOR: Nid = _

pub const ID_REGCTRL_PKIPUBLICATIONINFO: Nid = _

pub const ID_REGCTRL_PKIARCHIVEOPTIONS: Nid = _

pub const ID_REGCTRL_OLDCERTID: Nid = _

pub const ID_REGCTRL_PROTOCOLENCRKEY: Nid = _

pub const ID_REGINFO_UTF8PAIRS: Nid = _

pub const ID_REGINFO_CERTREQ: Nid = _

pub const ID_ALG_DES40: Nid = _

pub const ID_ALG_NOSIGNATURE: Nid = _

pub const ID_ALG_DH_SIG_HMAC_SHA1: Nid = _

pub const ID_ALG_DH_POP: Nid = _

pub const ID_CMC_STATUSINFO: Nid = _

pub const ID_CMC_IDENTIFICATION: Nid = _

pub const ID_CMC_IDENTITYPROOF: Nid = _

pub const ID_CMC_DATARETURN: Nid = _

pub const ID_CMC_TRANSACTIONID: Nid = _

pub const ID_CMC_SENDERNONCE: Nid = _

pub const ID_CMC_RECIPIENTNONCE: Nid = _

pub const ID_CMC_ADDEXTENSIONS: Nid = _

pub const ID_CMC_ENCRYPTEDPOP: Nid = _

pub const ID_CMC_DECRYPTEDPOP: Nid = _

pub const ID_CMC_LRAPOPWITNESS: Nid = _

pub const ID_CMC_GETCERT: Nid = _

pub const ID_CMC_GETCRL: Nid = _

pub const ID_CMC_REVOKEREQUEST: Nid = _

pub const ID_CMC_REGINFO: Nid = _

pub const ID_CMC_RESPONSEINFO: Nid = _

pub const ID_CMC_QUERYPENDING: Nid = _

pub const ID_CMC_POPLINKRANDOM: Nid = _

pub const ID_CMC_POPLINKWITNESS: Nid = _

pub const ID_CMC_CONFIRMCERTACCEPTANCE: Nid = _

pub const ID_ON_PERSONALDATA: Nid = _

pub const ID_ON_PERMANENTIDENTIFIER: Nid = _

pub const ID_PDA_DATEOFBIRTH: Nid = _

pub const ID_PDA_PLACEOFBIRTH: Nid = _

pub const ID_PDA_GENDER: Nid = _

pub const ID_PDA_COUNTRYOFCITIZENSHIP: Nid = _

pub const ID_PDA_COUNTRYOFRESIDENCE: Nid = _

pub const ID_ACA_AUTHENTICATIONINFO: Nid = _

pub const ID_ACA_ACCESSIDENTITY: Nid = _

pub const ID_ACA_CHARGINGIDENTITY: Nid = _

pub const ID_ACA_GROUP: Nid = _

pub const ID_ACA_ROLE: Nid = _

pub const ID_ACA_ENCATTRS: Nid = _

pub const ID_QCS_PKIXQCSYNTAX_V1: Nid = _

pub const ID_CCT_CRS: Nid = _

pub const ID_CCT_PKIDATA: Nid = _

pub const ID_CCT_PKIRESPONSE: Nid = _

pub const ID_PPL_ANYLANGUAGE: Nid = _

pub const ID_PPL_INHERITALL: Nid = _

pub const INDEPENDENT: Nid = _

pub const AD_OCSP: Nid = _

pub const AD_CA_ISSUERS: Nid = _

pub const AD_TIMESTAMPING: Nid = _

pub const AD_DVCS: Nid = _

pub const CAREPOSITORY: Nid = _

pub const ID_PKIX_OCSP_BASIC: Nid = _

pub const ID_PKIX_OCSP_NONCE: Nid = _

pub const ID_PKIX_OCSP_CRLID: Nid = _

pub const ID_PKIX_OCSP_ACCEPTABLERESPONSES: Nid = _

pub const ID_PKIX_OCSP_NOCHECK: Nid = _

pub const ID_PKIX_OCSP_ARCHIVECUTOFF: Nid = _

pub const ID_PKIX_OCSP_SERVICELOCATOR: Nid = _

pub const ID_PKIX_OCSP_EXTENDEDSTATUS: Nid = _

pub const ID_PKIX_OCSP_VALID: Nid = _

pub const ID_PKIX_OCSP_PATH: Nid = _

pub const ID_PKIX_OCSP_TRUSTROOT: Nid = _

pub const ALGORITHM: Nid = _

pub const MD5WITHRSA: Nid = _

pub const DES_ECB: Nid = _

pub const DES_CBC: Nid = _

pub const DES_OFB64: Nid = _

pub const DES_CFB64: Nid = _

pub const RSASIGNATURE: Nid = _

pub const DSA_2: Nid = _

pub const DSAWITHSHA: Nid = _

pub const SHAWITHRSAENCRYPTION: Nid = _

pub const DES_EDE_ECB: Nid = _

pub const DES_EDE3_ECB: Nid = _

pub const DES_EDE_CBC: Nid = _

pub const DES_EDE_CFB64: Nid = _

pub const DES_EDE3_CFB64: Nid = _

pub const DES_EDE_OFB64: Nid = _

pub const DES_EDE3_OFB64: Nid = _

pub const DESX_CBC: Nid = _

pub const SHA: Nid = _

pub const SHA1: Nid = _

pub const DSAWITHSHA1_2: Nid = _

pub const SHA1WITHRSA: Nid = _

pub const RIPEMD160: Nid = _

pub const RIPEMD160WITHRSA: Nid = _

pub const SXNET: Nid = _

pub const X500: Nid = _

pub const X509: Nid = _

pub const COMMONNAME: Nid = _

pub const SURNAME: Nid = _

pub const SERIALNUMBER: Nid = _

pub const COUNTRYNAME: Nid = _

pub const LOCALITYNAME: Nid = _

pub const STATEORPROVINCENAME: Nid = _

pub const STREETADDRESS: Nid = _

pub const ORGANIZATIONNAME: Nid = _

pub const ORGANIZATIONALUNITNAME: Nid = _

pub const TITLE: Nid = _

pub const DESCRIPTION: Nid = _

pub const SEARCHGUIDE: Nid = _

pub const BUSINESSCATEGORY: Nid = _

pub const POSTALADDRESS: Nid = _

pub const POSTALCODE: Nid = _

pub const POSTOFFICEBOX: Nid = _

pub const PHYSICALDELIVERYOFFICENAME: Nid = _

pub const TELEPHONENUMBER: Nid = _

pub const TELEXNUMBER: Nid = _

pub const TELETEXTERMINALIDENTIFIER: Nid = _

pub const FACSIMILETELEPHONENUMBER: Nid = _

pub const X121ADDRESS: Nid = _

pub const INTERNATIONALISDNNUMBER: Nid = _

pub const REGISTEREDADDRESS: Nid = _

pub const DESTINATIONINDICATOR: Nid = _

pub const PREFERREDDELIVERYMETHOD: Nid = _

pub const PRESENTATIONADDRESS: Nid = _

pub const SUPPORTEDAPPLICATIONCONTEXT: Nid = _

pub const MEMBER: Nid = _

pub const OWNER: Nid = _

pub const ROLEOCCUPANT: Nid = _

pub const SEEALSO: Nid = _

pub const USERPASSWORD: Nid = _

pub const USERCERTIFICATE: Nid = _

pub const CACERTIFICATE: Nid = _

pub const AUTHORITYREVOCATIONLIST: Nid = _

pub const CERTIFICATEREVOCATIONLIST: Nid = _

pub const CROSSCERTIFICATEPAIR: Nid = _

pub const NAME: Nid = _

pub const GIVENNAME: Nid = _

pub const INITIALS: Nid = _

pub const GENERATIONQUALIFIER: Nid = _

pub const X500UNIQUEIDENTIFIER: Nid = _

pub const DNQUALIFIER: Nid = _

pub const ENHANCEDSEARCHGUIDE: Nid = _

pub const PROTOCOLINFORMATION: Nid = _

pub const DISTINGUISHEDNAME: Nid = _

pub const UNIQUEMEMBER: Nid = _

pub const HOUSEIDENTIFIER: Nid = _

pub const SUPPORTEDALGORITHMS: Nid = _

pub const DELTAREVOCATIONLIST: Nid = _

pub const DMDNAME: Nid = _

pub const PSEUDONYM: Nid = _

pub const ROLE: Nid = _

pub const X500ALGORITHMS: Nid = _

pub const RSA: Nid = _

pub const MDC2WITHRSA: Nid = _

pub const MDC2: Nid = _

pub const ID_CE: Nid = _

pub const SUBJECT_DIRECTORY_ATTRIBUTES: Nid = _

pub const SUBJECT_KEY_IDENTIFIER: Nid = _

pub const KEY_USAGE: Nid = _

pub const PRIVATE_KEY_USAGE_PERIOD: Nid = _

pub const SUBJECT_ALT_NAME: Nid = _

pub const ISSUER_ALT_NAME: Nid = _

pub const BASIC_CONSTRAINTS: Nid = _

pub const CRL_NUMBER: Nid = _

pub const CRL_REASON: Nid = _

pub const INVALIDITY_DATE: Nid = _

pub const DELTA_CRL: Nid = _

pub const ISSUING_DISTRIBUTION_POINT: Nid = _

pub const CERTIFICATE_ISSUER: Nid = _

pub const NAME_CONSTRAINTS: Nid = _

pub const CRL_DISTRIBUTION_POINTS: Nid = _

pub const CERTIFICATE_POLICIES: Nid = _

pub const ANY_POLICY: Nid = _

pub const POLICY_MAPPINGS: Nid = _

pub const AUTHORITY_KEY_IDENTIFIER: Nid = _

pub const POLICY_CONSTRAINTS: Nid = _

pub const EXT_KEY_USAGE: Nid = _

pub const FRESHEST_CRL: Nid = _

pub const INHIBIT_ANY_POLICY: Nid = _

pub const TARGET_INFORMATION: Nid = _

pub const NO_REV_AVAIL: Nid = _

pub const ANYEXTENDEDKEYUSAGE: Nid = _

pub const NETSCAPE: Nid = _

pub const NETSCAPE_CERT_EXTENSION: Nid = _

pub const NETSCAPE_DATA_TYPE: Nid = _

pub const NETSCAPE_CERT_TYPE: Nid = _

pub const NETSCAPE_BASE_URL: Nid = _

pub const NETSCAPE_REVOCATION_URL: Nid = _

pub const NETSCAPE_CA_REVOCATION_URL: Nid = _

pub const NETSCAPE_RENEWAL_URL: Nid = _

pub const NETSCAPE_CA_POLICY_URL: Nid = _

pub const NETSCAPE_SSL_SERVER_NAME: Nid = _

pub const NETSCAPE_COMMENT: Nid = _

pub const NETSCAPE_CERT_SEQUENCE: Nid = _

pub const NS_SGC: Nid = _

pub const ORG: Nid = _

pub const DOD: Nid = _

pub const IANA: Nid = _

pub const DIRECTORY: Nid = _

pub const MANAGEMENT: Nid = _

pub const EXPERIMENTAL: Nid = _

pub const PRIVATE: Nid = _

pub const SECURITY: Nid = _

pub const SNMPV2: Nid = _

pub const MAIL: Nid = _

pub const ENTERPRISES: Nid = _

pub const DCOBJECT: Nid = _

pub const MIME_MHS: Nid = _

pub const MIME_MHS_HEADINGS: Nid = _

pub const MIME_MHS_BODIES: Nid = _

pub const ID_HEX_PARTIAL_MESSAGE: Nid = _

pub const ID_HEX_MULTIPART_MESSAGE: Nid = _

pub const ZLIB_COMPRESSION: Nid = _

pub const AES_128_ECB: Nid = _

pub const AES_128_CBC: Nid = _

pub const AES_128_OFB128: Nid = _

pub const AES_128_CFB128: Nid = _

pub const ID_AES128_WRAP: Nid = _

pub const AES_128_GCM: Nid = _

pub const AES_128_CCM: Nid = _

pub const ID_AES128_WRAP_PAD: Nid = _

pub const AES_192_ECB: Nid = _

pub const AES_192_CBC: Nid = _

pub const AES_192_OFB128: Nid = _

pub const AES_192_CFB128: Nid = _

pub const ID_AES192_WRAP: Nid = _

pub const AES_192_GCM: Nid = _

pub const AES_192_CCM: Nid = _

pub const ID_AES192_WRAP_PAD: Nid = _

pub const AES_256_ECB: Nid = _

pub const AES_256_CBC: Nid = _

pub const AES_256_OFB128: Nid = _

pub const AES_256_CFB128: Nid = _

pub const ID_AES256_WRAP: Nid = _

pub const AES_256_GCM: Nid = _

pub const AES_256_CCM: Nid = _

pub const ID_AES256_WRAP_PAD: Nid = _

pub const AES_128_CFB1: Nid = _

pub const AES_192_CFB1: Nid = _

pub const AES_256_CFB1: Nid = _

pub const AES_128_CFB8: Nid = _

pub const AES_192_CFB8: Nid = _

pub const AES_256_CFB8: Nid = _

pub const AES_128_CTR: Nid = _

pub const AES_192_CTR: Nid = _

pub const AES_256_CTR: Nid = _

pub const AES_128_XTS: Nid = _

pub const AES_256_XTS: Nid = _

pub const DES_CFB1: Nid = _

pub const DES_CFB8: Nid = _

pub const DES_EDE3_CFB1: Nid = _

pub const DES_EDE3_CFB8: Nid = _

pub const SHA256: Nid = _

pub const SHA384: Nid = _

pub const SHA512: Nid = _

pub const SHA224: Nid = _

pub const DSA_WITH_SHA224: Nid = _

pub const DSA_WITH_SHA256: Nid = _

pub const HOLD_INSTRUCTION_CODE: Nid = _

pub const HOLD_INSTRUCTION_NONE: Nid = _

pub const HOLD_INSTRUCTION_CALL_ISSUER: Nid = _

pub const HOLD_INSTRUCTION_REJECT: Nid = _

pub const DATA: Nid = _

pub const PSS: Nid = _

pub const UCL: Nid = _

pub const PILOT: Nid = _

pub const PILOTATTRIBUTETYPE: Nid = _

pub const PILOTATTRIBUTESYNTAX: Nid = _

pub const PILOTOBJECTCLASS: Nid = _

pub const PILOTGROUPS: Nid = _

pub const IA5STRINGSYNTAX: Nid = _

pub const CASEIGNOREIA5STRINGSYNTAX: Nid = _

pub const PILOTOBJECT: Nid = _

pub const PILOTPERSON: Nid = _

pub const ACCOUNT: Nid = _

pub const DOCUMENT: Nid = _

pub const ROOM: Nid = _

pub const DOCUMENTSERIES: Nid = _

pub const DOMAIN: Nid = _

pub const RFC822LOCALPART: Nid = _

pub const DNSDOMAIN: Nid = _

pub const DOMAINRELATEDOBJECT: Nid = _

pub const FRIENDLYCOUNTRY: Nid = _

pub const SIMPLESECURITYOBJECT: Nid = _

pub const PILOTORGANIZATION: Nid = _

pub const PILOTDSA: Nid = _

pub const QUALITYLABELLEDDATA: Nid = _

pub const USERID: Nid = _

pub const TEXTENCODEDORADDRESS: Nid = _

pub const RFC822MAILBOX: Nid = _

pub const INFO: Nid = _

pub const FAVOURITEDRINK: Nid = _

pub const ROOMNUMBER: Nid = _

pub const PHOTO: Nid = _

pub const USERCLASS: Nid = _

pub const HOST: Nid = _

pub const MANAGER: Nid = _

pub const DOCUMENTIDENTIFIER: Nid = _

pub const DOCUMENTTITLE: Nid = _

pub const DOCUMENTVERSION: Nid = _

pub const DOCUMENTAUTHOR: Nid = _

pub const DOCUMENTLOCATION: Nid = _

pub const HOMETELEPHONENUMBER: Nid = _

pub const SECRETARY: Nid = _

pub const OTHERMAILBOX: Nid = _

pub const LASTMODIFIEDTIME: Nid = _

pub const LASTMODIFIEDBY: Nid = _

pub const DOMAINCOMPONENT: Nid = _

pub const ARECORD: Nid = _

pub const PILOTATTRIBUTETYPE27: Nid = _

pub const MXRECORD: Nid = _

pub const NSRECORD: Nid = _

pub const SOARECORD: Nid = _

pub const CNAMERECORD: Nid = _

pub const ASSOCIATEDDOMAIN: Nid = _

pub const ASSOCIATEDNAME: Nid = _

pub const HOMEPOSTALADDRESS: Nid = _

pub const PERSONALTITLE: Nid = _

pub const MOBILETELEPHONENUMBER: Nid = _

pub const PAGERTELEPHONENUMBER: Nid = _

pub const FRIENDLYCOUNTRYNAME: Nid = _

pub const ORGANIZATIONALSTATUS: Nid = _

pub const JANETMAILBOX: Nid = _

pub const MAILPREFERENCEOPTION: Nid = _

pub const BUILDINGNAME: Nid = _

pub const DSAQUALITY: Nid = _

pub const SINGLELEVELQUALITY: Nid = _

pub const SUBTREEMINIMUMQUALITY: Nid = _

pub const SUBTREEMAXIMUMQUALITY: Nid = _

pub const PERSONALSIGNATURE: Nid = _

pub const DITREDIRECT: Nid = _

pub const AUDIO: Nid = _

pub const DOCUMENTPUBLISHER: Nid = _

pub const ID_SET: Nid = _

pub const SET_CTYPE: Nid = _

pub const SET_MSGEXT: Nid = _

pub const SET_ATTR: Nid = _

pub const SET_POLICY: Nid = _

pub const SET_CERTEXT: Nid = _

pub const SET_BRAND: Nid = _

pub const SETCT_PANDATA: Nid = _

pub const SETCT_PANTOKEN: Nid = _

pub const SETCT_PANONLY: Nid = _

pub const SETCT_OIDATA: Nid = _

pub const SETCT_PI: Nid = _

pub const SETCT_PIDATA: Nid = _

pub const SETCT_PIDATAUNSIGNED: Nid = _

pub const SETCT_HODINPUT: Nid = _

pub const SETCT_AUTHRESBAGGAGE: Nid = _

pub const SETCT_AUTHREVREQBAGGAGE: Nid = _

pub const SETCT_AUTHREVRESBAGGAGE: Nid = _

pub const SETCT_CAPTOKENSEQ: Nid = _

pub const SETCT_PINITRESDATA: Nid = _

pub const SETCT_PI_TBS: Nid = _

pub const SETCT_PRESDATA: Nid = _

pub const SETCT_AUTHREQTBS: Nid = _

pub const SETCT_AUTHRESTBS: Nid = _

pub const SETCT_AUTHRESTBSX: Nid = _

pub const SETCT_AUTHTOKENTBS: Nid = _

pub const SETCT_CAPTOKENDATA: Nid = _

pub const SETCT_CAPTOKENTBS: Nid = _

pub const SETCT_ACQCARDCODEMSG: Nid = _

pub const SETCT_AUTHREVREQTBS: Nid = _

pub const SETCT_AUTHREVRESDATA: Nid = _

pub const SETCT_AUTHREVRESTBS: Nid = _

pub const SETCT_CAPREQTBS: Nid = _

pub const SETCT_CAPREQTBSX: Nid = _

pub const SETCT_CAPRESDATA: Nid = _

pub const SETCT_CAPREVREQTBS: Nid = _

pub const SETCT_CAPREVREQTBSX: Nid = _

pub const SETCT_CAPREVRESDATA: Nid = _

pub const SETCT_CREDREQTBS: Nid = _

pub const SETCT_CREDREQTBSX: Nid = _

pub const SETCT_CREDRESDATA: Nid = _

pub const SETCT_CREDREVREQTBS: Nid = _

pub const SETCT_CREDREVREQTBSX: Nid = _

pub const SETCT_CREDREVRESDATA: Nid = _

pub const SETCT_PCERTREQDATA: Nid = _

pub const SETCT_PCERTRESTBS: Nid = _

pub const SETCT_BATCHADMINREQDATA: Nid = _

pub const SETCT_BATCHADMINRESDATA: Nid = _

pub const SETCT_CARDCINITRESTBS: Nid = _

pub const SETCT_MEAQCINITRESTBS: Nid = _

pub const SETCT_REGFORMRESTBS: Nid = _

pub const SETCT_CERTREQDATA: Nid = _

pub const SETCT_CERTREQTBS: Nid = _

pub const SETCT_CERTRESDATA: Nid = _

pub const SETCT_CERTINQREQTBS: Nid = _

pub const SETCT_ERRORTBS: Nid = _

pub const SETCT_PIDUALSIGNEDTBE: Nid = _

pub const SETCT_PIUNSIGNEDTBE: Nid = _

pub const SETCT_AUTHREQTBE: Nid = _

pub const SETCT_AUTHRESTBE: Nid = _

pub const SETCT_AUTHRESTBEX: Nid = _

pub const SETCT_AUTHTOKENTBE: Nid = _

pub const SETCT_CAPTOKENTBE: Nid = _

pub const SETCT_CAPTOKENTBEX: Nid = _

pub const SETCT_ACQCARDCODEMSGTBE: Nid = _

pub const SETCT_AUTHREVREQTBE: Nid = _

pub const SETCT_AUTHREVRESTBE: Nid = _

pub const SETCT_AUTHREVRESTBEB: Nid = _

pub const SETCT_CAPREQTBE: Nid = _

pub const SETCT_CAPREQTBEX: Nid = _

pub const SETCT_CAPRESTBE: Nid = _

pub const SETCT_CAPREVREQTBE: Nid = _

pub const SETCT_CAPREVREQTBEX: Nid = _

pub const SETCT_CAPREVRESTBE: Nid = _

pub const SETCT_CREDREQTBE: Nid = _

pub const SETCT_CREDREQTBEX: Nid = _

pub const SETCT_CREDRESTBE: Nid = _

pub const SETCT_CREDREVREQTBE: Nid = _

pub const SETCT_CREDREVREQTBEX: Nid = _

pub const SETCT_CREDREVRESTBE: Nid = _

pub const SETCT_BATCHADMINREQTBE: Nid = _

pub const SETCT_BATCHADMINRESTBE: Nid = _

pub const SETCT_REGFORMREQTBE: Nid = _

pub const SETCT_CERTREQTBE: Nid = _

pub const SETCT_CERTREQTBEX: Nid = _

pub const SETCT_CERTRESTBE: Nid = _

pub const SETCT_CRLNOTIFICATIONTBS: Nid = _

pub const SETCT_CRLNOTIFICATIONRESTBS: Nid = _

pub const SETCT_BCIDISTRIBUTIONTBS: Nid = _

pub const SETEXT_GENCRYPT: Nid = _

pub const SETEXT_MIAUTH: Nid = _

pub const SETEXT_PINSECURE: Nid = _

pub const SETEXT_PINANY: Nid = _

pub const SETEXT_TRACK2: Nid = _

pub const SETEXT_CV: Nid = _

pub const SET_POLICY_ROOT: Nid = _

pub const SETCEXT_HASHEDROOT: Nid = _

pub const SETCEXT_CERTTYPE: Nid = _

pub const SETCEXT_MERCHDATA: Nid = _

pub const SETCEXT_CCERTREQUIRED: Nid = _

pub const SETCEXT_TUNNELING: Nid = _

pub const SETCEXT_SETEXT: Nid = _

pub const SETCEXT_SETQUALF: Nid = _

pub const SETCEXT_PGWYCAPABILITIES: Nid = _

pub const SETCEXT_TOKENIDENTIFIER: Nid = _

pub const SETCEXT_TRACK2DATA: Nid = _

pub const SETCEXT_TOKENTYPE: Nid = _

pub const SETCEXT_ISSUERCAPABILITIES: Nid = _

pub const SETATTR_CERT: Nid = _

pub const SETATTR_PGWYCAP: Nid = _

pub const SETATTR_TOKENTYPE: Nid = _

pub const SETATTR_ISSCAP: Nid = _

pub const SET_ROOTKEYTHUMB: Nid = _

pub const SET_ADDPOLICY: Nid = _

pub const SETATTR_TOKEN_EMV: Nid = _

pub const SETATTR_TOKEN_B0PRIME: Nid = _

pub const SETATTR_ISSCAP_CVM: Nid = _

pub const SETATTR_ISSCAP_T2: Nid = _

pub const SETATTR_ISSCAP_SIG: Nid = _

pub const SETATTR_GENCRYPTGRM: Nid = _

pub const SETATTR_T2ENC: Nid = _

pub const SETATTR_T2CLEARTXT: Nid = _

pub const SETATTR_TOKICCSIG: Nid = _

pub const SETATTR_SECDEVSIG: Nid = _

pub const SET_BRAND_IATA_ATA: Nid = _

pub const SET_BRAND_DINERS: Nid = _

pub const SET_BRAND_AMERICANEXPRESS: Nid = _

pub const SET_BRAND_JCB: Nid = _

pub const SET_BRAND_VISA: Nid = _

pub const SET_BRAND_MASTERCARD: Nid = _

pub const SET_BRAND_NOVUS: Nid = _

pub const DES_CDMF: Nid = _

pub const RSAOAEPENCRYPTIONSET: Nid = _

pub const IPSEC3: Nid = _

pub const IPSEC4: Nid = _

pub const WHIRLPOOL: Nid = _

pub const CRYPTOPRO: Nid = _

pub const CRYPTOCOM: Nid = _

pub const ID_GOSTR3411_94_WITH_GOSTR3410_2001: Nid = _

pub const ID_GOSTR3411_94_WITH_GOSTR3410_94: Nid = _

pub const ID_GOSTR3411_94: Nid = _

pub const ID_HMACGOSTR3411_94: Nid = _

pub const ID_GOSTR3410_2001: Nid = _

pub const ID_GOSTR3410_94: Nid = _

pub const ID_GOST28147_89: Nid = _

pub const GOST89_CNT: Nid = _

pub const ID_GOST28147_89_MAC: Nid = _

pub const ID_GOSTR3411_94_PRF: Nid = _

pub const ID_GOSTR3410_2001DH: Nid = _

pub const ID_GOSTR3410_94DH: Nid = _

pub const ID_GOST28147_89_CRYPTOPRO_KEYMESHING: Nid = _

pub const ID_GOST28147_89_NONE_KEYMESHING: Nid = _

pub const ID_GOSTR3411_94_TESTPARAMSET: Nid = _

pub const ID_GOSTR3411_94_CRYPTOPROPARAMSET: Nid = _

pub const ID_GOST28147_89_TESTPARAMSET: Nid = _

pub const ID_GOST28147_89_CRYPTOPRO_A_PARAMSET: Nid = _

pub const ID_GOST28147_89_CRYPTOPRO_B_PARAMSET: Nid = _

pub const ID_GOST28147_89_CRYPTOPRO_C_PARAMSET: Nid = _

pub const ID_GOST28147_89_CRYPTOPRO_D_PARAMSET: Nid = _

pub const ID_GOST28147_89_CRYPTOPRO_OSCAR_1_1_PARAMSET: Nid = _

pub const ID_GOST28147_89_CRYPTOPRO_OSCAR_1_0_PARAMSET: Nid = _

pub const ID_GOST28147_89_CRYPTOPRO_RIC_1_PARAMSET: Nid = _

pub const ID_GOSTR3410_94_TESTPARAMSET: Nid = _

pub const ID_GOSTR3410_94_CRYPTOPRO_A_PARAMSET: Nid = _

pub const ID_GOSTR3410_94_CRYPTOPRO_B_PARAMSET: Nid = _

pub const ID_GOSTR3410_94_CRYPTOPRO_C_PARAMSET: Nid = _

pub const ID_GOSTR3410_94_CRYPTOPRO_D_PARAMSET: Nid = _

pub const ID_GOSTR3410_94_CRYPTOPRO_XCHA_PARAMSET: Nid = _

pub const ID_GOSTR3410_94_CRYPTOPRO_XCHB_PARAMSET: Nid = _

pub const ID_GOSTR3410_94_CRYPTOPRO_XCHC_PARAMSET: Nid = _

pub const ID_GOSTR3410_2001_TESTPARAMSET: Nid = _

pub const ID_GOSTR3410_2001_CRYPTOPRO_A_PARAMSET: Nid = _

pub const ID_GOSTR3410_2001_CRYPTOPRO_B_PARAMSET: Nid = _

pub const ID_GOSTR3410_2001_CRYPTOPRO_C_PARAMSET: Nid = _

pub const ID_GOSTR3410_2001_CRYPTOPRO_XCHA_PARAMSET: Nid = _

pub const ID_GOSTR3410_2001_CRYPTOPRO_XCHB_PARAMSET: Nid = _

pub const ID_GOSTR3410_94_A: Nid = _

pub const ID_GOSTR3410_94_ABIS: Nid = _

pub const ID_GOSTR3410_94_B: Nid = _

pub const ID_GOSTR3410_94_BBIS: Nid = _

pub const ID_GOST28147_89_CC: Nid = _

pub const ID_GOSTR3410_94_CC: Nid = _

pub const ID_GOSTR3410_2001_CC: Nid = _

pub const ID_GOSTR3411_94_WITH_GOSTR3410_94_CC: Nid = _

pub const ID_GOSTR3411_94_WITH_GOSTR3410_2001_CC: Nid = _

pub const ID_GOSTR3410_2001_PARAMSET_CC: Nid = _

pub const CAMELLIA_128_CBC: Nid = _

pub const CAMELLIA_192_CBC: Nid = _

pub const CAMELLIA_256_CBC: Nid = _

pub const ID_CAMELLIA128_WRAP: Nid = _

pub const ID_CAMELLIA192_WRAP: Nid = _

pub const ID_CAMELLIA256_WRAP: Nid = _

pub const CAMELLIA_128_ECB: Nid = _

pub const CAMELLIA_128_OFB128: Nid = _

pub const CAMELLIA_128_CFB128: Nid = _

pub const CAMELLIA_192_ECB: Nid = _

pub const CAMELLIA_192_OFB128: Nid = _

pub const CAMELLIA_192_CFB128: Nid = _

pub const CAMELLIA_256_ECB: Nid = _

pub const CAMELLIA_256_OFB128: Nid = _

pub const CAMELLIA_256_CFB128: Nid = _

pub const CAMELLIA_128_CFB1: Nid = _

pub const CAMELLIA_192_CFB1: Nid = _

pub const CAMELLIA_256_CFB1: Nid = _

pub const CAMELLIA_128_CFB8: Nid = _

pub const CAMELLIA_192_CFB8: Nid = _

pub const CAMELLIA_256_CFB8: Nid = _

pub const KISA: Nid = _

pub const SEED_ECB: Nid = _

pub const SEED_CBC: Nid = _

pub const SEED_CFB128: Nid = _

pub const SEED_OFB128: Nid = _

pub const HMAC: Nid = _

pub const CMAC: Nid = _

pub const RC4_HMAC_MD5: Nid = _

pub const AES_128_CBC_HMAC_SHA1: Nid = _

pub const AES_192_CBC_HMAC_SHA1: Nid = _

pub const AES_256_CBC_HMAC_SHA1: Nid = _

pub const SM3: Nid = _

pub const SHA3_224: Nid = _

pub const SHA3_256: Nid = _

pub const SHA3_384: Nid = _

pub const SHA3_512: Nid = _

pub const SHAKE128: Nid = _

pub const SHAKE256: Nid = _