Struct openssl::ocsp::OcspBasicResponseRef

pub struct OcspBasicResponseRef(_);

Methods

impl OcspBasicResponseRef

fn verify(&self,
certs: &StackRef<X509>,
store: &X509StoreRef,
flags: Flag)
-> Result<(), ErrorStack>

Verifies the validity of the response.

The certs parameter contains a set of certificates that will be searched when locating the OCSP response signing certificate. Some responders do not include this in the response.

fn find_status<'a>(&'a self, id: &OcspCertIdRef) -> Option<Status<'a>>

Looks up the status for the specified certificate ID.