Struct openssl::pkey::PKey

pub struct PKey(_);

Methods

impl PKey

fn from_rsa(rsa: Rsa) -> Result<PKey, ErrorStack>

Create a new PKey containing an RSA key.

fn from_dsa(dsa: Dsa) -> Result<PKey, ErrorStack>

Create a new PKey containing a DSA key.

fn hmac(key: &[u8]) -> Result<PKey, ErrorStack>

Create a new PKey containing an HMAC key.

fn private_key_from_pem(buf: &[u8]) -> Result<PKey, ErrorStack>

Reads private key from PEM, takes ownership of handle

fn private_key_from_pem_cb<F>(buf: &[u8], pass_cb: F) -> Result<PKey, ErrorStack> where F: FnOnce(&mut [c_char]) -> usize

Read a private key from PEM, supplying a password callback to be invoked if the private key is encrypted.

The callback will be passed the password buffer and should return the number of characters placed into the buffer.

fn public_key_from_pem(buf: &[u8]) -> Result<PKey, ErrorStack>

Reads public key from PEM, takes ownership of handle