Struct openssl::pkey::PKey

pub struct PKey<T>(_, _);

A public or private key.

Implementations

impl<T> PKey<T>

pub fn from_rsa(rsa: Rsa<T>) -> Result<PKey<T>, ErrorStack>

Creates a new PKey containing an RSA key.

This corresponds to EVP_PKEY_assign_RSA.

pub fn from_dsa(dsa: Dsa<T>) -> Result<PKey<T>, ErrorStack>

Creates a new PKey containing a DSA key.

This corresponds to EVP_PKEY_assign_DSA.

pub fn from_dh(dh: Dh<T>) -> Result<PKey<T>, ErrorStack>

Creates a new PKey containing a Diffie-Hellman key.

This corresponds to EVP_PKEY_assign_DH.

pub fn from_ec_key(ec_key: EcKey<T>) -> Result<PKey<T>, ErrorStack>

Creates a new PKey containing an elliptic curve key.

This corresponds to EVP_PKEY_assign_EC_KEY.

impl PKey<Private>

pub fn hmac(key: &[u8]) -> Result<PKey<Private>, ErrorStack>

Creates a new PKey containing an HMAC key.

Note

To compute HMAC values, use the sign module.

pub fn cmac(cipher: &Cipher, key: &[u8]) -> Result<PKey<Private>, ErrorStack>

Creates a new PKey containing a CMAC key.

Requires OpenSSL 1.1.0 or newer.

Note

To compute CMAC values, use the sign module.

pub fn generate_x25519() -> Result<PKey<Private>, ErrorStack>

Generates a new private Ed25519 key

pub fn generate_x448() -> Result<PKey<Private>, ErrorStack>

Generates a new private Ed448 key

pub fn generate_ed25519() -> Result<PKey<Private>, ErrorStack>

Generates a new private Ed25519 key

pub fn generate_ed448() -> Result<PKey<Private>, ErrorStack>

Generates a new private Ed448 key

pub fn private_key_from_pem(pem: &[u8]) -> Result<PKey<Private>, ErrorStack>

Deserializes a private key from a PEM-encoded key type specific format.

This corresponds to PEM_read_bio_PrivateKey.

pub fn private_key_from_pem_passphrase(
pem: &[u8],
passphrase: &[u8]
) -> Result<PKey<Private>, ErrorStack>

Deserializes a private key from a PEM-encoded encrypted key type specific format.

This corresponds to PEM_read_bio_PrivateKey.

pub fn private_key_from_pem_callback<F>(
pem: &[u8],
callback: F
) -> Result<PKey<Private>, ErrorStack> where
F: FnOnce(&mut [u8]) -> Result<usize, ErrorStack>,

Deserializes a private key from a PEM-encoded encrypted key type specific format.

The callback should fill the password into the provided buffer and return its length.

This corresponds to PEM_read_bio_PrivateKey.

pub fn private_key_from_der(der: &[u8]) -> Result<PKey<Private>, ErrorStack>

Decodes a DER-encoded private key.

This function will automatically attempt to detect the underlying key format, and supports the unencrypted PKCS#8 PrivateKeyInfo structures as well as key type specific formats.

This corresponds to d2i_AutoPrivateKey.

pub fn private_key_from_pkcs8(der: &[u8]) -> Result<PKey<Private>, ErrorStack>

Deserializes a DER-formatted PKCS#8 unencrypted private key.

This method is mainly for interoperability reasons. Encrypted keyfiles should be preferred.

pub fn private_key_from_pkcs8_callback<F>(
der: &[u8],
callback: F
) -> Result<PKey<Private>, ErrorStack> where
F: FnOnce(&mut [u8]) -> Result<usize, ErrorStack>,

Deserializes a DER-formatted PKCS#8 private key, using a callback to retrieve the password if the key is encrpyted.

The callback should copy the password into the provided buffer and return the number of bytes written.

pub fn private_key_from_pkcs8_passphrase(
der: &[u8],
passphrase: &[u8]
) -> Result<PKey<Private>, ErrorStack>

Deserializes a DER-formatted PKCS#8 private key, using the supplied password if the key is encrypted.

Panics

Panics if passphrase contains an embedded null.

pub fn private_key_from_raw_bytes(
bytes: &[u8],
key_type: Id
) -> Result<PKey<Private>, ErrorStack>

Creates a private key from its raw byte representation

Algorithm types that support raw private keys are HMAC, X25519, ED25519, X448 or ED448

This corresponds to EVP_PKEY_new_raw_private_key.

impl PKey<Public>

pub fn public_key_from_pem(pem: &[u8]) -> Result<PKey<Public>, ErrorStack>

Decodes a PEM-encoded SubjectPublicKeyInfo structure.

The input should have a header of -----BEGIN PUBLIC KEY-----.

This corresponds to PEM_read_bio_PUBKEY.

pub fn public_key_from_der(der: &[u8]) -> Result<PKey<Public>, ErrorStack>

Decodes a DER-encoded SubjectPublicKeyInfo structure.

This corresponds to d2i_PUBKEY.

pub fn public_key_from_raw_bytes(
bytes: &[u8],
key_type: Id
) -> Result<PKey<Public>, ErrorStack>

Creates a public key from its raw byte representation

Algorithm types that support raw public keys are X25519, ED25519, X448 or ED448

This corresponds to EVP_PKEY_new_raw_public_key.