Struct openssl::rsa::Rsa

pub struct Rsa<T>(_, _);

An RSA key.

Methods

impl Rsa<Public>

pub fn from_public_components(
n: BigNum,
e: BigNum
) -> Result<Rsa<Public>, ErrorStack>

Creates a new RSA key with only public components.

n is the modulus common to both public and private key. e is the public exponent.

This corresponds to RSA_new and uses RSA_set0_key.

pub fn public_key_from_pem(pem: &[u8]) -> Result<Rsa<Public>, ErrorStack>

Decodes a PEM-encoded SubjectPublicKeyInfo structure containing an RSA key.

The input should have a header of -----BEGIN PUBLIC KEY-----.

This corresponds to PEM_read_bio_RSA_PUBKEY.

pub fn public_key_from_pem_pkcs1(pem: &[u8]) -> Result<Rsa<Public>, ErrorStack>

Decodes a PEM-encoded PKCS#1 RSAPublicKey structure.

The input should have a header of -----BEGIN RSA PUBLIC KEY-----.

This corresponds to PEM_read_bio_RSAPublicKey.

pub fn public_key_from_der(der: &[u8]) -> Result<Rsa<Public>, ErrorStack>

Decodes a DER-encoded SubjectPublicKeyInfo structure containing an RSA key.

This corresponds to d2i_RSA_PUBKEY.

pub fn public_key_from_der_pkcs1(der: &[u8]) -> Result<Rsa<Public>, ErrorStack>

Decodes a DER-encoded PKCS#1 RSAPublicKey structure.

This corresponds to d2i_RSAPublicKey.

impl Rsa<Private>

pub fn from_private_components(
n: BigNum,
e: BigNum,
d: BigNum,
p: BigNum,
q: BigNum,
dmp1: BigNum,
dmq1: BigNum,
iqmp: BigNum
) -> Result<Rsa<Private>, ErrorStack>

Creates a new RSA key with private components (public components are assumed).

n is the modulus common to both public and private key. e is the public exponent and d is the private exponent. p and q are the first and second factors of n. dmp1, dmq1, and iqmp are the exponents and coefficient for Chinese Remainder Theorem calculations which is used to speed up RSA operations.

This corresponds to RSA_new and uses RSA_set0_key, RSA_set0_factors, and RSA_set0_crt_params.

pub fn generate(bits: u32) -> Result<Rsa<Private>, ErrorStack>

Generates a public/private key pair with the specified size.

The public exponent will be 65537.

pub fn private_key_from_pem(pem: &[u8]) -> Result<Rsa<Private>, ErrorStack>

Deserializes a private key from a PEM-encoded PKCS#1 RSAPrivateKey structure.

This corresponds to PEM_read_bio_RSAPrivateKey.

pub fn private_key_from_pem_passphrase(
pem: &[u8],
passphrase: &[u8]
) -> Result<Rsa<Private>, ErrorStack>

Deserializes a private key from a PEM-encoded encrypted PKCS#1 RSAPrivateKey structure.

This corresponds to PEM_read_bio_RSAPrivateKey.

pub fn private_key_from_pem_callback<F>(
pem: &[u8],
callback: F
) -> Result<Rsa<Private>, ErrorStack> where
F: FnOnce(&mut [u8]) -> Result<usize, ErrorStack>,

Deserializes a private key from a PEM-encoded encrypted PKCS#1 RSAPrivateKey structure.

The callback should fill the password into the provided buffer and return its length.

This corresponds to PEM_read_bio_RSAPrivateKey.

pub fn private_key_from_der(der: &[u8]) -> Result<Rsa<Private>, ErrorStack>

Decodes a DER-encoded PKCS#1 RSAPrivateKey structure.

This corresponds to d2i_RSAPrivateKey.