Struct openssl::rsa::Rsa

pub struct Rsa(_);

Methods

impl Rsa

fn from_public_components(n: BigNum, e: BigNum) -> Result<Rsa, ErrorStack>

only useful for associating the key material directly with the key, it's safer to use the supplied load and save methods for DER formatted keys.

fn from_private_components(n: BigNum, e: BigNum, d: BigNum, p: BigNum, q: BigNum, dp: BigNum, dq: BigNum, qi: BigNum) -> Result<Rsa, ErrorStack>

fn generate(bits: u32) -> Result<Rsa, ErrorStack>

Generates a public/private key pair with the specified size.

The public exponent will be 65537.

fn private_key_from_pem(pem: &[u8]) -> Result<Rsa, ErrorStack>

Deserializes a PEM-formatted private key.

fn private_key_from_pem_passphrase(pem: &[u8], passphrase: &[u8]) -> Result<Rsa, ErrorStack>

Deserializes a PEM-formatted private key, using the supplied password if the key is encrypted.

Panics

Panics if passphrase contains an embedded null.

fn private_key_from_pem_callback<F>(pem: &[u8], callback: F) -> Result<Rsa, ErrorStack> where F: FnOnce(&mut [u8]) -> Result<usize, ErrorStack>

Deserializes a PEM-formatted private key, using a callback to retrieve a password if the key is encrypted.

The callback should copy the password into the provided buffer and return the number of bytes written.

fn private_key_from_der(der: &[u8]) -> Result<Rsa, ErrorStack>

Deserializes a private key from DER-formatted data.

fn public_key_from_pem(pem: &[u8]) -> Result<Rsa, ErrorStack>

Deserializes a public key from PEM-formatted data.

fn public_key_from_der(der: &[u8]) -> Result<Rsa, ErrorStack>

Deserializes a public key from DER-formatted data.

fn private_key_from_pem_cb<F>(buf: &[u8], pass_cb: F) -> Result<Rsa, ErrorStack> where F: FnOnce(&mut [c_char]) -> usize

Deprecated since 0.9.2
: use private_key_from_pem_callback