Struct openssl::rsa::Rsa

pub struct Rsa(_);

Methods

impl Rsa

fn from_public_components(n: BigNum, e: BigNum) -> Result<Rsa, ErrorStack>

only useful for associating the key material directly with the key, it's safer to use the supplied load and save methods for DER formatted keys.

fn from_private_components(n: BigNum, e: BigNum, d: BigNum, p: BigNum, q: BigNum, dp: BigNum, dq: BigNum, qi: BigNum) -> Result<Rsa, ErrorStack>

fn generate(bits: u32) -> Result<Rsa, ErrorStack>

Generates a public/private key pair with the specified size.

The public exponent will be 65537.

fn private_key_from_pem(buf: &[u8]) -> Result<Rsa, ErrorStack>

Reads an RSA private key from PEM formatted data.

fn private_key_from_pem_cb<F>(buf: &[u8], pass_cb: F) -> Result<Rsa, ErrorStack> where F: FnOnce(&mut [c_char]) -> usize

Reads an RSA private key from PEM formatted data and supplies a password callback.

fn public_key_from_pem(buf: &[u8]) -> Result<Rsa, ErrorStack>

Reads an RSA public key from PEM formatted data.