Struct openssl::rsa::RsaRef

pub struct RsaRef<T>(_, _);

Reference to RSA

Implementations

impl<T> RsaRef<T> where
T: HasPrivate,

pub fn private_key_to_pem(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the private key to a PEM-encoded PKCS#1 RSAPrivateKey structure.

The output will have a header of -----BEGIN RSA PRIVATE KEY-----.

This corresponds to PEM_write_bio_RSAPrivateKey.

pub fn private_key_to_pem_passphrase(
&self,
cipher: Cipher,
passphrase: &[u8]
) -> Result<Vec<u8>, ErrorStack>

Serializes the private key to a PEM-encoded encrypted PKCS#1 RSAPrivateKey structure.

The output will have a header of -----BEGIN RSA PRIVATE KEY-----.

This corresponds to PEM_write_bio_RSAPrivateKey.

pub fn private_key_to_der(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the private key to a DER-encoded PKCS#1 RSAPrivateKey structure.

This corresponds to i2d_RSAPrivateKey.

pub fn private_decrypt(
&self,
from: &[u8],
to: &mut [u8],
padding: Padding
) -> Result<usize, ErrorStack>

Decrypts data using the private key, returning the number of decrypted bytes.

Panics

Panics if self has no private components, or if to is smaller than self.size().

pub fn private_encrypt(
&self,
from: &[u8],
to: &mut [u8],
padding: Padding
) -> Result<usize, ErrorStack>

Encrypts data using the private key, returning the number of encrypted bytes.

Panics

Panics if self has no private components, or if to is smaller than self.size().

pub fn d(&self) -> &BigNumRef

Returns a reference to the private exponent of the key.

This corresponds to RSA_get0_key.

pub fn p(&self) -> Option<&BigNumRef>

Returns a reference to the first factor of the exponent of the key.

This corresponds to RSA_get0_factors.

pub fn q(&self) -> Option<&BigNumRef>

Returns a reference to the second factor of the exponent of the key.

This corresponds to RSA_get0_factors.

pub fn dmp1(&self) -> Option<&BigNumRef>

Returns a reference to the first exponent used for CRT calculations.

This corresponds to RSA_get0_crt_params.

pub fn dmq1(&self) -> Option<&BigNumRef>

Returns a reference to the second exponent used for CRT calculations.

This corresponds to RSA_get0_crt_params.

pub fn iqmp(&self) -> Option<&BigNumRef>

Returns a reference to the coefficient used for CRT calculations.

This corresponds to RSA_get0_crt_params.

pub fn check_key(&self) -> Result<bool, ErrorStack>

Validates RSA parameters for correctness

This corresponds to RSA_check_key.

impl<T> RsaRef<T> where
T: HasPublic,

pub fn public_key_to_pem(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the public key into a PEM-encoded SubjectPublicKeyInfo structure.

The output will have a header of -----BEGIN PUBLIC KEY-----.

This corresponds to PEM_write_bio_RSA_PUBKEY.

pub fn public_key_to_der(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the public key into a DER-encoded SubjectPublicKeyInfo structure.

This corresponds to i2d_RSA_PUBKEY.

pub fn public_key_to_pem_pkcs1(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the public key into a PEM-encoded PKCS#1 RSAPublicKey structure.

The output will have a header of -----BEGIN RSA PUBLIC KEY-----.

This corresponds to PEM_write_bio_RSAPublicKey.

pub fn public_key_to_der_pkcs1(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the public key into a DER-encoded PKCS#1 RSAPublicKey structure.

This corresponds to i2d_RSAPublicKey.

pub fn size(&self) -> u32

Returns the size of the modulus in bytes.

This corresponds to RSA_size.

pub fn public_decrypt(
&self,
from: &[u8],
to: &mut [u8],
padding: Padding
) -> Result<usize, ErrorStack>

Decrypts data using the public key, returning the number of decrypted bytes.

Panics

Panics if to is smaller than self.size().

pub fn public_encrypt(
&self,
from: &[u8],
to: &mut [u8],
padding: Padding
) -> Result<usize, ErrorStack>

Encrypts data using the public key, returning the number of encrypted bytes.

Panics

Panics if to is smaller than self.size().

pub fn n(&self) -> &BigNumRef

Returns a reference to the modulus of the key.

This corresponds to RSA_get0_key.

pub fn e(&self) -> &BigNumRef

Returns a reference to the public exponent of the key.

This corresponds to RSA_get0_key.