Module openssl::ssl

Modules

error

Structs

CipherBits

NonblockingSslStream

Ssl

SslCipher

SslContext
An SSL context object

SslContextOptions

SslStream
A stream wrapper which handles SSL encryption for an underlying stream.

SslVerifyMode

Enums

Error
An SSL error.

MaybeSslStream
A utility type to help in cases where the use of SSL is decided at runtime.

SslMethod
Determines the SSL method supported

Constants

SSL_OP_ALL

SSL_OP_ALLOW_UNSAFE_LEGACY_RENEGOTIATION

SSL_OP_CIPHER_SERVER_PREFERENCE

SSL_OP_CISCO_ANYCONNECT

SSL_OP_COOKIE_EXCHANGE

SSL_OP_CRYPTOPRO_TLSEXT_BUG

SSL_OP_DONT_INSERT_EMPTY_FRAGMENTS

SSL_OP_EPHEMERAL_RSA

SSL_OP_LEGACY_SERVER_CONNECT

SSL_OP_MICROSOFT_BIG_SSLV3_BUFFER

SSL_OP_MICROSOFT_SESS_ID_BUG

SSL_OP_MSIE_SSLV2_RSA_PADDING

SSL_OP_NETSCAPE_CA_DN_BUG

SSL_OP_NETSCAPE_CHALLENGE_BUG

SSL_OP_NETSCAPE_DEMO_CIPHER_CHANGE_BUG

SSL_OP_NETSCAPE_REUSE_CIPHER_CHANGE_BUG

SSL_OP_NO_COMPRESSION

SSL_OP_NO_DTLSV1

SSL_OP_NO_DTLSV1_2

SSL_OP_NO_QUERY_MTU

SSL_OP_NO_SESSION_RESUMPTION_ON_RENEGOTIATION

SSL_OP_NO_SSLV2

SSL_OP_NO_SSLV3

SSL_OP_NO_SSL_MASK

SSL_OP_NO_TICKET

SSL_OP_NO_TLSV1

SSL_OP_NO_TLSV1_1

SSL_OP_NO_TLSV1_2

SSL_OP_PKCS1_CHECK_1

SSL_OP_PKCS1_CHECK_2

SSL_OP_SAFARI_ECDHE_ECDSA_BUG

SSL_OP_SINGLE_DH_USE

SSL_OP_SINGLE_ECDH_USE

SSL_OP_SSLEAY_080_CLIENT_DH_BUG

SSL_OP_SSLREF2_REUSE_CERT_TYPE_BUG

SSL_OP_TLSEXT_PADDING

SSL_OP_TLS_BLOCK_PADDING_BUG

SSL_OP_TLS_D5_BUG

SSL_OP_TLS_ROLLBACK_BUG

SSL_VERIFY_FAIL_IF_NO_PEER_CERT
Terminate handshake if client did not return a certificate. Use together with SSL_VERIFY_PEER.

SSL_VERIFY_NONE
Do not verify the server's certificate

SSL_VERIFY_PEER
Verify that the server's certificate is trusted

Traits

IntoSsl

Functions

init
Manually initialize SSL. It is optional to call this function and safe to do so more than once.

Type Definitions

ServerNameCallback
The signature of functions that can be used to choose the context depending on the server name

ServerNameCallbackData

VerifyCallback
The signature of functions that can be used to manually verify certificates

VerifyCallbackData
The signature of functions that can be used to manually verify certificates when user-data should be carried for all verification process