Module openssl::ssl

Modules

error

Structs

Ssl

SslContext
An SSL context object

SslStream
A stream wrapper which handles SSL encryption for an underlying stream.

Enums

MaybeSslStream
A utility type to help in cases where the use of SSL is decided at runtime.

SslMethod
Determines the SSL method supported

SslVerifyMode
Determines the type of certificate verification used

Type Definitions

VerifyCallback
The signature of functions that can be used to manually verify certificates

VerifyCallbackData
The signature of functions that can be used to manually verify certificates when user-data should be carried for all verification process