Struct openssl::ssl::ConnectConfiguration

pub struct ConnectConfiguration { /* private fields */ }

A type which allows for configuration of a client-side TLS session before connection.

Implementations

impl ConnectConfiguration

pub fn use_server_name_indication(self, use_sni: bool) -> ConnectConfiguration

A builder-style version of set_use_server_name_indication.

pub fn set_use_server_name_indication(&mut self, use_sni: bool)

Configures the use of Server Name Indication (SNI) when connecting.

Defaults to true.

pub fn verify_hostname(self, verify_hostname: bool) -> ConnectConfiguration

A builder-style version of set_verify_hostname.

pub fn set_verify_hostname(&mut self, verify_hostname: bool)

Configures the use of hostname verification when connecting.

Defaults to true.

Warning

You should think very carefully before you use this method. If hostname verification is not used, any valid certificate for any site will be trusted for use from any other. This introduces a significant vulnerability to man-in-the-middle attacks.

pub fn into_ssl(self, domain: &str) -> Result<Ssl, ErrorStack>

Returns an Ssl configured to connect to the provided domain.

The domain is used for SNI (if it is not an IP address) and hostname verification if enabled.

pub fn connect<S>( self, domain: &str, stream: S ) -> Result<SslStream<S>, HandshakeError<S>>where S: Read + Write,

Initiates a client-side TLS session on a stream.

The domain is used for SNI and hostname verification if enabled.