Struct openssl::ssl::Ssl

pub struct Ssl(_);

Methods

impl Ssl

fn new(ctx: &SslContext) -> Result<Ssl, ErrorStack>

fn connect<S>(self, stream: S) -> Result<SslStream<S>, HandshakeError<S>> where S: Read + Write

Creates an SSL/TLS client operating over the provided stream.

Warning

OpenSSL's default configuration is insecure. It is highly recommended to use SslConnector rather than Ssl directly, as it manages that configuration.

fn accept<S>(self, stream: S) -> Result<SslStream<S>, HandshakeError<S>> where S: Read + Write

Creates an SSL/TLS server operating over the provided stream.

Warning

OpenSSL's default configuration is insecure. It is highly recommended to use SslAcceptor rather than Ssl directly, as it manages that configuration.