Struct openssl::ssl::SslAcceptorBuilder

pub struct SslAcceptorBuilder(_);

A builder for SslAcceptors.

Methods

impl SslAcceptorBuilder

fn mozilla_intermediate<I>(method: SslMethod, private_key: &PKeyRef, certificate: &X509Ref, chain: I) -> Result<SslAcceptorBuilder, ErrorStack> where I: IntoIterator, I::Item: AsRef<X509Ref>

Creates a new builder configured to connect to non-legacy clients. This should generally be considered a reasonable default choice.

This corresponds to the intermediate configuration of Mozilla's server side TLS recommendations. See its documentation for more details on specifics.

fn mozilla_modern<I>(method: SslMethod, private_key: &PKeyRef, certificate: &X509Ref, chain: I) -> Result<SslAcceptorBuilder, ErrorStack> where I: IntoIterator, I::Item: AsRef<X509Ref>

Creates a new builder configured to connect to modern clients.

This corresponds to the modern configuration of Mozilla's server side TLS recommendations. See its documentation for more details on specifics.

fn builder(&self) -> &SslContextBuilder

Returns a shared reference to the inner SslContextBuilder.

fn builder_mut(&mut self) -> &mut SslContextBuilder

Returns a mutable reference to the inner SslContextBuilder.

fn build(self) -> SslAcceptor

Consumes the builder, returning a SslAcceptor.