openssl::ssl
Struct SslContext

pub struct SslContext(/* private fields */);

A context object for TLS streams.

Applications commonly configure a single SslContext that is shared by all of its SslStreams.

Implementations

impl SslContext

pub fn builder(method: SslMethod) -> Result<SslContextBuilder, ErrorStack>

Creates a new builder object for an SslContext.

pub fn new_ex_index<T>() -> Result<Index<SslContext, T>, ErrorStack>
where T: 'static + Sync + Send,

Returns a new extra data index.

Each invocation of this function is guaranteed to return a distinct index. These can be used to store data in the context that can be retrieved later by callbacks, for example.

This corresponds to SSL_CTX_get_ex_new_index.