Struct openssl::ssl::SslRef

pub struct SslRef(_);

Reference to an Ssl.

Methods

impl SslRef

pub fn set_verify(&mut self, mode: SslVerifyMode)

Like SslContextBuilder::set_verify.

This corresponds to SSL_set_verify.

pub fn set_verify_callback<F>(&mut self, mode: SslVerifyMode, verify: F) where
F: Fn(bool, &X509StoreContextRef) -> bool + Any + 'static + Sync + Send,

Like SslContextBuilder::set_verify_callback.

This corresponds to SSL_set_verify.

pub fn set_tmp_dh(&mut self, dh: &DhRef) -> Result<(), ErrorStack>

Like SslContextBuilder::set_tmp_dh.

This corresponds to SSL_set_tmp_dh.

pub fn set_tmp_dh_callback<F>(&mut self, callback: F) where
F: Fn(&mut SslRef, bool, u32) -> Result<Dh, ErrorStack> + Any + 'static + Sync + Send,

Like SslContextBuilder::set_tmp_dh_callback.

This corresponds to SSL_set_tmp_dh_callback.

pub fn set_tmp_ecdh(&mut self, key: &EcKeyRef) -> Result<(), ErrorStack>

Like SslContextBuilder::set_tmp_ecdh.

This corresponds to SSL_set_tmp_ecdh.

pub fn current_cipher(&self) -> Option<&SslCipherRef>

Returns the current cipher if the session is active.

This corresponds to SSL_get_current_cipher.

pub fn state_string(&self) -> &'static str

Returns a short string describing the state of the session.

This corresponds to SSL_state_string.

pub fn state_string_long(&self) -> &'static str

Returns a longer string describing the state of the session.

This corresponds to SSL_state_string_long.

pub fn set_hostname(&mut self, hostname: &str) -> Result<(), ErrorStack>

Sets the host name to be sent to the server for Server Name Indication (SNI).

It has no effect for a server-side connection.

This corresponds to SSL_set_tlsext_host_name.

pub fn peer_certificate(&self) -> Option<X509>

Returns the peer's certificate, if present.

This corresponds to SSL_get_peer_certificate.

pub fn peer_cert_chain(&self) -> Option<&StackRef<X509>>

Returns the certificate chain of the peer, if present.

On the client side, the chain includes the leaf certificate, but on the server side it does not. Fun!

This corresponds to SSL_get_peer_cert_chain.

pub fn certificate(&self) -> Option<&X509Ref>

Like SslContext::certificate.

This corresponds to SSL_get_certificate.

pub fn private_key(&self) -> Option<&PKeyRef>

Like SslContext::private_key.

This corresponds to SSL_get_privatekey.

pub fn version(&self) -> &'static str

Returns a string describing the protocol version of the session.

This corresponds to SSL_get_version.

pub fn selected_npn_protocol(&self) -> Option<&[u8]>

Returns the protocol selected by performing Next Protocol Negotiation, if any.

The protocol's name is returned is an opaque sequence of bytes. It is up to the client to interpret it.

This corresponds to SSL_get0_next_proto_negotiated.

pub fn pending(&self) -> usize

Returns the number of bytes remaining in the currently processed TLS record.

If this is greater than 0, the next call to read will not call down to the underlying stream.

This corresponds to [SSL_pending].

pub fn compression(&self) -> Option<&str>

Returns the compression method currently in use.

This corresponds to SSL_get_current_compression.

pub fn servername(&self) -> Option<&str>

Returns the servername sent by the client via Server Name Indication (SNI).

It is only useful on the server side.

This corresponds to SSL_get_servername.

pub fn set_ssl_context(&mut self, ctx: &SslContextRef) -> Result<(), ErrorStack>

Changes the context corresponding to the current connection.

It is most commonly used in the Server Name Indication (SNI) callback.

This corresponds to SSL_set_SSL_CTX.

pub fn ssl_context(&self) -> &SslContextRef

Returns the context corresponding to the current connection.

This corresponds to SSL_get_SSL_CTX.

pub fn verify_result(&self) -> Option<X509VerifyError>

Returns the certificate verification result.

This corresponds to SSL_get_verify_result.

pub fn session(&self) -> Option<&SslSessionRef>

Returns a shared reference to the SSL session.

This corresponds to SSL_get_session.

pub unsafe fn set_session(
&mut self,
session: &SslSessionRef
) -> Result<(), ErrorStack>

Sets the session to be used.

This should be called before the handshake to attempt to reuse a previously established session. If the server is not willing to reuse the session, a new one will be transparently negotiated.

This corresponds to SSL_set_session.

Safety

The caller of this method is responsible for ensuring that the session is associated with the same SslContext as this Ssl.

pub fn session_reused(&self) -> bool

Determines if the session provided to set_session was successfully reused.

This corresponds to SSL_session_reused.

pub fn set_status_type(&mut self, type_: StatusType) -> Result<(), ErrorStack>

Sets the status response a client wishes the server to reply with.

This corresponds to SSL_set_tlsext_status_type.

pub fn ocsp_status(&self) -> Option<&[u8]>

Returns the server's OCSP response, if present.

This corresponds to [SSL_get_tlsext_status_oscp_resp].

pub fn set_ocsp_status(&mut self, response: &[u8]) -> Result<(), ErrorStack>

Sets the OCSP response to be returned to the client.

This corresponds to [SSL_set_tlsext_status_oscp_resp].

pub fn is_server(&self) -> bool

Determines if this Ssl is configured for server-side or client-side use.

This corresponds to SSL_is_server.

pub fn set_ex_data<T>(&mut self, index: Index<Ssl, T>, data: T)

Sets the extra data at the specified index.

This can be used to provide data to callbacks registered with the context. Use the Ssl::new_ex_index method to create an Index.

This corresponds to SSL_set_ex_data.

pub fn ex_data<T>(&self, index: Index<Ssl, T>) -> Option<&T>

Returns a reference to the extra data at the specified index.

This corresponds to SSL_get_ex_data.