Struct openssl::ssl::SslSessionRef

pub struct SslSessionRef(_);

Reference to SslSession.

Implementations

impl SslSessionRef

pub fn id(&self) -> &[u8]

Returns the SSL session ID.

This corresponds to SSL_SESSION_get_id.

pub fn master_key_len(&self) -> usize

Returns the length of the master key.

This corresponds to SSL_SESSION_get_master_key.

pub fn master_key(&self, buf: &mut [u8]) -> usize

Copies the master key into the provided buffer.

Returns the number of bytes written, or the size of the master key if the buffer is empty.

This corresponds to SSL_SESSION_get_master_key.

pub fn max_early_data(&self) -> u32

Gets the maximum amount of early data that can be sent on this session.

Requires OpenSSL 1.1.1 or LibreSSL 3.4.0 or newer.

This corresponds to SSL_SESSION_get_max_early_data.

pub fn time(&self) -> i64

Returns the time at which the session was established, in seconds since the Unix epoch.

This corresponds to SSL_SESSION_get_time.

pub fn timeout(&self) -> i64

Returns the sessions timeout, in seconds.

A session older than this time should not be used for session resumption.

This corresponds to SSL_SESSION_get_timeout.

pub fn protocol_version(&self) -> SslVersion

Returns the session’s TLS protocol version.

Requires OpenSSL 1.1.0 or LibreSSL 2.7.0 or newer.

This corresponds to SSL_SESSION_get_protocol_version.

pub fn to_der(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the session into a DER-encoded structure.

This corresponds to i2d_SSL_SESSION.