Struct openssl::ssl::SslStreamBuilder

pub struct SslStreamBuilder<S> { /* fields omitted */ }

A partially constructed SslStream, useful for unusual handshakes.

Methods

impl<S> SslStreamBuilder<S> where
S: Read + Write,

pub fn new(ssl: Ssl, stream: S) -> Self

Begin creating an SslStream atop stream

pub fn stateless(&mut self) -> Result<bool, ErrorStack>

Perform a stateless server-side handshake

Requires that cookie generation and verification callbacks were set on the SSL context.

Returns Ok(true) if a complete ClientHello containing a valid cookie was read, in which case the handshake should be continued via accept. If a HelloRetryRequest containing a fresh cookie was transmitted, Ok(false) is returned instead. If the handshake cannot proceed at all, Err is returned.

This corresponds to SSL_stateless

pub fn connect(self) -> Result<SslStream<S>, HandshakeError<S>>

See Ssl::connect

pub fn accept(self) -> Result<SslStream<S>, HandshakeError<S>>

See Ssl::accept

impl<S> SslStreamBuilder<S>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_ref(&self) -> &S

Returns a shared reference to the underlying stream.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_mut(&mut self) -> &mut S

Returns a mutable reference to the underlying stream.

Warning

It is inadvisable to read from or write to the underlying stream as it will most likely corrupt the SSL session.

pub fn ssl(&self) -> &SslRef

Returns a shared reference to the Ssl object associated with this builder.