Struct openssl::stack::StackRef

pub struct StackRef<T: Stackable>(_, _);

Methods

impl<T: Stackable> StackRef<T>

pub fn len(&self) -> usize

Returns the number of items in the stack

pub fn iter(&self) -> Iter<T>

pub fn iter_mut(&mut self) -> IterMut<T>

pub fn get(&self, idx: usize) -> Option<&T::Ref>

Returns a reference to the element at the given index in the stack or None if the index is out of bounds

pub fn get_mut(&mut self, idx: usize) -> Option<&mut T::Ref>

Returns a mutable reference to the element at the given index in the stack or None if the index is out of bounds

pub fn push(&mut self, data: T) -> Result<(), ErrorStack>

Pushes a value onto the top of the stack.

pub fn pop(&mut self) -> Option<T>

Removes the last element from the stack and returns it.