Trait openssl::stack::Stackable

pub trait Stackable: OpenSslType {
    type StackType;
}

Trait implemented by types which can be placed in a stack.

Like OpenSslType, it should not be implemented for any type outside of this crate.

Associated Types

type StackType

The C stack type for this element.

Generally called stack_st_{ELEMENT_TYPE}, normally hidden by the STACK_OF(ELEMENT_TYPE) macro in the OpenSSL API.