Struct openssl::symm::Cipher

pub struct Cipher(_);

Methods

impl Cipher

fn aes_128_ecb() -> Cipher

fn aes_128_cbc() -> Cipher

fn aes_128_xts() -> Cipher

fn aes_128_ctr() -> Cipher

fn aes_128_cfb1() -> Cipher

fn aes_128_cfb128() -> Cipher

fn aes_128_cfb8() -> Cipher

fn aes_128_gcm() -> Cipher

fn aes_256_ecb() -> Cipher

fn aes_256_cbc() -> Cipher

fn aes_256_xts() -> Cipher

fn aes_256_ctr() -> Cipher

fn aes_256_cfb1() -> Cipher

fn aes_256_cfb128() -> Cipher

fn aes_256_cfb8() -> Cipher

fn aes_256_gcm() -> Cipher

fn des_cbc() -> Cipher

fn des_ecb() -> Cipher

fn rc4() -> Cipher

unsafe fn from_ptr(ptr: *const EVP_CIPHER) -> Cipher

fn as_ptr(&self) -> *const EVP_CIPHER

fn key_len(&self) -> usize

Returns the length of keys used with this cipher.

fn iv_len(&self) -> Option<usize>

Returns the length of the IV used with this cipher, or None if the cipher does not use an IV.

fn block_size(&self) -> usize

Returns the block size of the cipher.

Note

Stream ciphers such as RC4 have a block size of 1.