Module openssl::version

Functions

built_on
The date of the build process in the form "built on: ..." if available or "built on: date not available" otherwise.

c_flags
The compiler flags set for the compilation process in the form "compiler: ..." if available or "compiler: information not available" otherwise.

dir
The "OPENSSLDIR" setting of the library build in the form "OPENSSLDIR: "..."" if available or "OPENSSLDIR: N/A" otherwise.

number
OPENSSL_VERSION_NUMBER is a numeric release version identifier:

platform
The "Configure" target of the library build in the form "platform: ..." if available or "platform: information not available" otherwise.

version
The text variant of the version number and the release date. For example, "OpenSSL 0.9.5a 1 Apr 2000".