Struct openssl::x509::X509StoreContext

pub struct X509StoreContext(_);

An X509 certificate store context.

Implementations

impl X509StoreContext

pub fn ssl_idx() -> Result<Index<X509StoreContext, SslRef>, ErrorStack>

Returns the index which can be used to obtain a reference to the Ssl associated with a context.

This corresponds to SSL_get_ex_data_X509_STORE_CTX_idx.

pub fn new() -> Result<X509StoreContext, ErrorStack>

Creates a new X509StoreContext instance.

This corresponds to X509_STORE_CTX_new.