Struct openssl::x509::X509StoreContextRef

pub struct X509StoreContextRef(_);

Reference to X509StoreContext.

Implementations

impl X509StoreContextRef

pub fn ex_data<T>(&self, index: Index<X509StoreContext, T>) -> Option<&T>

Returns application data pertaining to an X509 store context.

This corresponds to X509_STORE_CTX_get_ex_data.

pub fn error(&self) -> X509VerifyResult

Returns the error code of the context.

This corresponds to X509_STORE_CTX_get_error.

pub fn init<F, T>(
&mut self,
trust: &X509StoreRef,
cert: &X509Ref,
cert_chain: &StackRef<X509>,
with_context: F
) -> Result<T, ErrorStack> where
F: FnOnce(&mut X509StoreContextRef) -> Result<T, ErrorStack>,

Initializes this context with the given certificate, certificates chain and certificate store. After initializing the context, the with_context closure is called with the prepared context. As long as the closure is running, the context stays initialized and can be used to e.g. verify a certificate. The context will be cleaned up, after the closure finished.

trust - The certificate store with the trusted certificates.

cert - The certificate that should be verified.

cert_chain - The certificates chain.

with_context - The closure that is called with the initialized context.

This corresponds to X509_STORE_CTX_init before calling with_context and to X509_STORE_CTX_cleanup after calling with_context.

pub fn verify_cert(&mut self) -> Result<bool, ErrorStack>

Verifies the stored certificate.

Returns true if verification succeeds. The error method will return the specific validation error if the certificate was not valid.

This will only work inside of a call to init.

This corresponds to X509_verify_cert.

pub fn set_error(&mut self, result: X509VerifyResult)

Set the error code of the context.

This corresponds to X509_STORE_CTX_set_error.

pub fn current_cert(&self) -> Option<&X509Ref>

Returns a reference to the certificate which caused the error or None if no certificate is relevant to the error.

This corresponds to X509_STORE_CTX_get_current_cert.

pub fn error_depth(&self) -> u32

Returns a non-negative integer representing the depth in the certificate chain where the error occurred. If it is zero it occurred in the end entity certificate, one if it is the certificate which signed the end entity certificate and so on.

This corresponds to X509_STORE_CTX_get_error_depth.

pub fn chain(&self) -> Option<&StackRef<X509>>

Returns a reference to a complete valid X509 certificate chain.

This corresponds to X509_STORE_CTX_get0_chain.